### 2021-01-14 #31 @NickleDave
Hi all, just adding editor checks.

Thank you @JonnyTran for your detailed submission

## Editor checks:

    + yes, tool for working with different types of -omics data
    + yes, Travis CI
- [x] **License:** The package has an OSI accepted license
    + yes, MIT
    + yup

submitter did not check "yes" for submitting to JOSS

---

## Editor comments
Overall:
- perfect fit for PyOpenSci, definitely looks like it has the potential to be an incredibly useful data retrieval / extraction / munging tool that faciliates reproducible pipelines
    - many use cases e.g. buildng datasets for machine learning from multiple omics sources

Minor comments:
 things I notice on first glance
- no DOI for releases, e.g. using Zenodo integration. Good idea to have DOI to make releases citable, e.g. in papers to make explicit which version was used
- no explicit link to docs page
  + this could be in README and / or in the "about" section -- GitHub lets you add a link there in addition to the text description
  + I did finally find the docs by clicking on the badge ... but if it were me I would have a big giant link point to the docs as obvious as possible
- some info could be added to README
> This is definitely an "elements of style" question / matter of personal preference, and not a thing that is directly addressed by our developer guide, but:
> I really like the readthedocs suggestions for what to include in your README:
> https://www.writethedocs.org/guide/writing/beginners-guide-to-docs/#readme
> I don't see a lot of those things included in yours now
> Don't get me wrong, It is definitely great that you have many concrete examples built around code snippets
> but there's a trade-off involved in having many of them at the expense of having all the other information, e.g. more detailed install instructions, info on where/how to raise issues, license type, etc.
> Maybe have one good  code snippet + then link to the other examples via a link to the docs?
> You don't want to miss getting all that other info to potential contributors / other developers who are kind of most likely to look at your GitHub README, IMHO
> but for sure many successful projects that don't follow this pattern (see for example the `tqdm` README)

---

Reviewers:
Due date:
