### 2021-01-17 #31 @JonnyTran
Hi @NickleDave, will it be possible for OpenOmics to make submission to JOSS at this stage of the review? I just changed my mind about this, but I already have a complete manuscript and I'm ready to provide the `paper.md` within a few days.
