### 2021-01-17 #31 @NickleDave
Hey @JonnyTran I think that could be okay. But first let me make sure I understand what you're asking.

> Hi @NickleDave and @lwasser, will it be possible for OpenOmics to make submission to JOSS at this stage of the review?
Do you mean that you want to change "Publication Options" in your submission above, and check the box for "automatically submit to" JOSS?
If so, yes I think that's fine. @lwasser please confirm

> I just changed my mind about this, but I already have a complete manuscript and I'm ready to provide the `paper.md` within a few days.
Do you mean that you have a complete, separate manuscript written about OpenOmics, in addition to the `paper.md` you would submit to JOSS?
E.g., like Physcraper
https://github.com/pyOpenSci/software-review/issues/26
which has a paper on biorxiv
https://www.biorxiv.org/content/10.1101/2020.09.15.299156v1

If that's the case, we should discuss more whether you want to submit to JOSS. We can perhaps tag some editors and ask if they can give us input. My impression is that JOSS is usually meant to provide a mechanism for getting publication credit for software in cases where the developer/maintainer **can't** easily publish a paper about it.
Although I think they might have changed some of the language in their submission guidelines about this.
See for example: https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software

so:
if you just want to go through PyOpenSci review and then submit to JOSS at the end, yes, totally fine assuming @lwasser agrees with me.
Anything else, we should probably discuss a little more first
