### 2021-01-21 #31 @JonnyTran
Hi @NickleDave, sorry it took awhile to consult with my advisor.
Yes, I meant that I'd like to check the box on automatic submission, and no, that I have not submitted a separate manuscript elsewhere. My intention for the JOSS submission is to publish contributions on the technical software aspects that I have thus far. But in later months (after finishing the web-app features in openomics), I do plan on making another contribution to a bioinformatics journal on the scientific bioinformatics use-cases. So I think this would fall under the "co-publication".

Thank you for the clarifications!
