### 2021-01-25 #31 @NickleDave
Hi again @JonnyTran excited to let you know that @gawbul and @ksielemann have both kindly accepted our invitation to review

@gawbul actually started PyOpenSci back in 2013 (see [this ROpenSci blogpost](https://ropensci.org/blog/2013/05/16/pyopensci/)) and develops related tools such as [pyEnsemblREST](https://github.com/gawbul/pyEnsemblRest)

@ksielemann [has significant experience with omics datasets](https://scholar.google.com/citations?user=UVeUz10AAAAJ&hl=en&oi=sra) and was recommended to us by @bpucker as developer of the tool [QUOD](https://github.com/ksielemann/QUOD) (from their publication https://www.biorxiv.org/content/10.1101/2020.04.28.065714v1.abstract)

@gawbul and @ksielemann here are related links again for your convenience:
Our [reviewers guide](https://www.pyopensci.org/contributing-guide/open-source-software-submissions/reviewer-guide.html) details what we look for in a package review, and includes [links to sample reviews](https://www.pyopensci.org/contributing-guide/open-source-software-submissions/reviewer-guide.html#examples-of-past-pyopensci-reviews). Our standards are detailed in our [packaging guide](https://www.pyopensci.org/contributing-guide/authoring/overview.html), and we provide a [reviewer template](https://www.pyopensci.org/contributing-guide/appendices/templates#review-template) for you to use. Please make sure you do not have a [conflict of interest](https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/policies-and-guidelines.html#conflict-of-interest) preventing you from reviewing this package. If you have questions or feedback, feel free to ask me here or by email, or post to the [pyOpenSci forum](https://pyopensci.discourse.group/).

I will update my editor checks above to add you both as reviewers, and set an initial due date of three weeks: February 15, 2021
