### 2021-02-01 #31 @ksielemann
>  It is not completely clear what OpenOmics can be used for. An overview of the available functions and methods would be great: What exactly does OpenOmics do? What are specific usage examples (also e.g. after using OpenOmics: What's next?)?:

> 'OpenOmics facilitates the common coding tasks when preparing data for bioinformatics analysis.': For which bioinformatic analyses exactly?

> '# Load each expression dataframe': additional ')' in lines should be removed! as current form results in an error

> `mRNA = MessengerRNA(data=folder_path+"LUAD__geneExp.txt", transpose=True, usecols="GeneSymbol|TCGA", gene_index="GeneSymbol", gene_level="gene_name") `
Results in warnings (only in first time use):
_'/homes/.local/lib/python3.7/site-packages/openomics/transcriptomics.py:95: FutureWarning:
read_table is deprecated, use read_csv instead.
/homes/.local/lib/python3.7/site-packages/openomics/transcriptomics.py:95: ParserWarning:
Falling back to the 'python' engine because the 'c' engine does not support sep=None with delim_whitespace=False; you can avoid this warning by specifying engine='python'.'_

>  `som = SomaticMutation(data=folder_path+"LUAD__somaticMutation_geneLevel.txt", transpose=True, usecols="GeneSymbol|TCGA", gene_index="gene_name")`
Results in: _'KeyError: 'gene_name''_. This should probably be _'gene_index="GeneSymbol"'_.

> `luad_data.add_clinical_data(clinical_data=folder_path+"nationwidechildrens.org_clinical_patient_luad.txt")`
Results in warning: _'/homes/.local/lib/python3.7/site-packages/openomics/clinical.py:51: FutureWarning:
read_table is deprecated, use read_csv instead, passing sep='\t'.'_

> `gencode = GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz","basic.annotation.gtf": "gencode.v32.basic.annotation.gtf.gz","lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz","transcripts.fa": "gencode.v32.transcripts.fa.gz"},remove_version_num=True,npartitions=5)`
Results in: _'AttributeError: '_io.TextIOWrapper' object has no attribute 'startswith''_.

- [ ] **Function Documentation:** for all user-facing functions

> Please see above (in _'A statement of need'_).

- [x] **Examples** for all user-facing functions
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be more wide than high. (Note that the badge for pyOpenSci peer-review will be provided upon acceptance.)
- [ ] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.

> Please see above (in _'A statement of need'_). The goals could be communicated more specifically.

- [x] Any additional setup required (authentication tokens, etc)
- [x] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages

> Citation information is missing at the end of the README.

Package structure should follow general community best-practices. In general please consider:

- [x] The documentation is easy to find and understand

> Please see above (in _'A statement of need'_).  Specific examples for bioinformatic analyses after the use of OpenOmics could be added.

> Please see above (in _'A statement of need'_). An overview of all methods and functions of the package would be helpful.

> Installation with `pip install openomics` worked fine on the Linux system I am using. However, using my windows computer, I got the following error: _error: Microsoft Visual C++ 14.0 is required_

> A list of dependencies/requirements in the README would be great.

> `from openomics import MultiOmics` results in the "UserWarning: Tensorflow not installed; ParametricUMAP will be unavailable"

> Please see above.

> Please see above.

- [ ] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

#### For packages co-submitting to JOSS

---

The concept of OpenOmics seems interesting and useful for the integration of various omics datasets. Overall, the package has a clear documentation. However, there are still a few issues that should be addressed. Please see the points above and below.

- Download of test data: Add code to download the data within a Python script so that the user does not have to download the whole repository or describe exactly how to download the test data. (Import TCGA LUAD data included in tests dataset (preprocessed from TCGA-Assembler) `folder_path = "tests/data/TCGA_LUAD/" # Located at https://github.com/BioMeCIS-Lab/OpenOmics/tree/master/tests) `is not sufficient.) Here is an example on how I did this:

```
#fetch data
import tarfile
import urllib.request

def fetch_data(file_url, own_path, file_name):
    if not os.path.isdir(own_path):
        os.makedirs(own_path)
    own_file_path = os.path.join(own_path, file_name)
    urllib.request.urlretrieve(file_url, own_file_path)

FILE_NAMES = ["LUAD__geneExp.txt",
              "LUAD__miRNAExp__RPM.txt",
              "LUAD__protein_RPPA.txt",
              "LUAD__somaticMutation_geneLevel.txt",
              "TCGA-rnaexpr.tsv",
              "genome.wustl.edu_biospecimen_sample_luad.txt",
              "nationwidechildrens.org_clinical_drug_luad.txt",
              "nationwidechildrens.org_clinical_patient_luad.txt",
              "protein_RPPA.txt"]

DOWNLOAD_ROOT = "https://raw.githubusercontent.com/BioMeCIS-Lab/OpenOmics/master/tests/data/TCGA_LUAD/"
OWN_PATH = os.path.join("data", "omics")

for file_name in FILE_NAMES:
  FILE_URL = DOWNLOAD_ROOT + file_name
  fetch_data(FILE_URL, OWN_PATH, file_name)
```

- In the ReadTheDocs documentation:

`gtex = GTEx(path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/")` Results in: _OSError: Not enough free space in /homes/.astropy/cache/download/url to download a 3.6G file, only 2.7G left_.

Is there a possibility to choose the directory in which the files should be downloaded? This would be great.
