### 2021-02-02 #31 @JonnyTran
Thanks so much for the fantastic reviews @ksielemann! I can work on the revisions, which should be available in 2 weeks.

> Download of test data. Add code to download the data within a Python script so that the user does not have to download the whole repository or describe exactly how to download the test data.

Thanks for pointing this out and providing the automated scripts. I initially placed the test data at `tests/data/` for automated tests when you run `pytest ./` at the root directory. You can actually copy the codes at [tests/test_multiomics.py](https://github.com/BioMeCIS-Lab/OpenOmics/blob/master/tests/test_multiomics.py) to load the -omics data.

> gtex = GTEx(path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/") Results in: OSError: Not enough free space in /homes/.astropy/cache/download/url to download a 3.6G file, only 2.7G left.

I used the package `astropy` to automatically cache downloaded files. It defaults to saving the files at `/homes/.astropy/cache/download/url`, and ideally it should be in one location for each user-session. But I can see how useful it is for the user to be able to make a setting to choose a directory of their choice - I will see about making an openomics configuration file at the user directory, located at `~/.openomics/conf.json`.
