### 2021-02-08 #31 @gawbul
Just in case it gets lost in the review, I submitted a PR for one of the issues I hit https://github.com/BioMeCIS-Lab/OpenOmics/pull/103
