### 2021-02-08 #31 @gawbul
Also submitted a PR for some issues I found in the pyOpenSci Contributing Guidelines https://github.com/pyOpenSci/contributing-guide/pull/85
