### 2021-02-08 BioMeCIS-Lab/OpenOmics#104 @gawbul
I initially opened an issue with `fsspec` here (https://github.com/intake/filesystem_spec/issues/529) but realised it was due to `gzip.open` here (https://github.com/BioMeCIS-Lab/OpenOmics/blob/master/openomics/database/base.py#L89-L90) returning an `io.TextIOWrapper` object.

`fsspec` is unable to get a path (see https://github.com/intake/filesystem_spec/blob/fb406453b6418052f98b64d405bd4e6a4be1def1/fsspec/utils.py#L304-L308) from the object and so it just returns it in its original state, which, of course, doesn't have a `startswith` method.
