### 2021-02-08 BioMeCIS-Lab/OpenOmics#104 @gawbul
Found this too https://stackoverflow.com/questions/65998183/python-dask-module-error-attributeerror-io-textiowrapper-object-has-no-at from a few days ago. Added a comment asking if they managed to figure it out.
