+++ requirements.txt
@@ -29,3 +29,4 @@ rarfile
 filetype
 goatools
 obonet
+tensorflow
