### 2021-02-08 BioMeCIS-Lab/OpenOmics#104 @gawbul
This might help? https://stackoverflow.com/q/39924518/393634 🤔
