### 2021-02-15 #31 @NickleDave
Thank you @gawbul for the very thorough review, really appreciate that you could

@JonnyTran just want to check where you are at with this. I know it might feel like a lot, and you could have other things going on.

Our guidelines suggest aiming for a two-week turnaround time after reviews are in. We definitely don't have to hold strictly to that especially if you have other obligations to deal with.

But please when you can just give me some idea of how you'll move forward. I would suggest converting reviewer comments into issues on OpenOmics and linking to them where you can. See for example issues on `physcraper` from their review: https://github.com/McTavishLab/physcraper/issues
