### 2021-02-16 #31 @JonnyTran
Thanks so much for the thorough review @gawbul! 🥰

@NickleDave I’ve been trying to go over the issues. Actually it has been difficult because of the constant rolling electricity blackouts from the snowstorm currently in my state, Texas 🥶.

I will work on making the issues on GH and fix them soon!
