### 2021-02-16 #31 @NickleDave
!!! I'm sorry, didn't realize you were in Texas!
The situation with the power grid is insane.
Hope you can stay warm and safe!

Thank you for letting us know!
