### 2021-02-16 BioMeCIS-Lab/OpenOmics#104 @JonnyTran
@gawbul Thank you so much for getting to the bottom of this issue. Amazing to see the detective work in action!

So originally, my intention was to have Dask handle the `read_table()`, since reading a large gtf file with Pandas' `read_table()` can have a huge memory footprint.

I've just tried `dask.dataframe.read_table()` with `compression="gzip"` on the compressed gzip file and it worked beautifully. Thanks for the suggestion!

Will be applying the fix and refactoring codes. I'll close this when it's done.
