### 2021-02-17 76bfe31
fixed read_gtf.py when loading uncompressed gzip files. Instead, let dask.dataframe handles it with read_table(compressed="gzip")
