### 2021-02-17 f7aea4e
fixed read_gtf.py when loading uncompressed gzip files. Instead, let dask.dataframe handles it with read_table(compressed="gzip")
