### 2021-02-17 4794937
added regex=True in pandas.DataFrame.str.replace()
