+++ openomics/database/sequence.py
@@ -113,8 +113,8 @@ class GENCODE(SequenceDataset):
             annotation_df = pd.concat(dfs)
 
         if self.remove_version_num:
-            annotation_df['gene_id'] = annotation_df['gene_id'].str.replace("[.].*", "")
-            annotation_df['transcript_id'] = annotation_df['transcript_id'].str.replace("[.].*", "")
+            annotation_df['gene_id'] = annotation_df['gene_id'].str.replace("[.].*", "", regex=True)
+            annotation_df['transcript_id'] = annotation_df['transcript_id'].str.replace("[.].*", "", regex=True)
 
         return annotation_df
 
