+++ openomics/utils/read_gtf.py
@@ -16,6 +16,7 @@ import logging
 from collections import OrderedDict
 from os.path import exists
 
+import dask
 import dask.dataframe as dd
 import numpy as np
 import pandas as pd
@@ -23,11 +24,7 @@ from six import string_types
 from six.moves import intern
 
 
-def expand_attribute_strings(
-    attribute_strings,
-    quote_char='\"',
-    missing_value="",
-    usecols=None):
+def expand_attribute_strings(attribute_strings, quote_char='\"', nan_value=None, usecols=None):
     """The last column of GTF has a variable number of key value pairs of the
     format: "key1 value1; key2 value2;" Parse these into a dictionary mapping
     each key onto a list of values, where the value is None for any row where
@@ -36,7 +33,7 @@ def expand_attribute_strings(
     Args:
         attribute_strings (list of str):
         quote_char (str): Quote character to remove from values
-        missing_value (any): If an attribute is missing from a row, give it this
+        nan_value (any): If an attribute is missing from a row, give it this
             value.
         usecols (list of str or None): If not None, then only expand columns
             included in this set, otherwise use all columns.
@@ -86,7 +83,7 @@ def expand_attribute_strings(
             try:
                 column = extra_columns[column_name]
             except KeyError:
-                column = [missing_value] * n
+                column = [nan_value] * n
                 extra_columns[column_name] = column
                 column_order.append(column_name)
 
@@ -101,7 +98,7 @@ def expand_attribute_strings(
             # if an attribute is used repeatedly then
             # keep track of all its values in a list
             old_value = column[i]
-            if old_value is missing_value:
+            if old_value is nan_value:
                 column[i] = value
             else:
                 column[i] = "%s,%s" % (old_value, value)
@@ -128,8 +125,93 @@ logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
-def parse_gtf(filepath_or_buffer, npartitions=None, compression=None, chunksize=1024 * 1024, features=None,
+def parse_gtf(filepath_or_buffer, chunksize=1024 * 1024, features=None,
               intern_columns=["seqname", "source", "strand", "frame"], fix_quotes_columns=["attribute"]):
+    """
+    Args:
+        filepath_or_buffer (str or buffer object):
+        chunksize (int): Default 1048576.
+        features (set or None): Drop entries which aren't one of these features
+        intern_columns (list): These columns are short strings which should be
+            interned
+        fix_quotes_columns (list): Most commonly the 'attribute' column which
+            had broken quotes on some Ensembl release GTF files.
+    """
+    if features is not None:
+        features = set(features)
+
+    def parse_frame(s):
+        if s == ".":
+            return 0
+        else:
+            return int(s)
+
+    # GTF columns:
+    # 1) seqname: str ("1", "X", "chrX", etc...)
+    # 2) source : str
+    #      Different versions of GTF use second column as of:
+    #      (a) gene biotype
+    #      (b) transcript biotype
+    #      (c) the annotation source
+    #      See: https://www.biostars.org/p/120306/#120321
+    # 3) feature : str ("gene", "transcript", &c)
+    # 4) start : int
+    # 5) end : int
+    # 6) score : float or "."
+    # 7) strand : "+", "-", or "."
+    # 8) frame : 0, 1, 2 or "."
+    # 9) attribute : key-value pairs separated by semicolons
+    # (see more complete description in docstring at top of file)
+
+    chunk_iterator = pd.read_csv(
+        filepath_or_buffer,
+        sep="\t",
+        comment="#",
+        names=REQUIRED_COLUMNS,
+        skipinitialspace=True,
+        skip_blank_lines=True,
+        error_bad_lines=True,
+        warn_bad_lines=True,
+        chunksize=chunksize,
+        engine="c",
[…]
