### 2021-02-17 ced7504
Format code with yapf and autopep8
This commit fixes the style issues introduced in 5ec9d93 according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/26e10218-bb28-4637-b97b-b177d9228b87/
