+++ openomics/database/sequence.py
@@ -9,8 +9,8 @@ import openomics
 from openomics.utils.read_gtf import read_gtf
 from .base import Dataset
 
-
 # from gtfparse import read_gtf
+
 
 class SequenceDataset(Dataset):
     def __init__(self, replace_U2T=False, **kwargs):
@@ -60,20 +60,33 @@ class SequenceDataset(Dataset):
             agg:
         """
         if agg == "all":
-            agg_func = lambda x: list(x) if not isinstance(x, str) else x
+
+            def agg_func(x):
+                return list(x) if not isinstance(x, str) else x
         elif agg == "shortest":
-            agg_func = lambda x: min(x, key=len)
+
+            def agg_func(x):
+                return min(x, key=len)
         elif agg == "longest":
-            agg_func = lambda x: max(x, key=len)
+
+            def agg_func(x):
+                return max(x, key=len)
         else:
-            raise Exception("agg_sequences argument must be one of {'all', 'shortest', 'longest'}")
+            raise Exception(
+                "agg_sequences argument must be one of {'all', 'shortest', 'longest'}"
+            )
         return agg_func
 
 
 class GENCODE(SequenceDataset):
-    def __init__(self, path='ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/',
-                 file_resources=None, col_rename=None, npartitions=0,
-                 replace_U2T=False, remove_version_num=False):
+    def __init__(
+            self,
+            path='ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/',
+            file_resources=None,
+            col_rename=None,
+            npartitions=0,
+            replace_U2T=False,
+            remove_version_num=False):
         """
         Args:
             path:
@@ -86,14 +99,20 @@ class GENCODE(SequenceDataset):
         if file_resources is None:
             file_resources = {
                 "basic.annotation.gtf": "gencode.v32.basic.annotation.gtf.gz",
-                "long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
-                "lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz",
-                "transcripts.fa": "gencode.v32.transcripts.fa.gz"}
+                "long_noncoding_RNAs.gtf":
+                "gencode.v32.long_noncoding_RNAs.gtf.gz",
+                "lncRNA_transcripts.fa":
+                "gencode.v32.lncRNA_transcripts.fa.gz",
+                "transcripts.fa": "gencode.v32.transcripts.fa.gz"
+            }
 
         self.remove_version_num = remove_version_num
 
-        super(GENCODE, self).__init__(path=path, file_resources=file_resources, col_rename=col_rename,
-                                      replace_U2T=replace_U2T, npartitions=npartitions)
+        super(GENCODE, self).__init__(path=path,
+                                      file_resources=file_resources,
+                                      col_rename=col_rename,
+                                      replace_U2T=replace_U2T,
+                                      npartitions=npartitions)
 
     def load_dataframe(self, file_resources, npartitions=None):
         """
@@ -104,7 +123,9 @@ class GENCODE(SequenceDataset):
         dfs = []
         for filename, content in file_resources.items():
             if '.gtf' in filename:
-                df = read_gtf(content, npartitions=npartitions, compression="gzip")
+                df = read_gtf(content,
+                              npartitions=npartitions,
+                              compression="gzip")
                 dfs.append(df)
 
         if npartitions:
@@ -113,8 +134,10 @@ class GENCODE(SequenceDataset):
             annotation_df = pd.concat(dfs)
 
         if self.remove_version_num:
-            annotation_df['gene_id'] = annotation_df['gene_id'].str.replace("[.].*", "", regex=True)
[…]
