### 2021-02-17 9e7b74a
Format code with yapf and autopep8
This commit fixes the style issues introduced in f51820e according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/923d62bf-5aa4-47da-92e8-edcc91e4283a/
