### 2021-02-17 5e88899
Format code with yapf and autopep8
This commit fixes the style issues introduced in 581e961 according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/95fb0582-98f9-4b33-bd71-dc87a02f6d4f/
