### 2021-02-17 2a338bc
Format code with yapf and autopep8
This commit fixes the style issues introduced in eacf4ac according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/6220cab6-de90-4b53-8d31-e460e7b8ab52/
