### 2021-02-17 b254239
Format code with yapf and autopep8
This commit fixes the style issues introduced in 593975a according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/4a392364-ef8a-4c59-b608-fe31d8f99a72/
