+++ openomics/utils/read_gtf.py
@@ -24,7 +24,10 @@ from six import string_types
 from six.moves import intern
 
 
-def expand_attribute_strings(attribute_strings, quote_char='\"', nan_value=None, usecols=None):
+def expand_attribute_strings(attribute_strings,
+                             quote_char='\"',
+                             nan_value=None,
+                             usecols=None):
     """The last column of GTF has a variable number of key value pairs of the
     format: "key1 value1; key2 value2;" Parse these into a dictionary mapping
     each key onto a list of values, where the value is None for any row where
@@ -87,7 +90,8 @@ def expand_attribute_strings(attribute_strings, quote_char='\"', nan_value=None,
                 extra_columns[column_name] = column
                 column_order.append(column_name)
 
-            value = value.replace(quote_char, "") if value.startswith(quote_char) else value
+            value = value.replace(
+                quote_char, "") if value.startswith(quote_char) else value
 
             try:
                 value = value_interned_strings[value]
@@ -103,9 +107,8 @@ def expand_attribute_strings(attribute_strings, quote_char='\"', nan_value=None,
             else:
                 column[i] = "%s,%s" % (old_value, value)
 
-    return OrderedDict(
-        (column_name, extra_columns[column_name])
-        for column_name in column_order)
+    return OrderedDict((column_name, extra_columns[column_name])
+                       for column_name in column_order)
 
 
 REQUIRED_COLUMNS = [
@@ -124,8 +127,11 @@ logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
-def parse_gtf(filepath_or_buffer, chunksize=1024 * 1024, features=None,
-              intern_columns=["seqname", "source", "strand", "frame"], fix_quotes_columns=["attribute"]):
+def parse_gtf(filepath_or_buffer,
+              chunksize=1024 * 1024,
+              features=None,
+              intern_columns=["seqname", "source", "strand", "frame"],
+              fix_quotes_columns=["attribute"]):
     """
     Args:
         filepath_or_buffer (str or buffer object):
@@ -162,25 +168,24 @@ def parse_gtf(filepath_or_buffer, chunksize=1024 * 1024, features=None,
     # 9) attribute : key-value pairs separated by semicolons
     # (see more complete description in docstring at top of file)
 
-    chunk_iterator = pd.read_csv(
-        filepath_or_buffer,
-        sep="\t",
-        comment="#",
-        names=REQUIRED_COLUMNS,
-        skipinitialspace=True,
-        skip_blank_lines=True,
-        error_bad_lines=True,
-        warn_bad_lines=True,
-        chunksize=chunksize,
-        engine="c",
-        dtype={
-            "start": np.int64,
-            "end": np.int64,
-            "score": np.float32,
-            "seqname": str,
-        },
-        na_values=".",
-        converters={"frame": parse_frame})
+    chunk_iterator = pd.read_csv(filepath_or_buffer,
+                                 sep="\t",
+                                 comment="#",
+                                 names=REQUIRED_COLUMNS,
+                                 skipinitialspace=True,
+                                 skip_blank_lines=True,
+                                 error_bad_lines=True,
+                                 warn_bad_lines=True,
+                                 chunksize=chunksize,
+                                 engine="c",
+                                 dtype={
+                                     "start": np.int64,
+                                     "end": np.int64,
+                                     "score": np.float32,
+                                     "seqname": str,
+                                 },
+                                 na_values=".",
+                                 converters={"frame": parse_frame})
 
     dataframes = []
     try:
@@ -210,7 +215,11 @@ def parse_gtf(filepath_or_buffer, chunksize=1024 * 1024, features=None,
 
[…]
