+++ openomics/utils/GTF.py
@@ -1,146 +1,0 @@
-#!/usr/bin/env python
-"""
-GTF.py
-Kamil Slowikowski
-December 24, 2013
-
-Read GFF/GTF files. Works with gzip compressed files and pandas.
-
-    http://useast.ensembl.org/info/website/upload/gff.html
-
-LICENSE
-
-This is free and unencumbered software released into the public domain.
-Anyone is free to copy, modify, publish, use, compile, sell, or
-distribute this software, either in source code form or as a compiled
-binary, for any purpose, commercial or non-commercial, and by any
-means.
-
-In jurisdictions that recognize copyright laws, the author or authors
-of this software dedicate any and all copyright interest in the
-software to the public domain. We make this dedication for the benefit
-of the public at large and to the detriment of our heirs and
-successors. We intend this dedication to be an overt act of
-relinquishment in perpetuity of all present and future rights to this
-software under copyright law.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR
-OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
-ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-OTHER DEALINGS IN THE SOFTWARE.
-
-For more information, please refer to <http://unlicense.org/>
-"""
-
-
-import gzip
-import re
-from collections import defaultdict
-
-import dask.dataframe as dd
-import pandas as pd
-
-GTF_HEADER  = ['seqname', 'source', 'feature', 'start', 'end', 'score',
-               'strand', 'frame']
-R_SEMICOLON = re.compile(r'\s*;\s*')
-R_COMMA     = re.compile(r'\s*,\s*')
-R_KEYVALUE  = re.compile(r'(\s+|\s*=\s*)')
-
-
-def dataframe(filename, npartitions):
-    """Open an optionally gzipped GTF file and return a pandas.DataFrame.
-
-    Args:
-        filename:
-        npartitions:
-    """
-    # Each column is a list stored as a value in this dict.
-    result = defaultdict(list)
-
-    for i, line in enumerate(lines(filename)):
-        for key in line.keys():
-            # This key has not been seen yet, so set it to None for all
-            # previous lines.
-            if key not in result:
-                result[key] = [None] * i
-
-        # Ensure this row has some value for each column.
-        for key in result.keys():
-            result[key].append(line.get(key, None))
-
-    df = pd.DataFrame(result)
-    if npartitions:
-        return dd.from_pandas(df)
-    else:
-        return df
-
-
-def lines(filename):
-    """Open an optionally gzipped GTF file and generate a dict for each line.
-
-    Args:
-        filename:
-    """
-    fn_open = gzip.open if filename.endswith('.gz') else open
-
-    with fn_open(filename) as fh:
-        for line in fh:
-            if line.startswith('#'):
-                continue
-            else:
-                yield parse(line)
-
-
-def parse(line):
-    """Parse a single GTF line and return a dict.
-
-    Args:
-        line:
-    """
-    result = {}
-
-    fields = line.rstrip().split('\t')
-
-    for i, col in enumerate(GTF_HEADER):
-        result[col] = _get_value(fields[i])
-
-    # INFO field consists of "key1=value;key2=value;...".
[…]
