### 2021-02-17 ad68a54
Format code with yapf and autopep8
This commit fixes the style issues introduced in 2e89102 according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/c864ce3e-6073-4855-9d63-1cb7c7e428f2/
