### 2021-02-17 BioMeCIS-Lab/OpenOmics#104 @JonnyTran
I added this line [openomics/database/base.py#L74](https://github.com/BioMeCIS-Lab/OpenOmics/blob/2e891028d9df0af6ab38b65b05dbdcd7b906cfdd/openomics/database/base.py#L74) which now lets dask handles uncompression when dealing with GTF files. Running `GENCODE(npartitions=5)` now works.

I also added some functionalities to parse attributes from GTF files into Dask dataframes at `openomics/utils/read_gtf.py`. Now, creating dask dataframes from GTF files is working (albeit no time improvement) with `read_gtf(npartitions=npartitions)` where `npartitions >0`. In the future, will look into optimizing GTF attributes parsing by using ddf.map_partitions(func) or ddf["attributes"].apply(func).
