### 2021-02-17 2ef4044
removed umap from requirements.txt
