### 2021-02-17 5c96c31
Format code with autopep8 and yapf
This commit fixes the style issues introduced in 3323d27 according to the output
from autopep8 and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/e434780b-1ebc-4179-afd1-d53ef44e5ff8/
