### 2021-02-17 BioMeCIS-Lab/OpenOmics#115 @JonnyTran
fixed the wrong `path` at `MiRTarBase(path="/data/datasets/Bioinformatics_ExternalData/miRTarBase/")` which was previously hard-coded, with this commit https://github.com/BioMeCIS-Lab/OpenOmics/commit/d143ddf43d512018bb280f513f3b4256aa53c50e
