### 2021-02-17 BioMeCIS-Lab/OpenOmics#117 @JonnyTran
Review 2 - Automated tests - Failed builds in Travis CI
* openOmics version: 0.8.4
* Python version: 3.5, 3.7, 3.8, 3.9
* Operating System: Linux

### Description

When raising a PR the build in Travis CI failed for all python versions. There is some work needed to get these issues resolved though I didn't inspect the output in detail. See https://travis-ci.com/github/gawbul/OpenOmics/builds/216404830
