### 2021-02-17 BioMeCIS-Lab/OpenOmics#117 @JonnyTran
I've been having trouble with the automated tests on Travis CI as I've ran out of free credits. Currently it's not automatically running until I've set up a paid subscription. Will look into this further as it's very important to have OpenOmics be compatible with Python 3.5, 3.7, 3.8, 3.9 as well as across Mac OS, Windows, and Linux.
