### 2021-02-17 46dea67
Format code with autopep8 and yapf
This commit fixes the style issues introduced in d862fb3 according to the output
from autopep8 and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/a079907e-f9b7-49e4-a31c-50430785d5ff/
