### 2021-02-17 BioMeCIS-Lab/OpenOmics#114 @JonnyTran
As a temporary fix, I've removed `umap-learn` from `requirements.txt` and removed the `import umap` as currently openomics doesn't yet need the visualization library.

See this commit https://github.com/BioMeCIS-Lab/OpenOmics/commit/2ef40446eb62b19850d7736e1204668b15a96b11
