### 2021-02-17 BioMeCIS-Lab/OpenOmics#115 @JonnyTran
Oops my mistake, did not actually fix this issue with the commit d143ddf
