### 2021-02-17 BioMeCIS-Lab/OpenOmics#104 @gawbul
Awesome 🥳 Glad you managed to get this sorted 🙂
