### 2021-02-17 BioMeCIS-Lab/OpenOmics#115 @JonnyTran
The path at MiRTarBase(path="/data/datasets/Bioinformatics_ExternalData/miRTarBase/") is hard-coded to the directory at my local computer.

There are many bioinformatics datasets that are not publicly available via direct download (e.g. lncbase) and may require access permission. In that case, it is expected that the users would obtain the data themself then provide a `path` to their own local directory.

Since currently, it is difficult to get a reliable ftp download connection from the server at http://mirtarbase.mbc.nctu.edu.tw, I suggest we remove `test_import_MiRTarBase` from the automated test set that relies on ftp download.
