### 2021-02-17 #31 @JonnyTran
Hi @gawbul, I've just gone over your reviews. Thanks for the care and attention on testing this software. There were many issues that I've missed and I've created Issues for most of your comments.

> I had several issues with installation as described above, though I only tested on a macOS system running Big Sur version 11.2. It appears there are issues with Python 3.9.x in particular, though I can't confirm whether this is platform specific. There were some failures in building scipy on Python 3.8.7 due to dependent libraries missing on my machine.

I'm sorry you've had troubles running OpenOmics because of issues with installing the package dependencies from `requirement.txt` https://github.com/BioMeCIS-Lab/OpenOmics/issues/113, and importing `umap` when running the SomaticMutation vignettes https://github.com/BioMeCIS-Lab/OpenOmics/issues/114.
I have not seen these errors before as I've never ran tests on Mac OS X with Python 3.9. I've primarily been developing with the Anaconda Python 3.7 environment which already comes with a `llvm` installation, so the `pip install openomics` did not show any errors. Since this is still an issue on Mac OS X, I try to use more TravisCI automated tests against a variety of different Python versions and debug these problems. I've made an issue at https://github.com/BioMeCIS-Lab/OpenOmics/issues/117.

> Docstrings seem to be available throughout the codebase for all relevant functions, however, I found the documentation on Read the Docs to be lacking in comparison, particularly in providing detail of function parameters that are otherwise available in the docstrings.

Currently documentations on Read the Docs has mostly been auto-generated by Sphinx. As pointed out by both @gawbul and @lwasser , I agree that Read the Docs documentations could be better, particularly Usage guides and Vignettes. Also, it is a great suggestion to use only markdown or reStructredText (rather than both). I've opened the issue https://github.com/BioMeCIS-Lab/OpenOmics/issues/119.

> Running the example under Annotate LncRNAs with GENCODE genomic annotations returns the following:
> AttributeError: '_io.TextIOWrapper' object has no attribute 'startswith'
This issue has been fixed at https://github.com/BioMeCIS-Lab/OpenOmics/issues/104

There were many other issues listed at https://github.com/BioMeCIS-Lab/OpenOmics/issues. Is there a deadline to address all these? I think the issues with Read the Docs and automated tests targeting MacOS + Python3.9 can take a week to finish.
