### 2021-02-17 d1c32fd
fixed FutureWarning's when using pandas.Series.str.replace(regex=True)
