### 2021-02-17 1fabe6c
fixed FutureWarning's when using pandas.Series.str.replace(regex=True)
