+++ openomics/transcriptomics.py
@@ -105,7 +105,7 @@ class ExpressionData(object):
             data.seek(0)  # Needed since the file was previous read to extract columns information
             df = pd.read_table(data)
         elif type(data) is str and os.path.isfile(data):
-            df = pd.read_table(data, sep=None)
+            df = pd.read_table(data, sep=None, engine='python')
         else:
             raise IOError(data)
 
