### 2021-02-17 f7f2fd9
fixed FutureWarning's when using pandas.Series.str.replace(regex=True)
