### 2021-02-18 6cc49c0
Format code with yapf and autopep8
This commit fixes the style issues introduced in 4c9aaf8 according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/57cdef17-4b28-494b-a11d-dd024ebdef27/
