### 2021-02-18 458dd95
Remove length check in favour of truthiness of the object
