+++ openomics_web/app.py
@@ -105,7 +105,7 @@ def update_table(page_current, page_size, sort_by, filter):
             # only works with complete fields in standard format
             dff = dff.loc[dff[col_name].str.startswith(filter_value)]
 
-    if len(sort_by):
+    if sort_by:
         dff = dff.sort_values(
             [col['column_id'] for col in sort_by],
             ascending=[
