### 2021-02-18 c13d2c3
Format code with yapf and autopep8
