+++ openomics_web/app.py
@@ -20,17 +20,29 @@ app.layout = app_layout.app_main()
 
 user_multiomics = MultiOmics(cohort_name="TEST", )
 
-@app.callback([Output('data-table-column-select', 'children'), Output('upload-data-table', 'children')],
-              [Input('upload-data-table', 'contents'), Input('upload-data-table', 'filename')],
-              [State('data-table-type', 'value'), ])
-def update_datatable_metadata(list_of_contents, list_of_names, data_type, ):
+
+@app.callback([
+    Output('data-table-column-select', 'children'),
+    Output('upload-data-table', 'children')
+], [
+    Input('upload-data-table', 'contents'),
+    Input('upload-data-table', 'filename')
+], [
+    State('data-table-type', 'value'),
+])
+def update_datatable_metadata(
+    list_of_contents,
+    list_of_names,
+    data_type,
+):
     """
     Args:
         list_of_contents:
         list_of_names:
         data_type:
     """
-    if list_of_contents is None: return None, ['Drag and Drop or ', html.A('Select Files')]
+    if list_of_contents is None:
+        return None, ['Drag and Drop or ', html.A('Select Files')]
 
     try:
         columns = get_table_columns(list_of_contents, list_of_names)
@@ -43,14 +55,18 @@ def update_datatable_metadata(list_of_contents, list_of_names, data_type, ):
 
 
 @app.callback(Output('output-data-upload', 'children'),
-              [Input('upload-data-table-submit', 'n_clicks')],
-              [State('data-table-cohort', 'value'), State('data-table-type', 'value'),
-               State('upload-data-table', 'contents'), State('upload-data-table', 'filename'),
-               State('data-table-genes-col-name', 'value'), State('data-table-columns-select', 'value'),
-               State('data-table-transpose', 'value')
-               ])
-def import_datatable_upload(n_clicks, cohort_name, data_type, list_of_contents, list_of_names, genes_col_name,
-                            columns_select, transposed):
+              [Input('upload-data-table-submit', 'n_clicks')], [
+                  State('data-table-cohort', 'value'),
+                  State('data-table-type', 'value'),
+                  State('upload-data-table', 'contents'),
+                  State('upload-data-table', 'filename'),
+                  State('data-table-genes-col-name', 'value'),
+                  State('data-table-columns-select', 'value'),
+                  State('data-table-transpose', 'value')
+])
+def import_datatable_upload(n_clicks, cohort_name, data_type, list_of_contents,
+                            list_of_names, genes_col_name, columns_select,
+                            transposed):
     """
     Args:
         n_clicks:
@@ -62,11 +78,14 @@ def import_datatable_upload(n_clicks, cohort_name, data_type, list_of_contents, 
         columns_select:
         transposed:
     """
-    if list_of_contents is None: return []
+    if list_of_contents is None:
+        return []
     try:
 
-        omics_data = get_expression_data(list_of_contents, list_of_names, data_type, cohort_name, genes_col_name,
-                                         columns_select, transposed)
+        omics_data = get_expression_data(list_of_contents, list_of_names,
+                                         data_type, cohort_name,
+                                         genes_col_name, columns_select,
+                                         transposed)
         user_multiomics.add_omic(omics_data)
     except Exception as e:
         print(e)
@@ -75,12 +94,12 @@ def import_datatable_upload(n_clicks, cohort_name, data_type, list_of_contents, 
     return ExpressionDataTable(omics_data.expressions.head(20))
 
 
-@app.callback(
-    Output('expression-datatable', "data"),
-    [Input('expression-datatable', "page_current"),
-     Input('expression-datatable', "page_size"),
-     Input('expression-datatable', "sort_by"),
[…]
