### 2021-02-18 861b50b
Use literal syntax instead of function calls to create data structure
