### 2021-02-18 c0684ce
Format code with yapf and autopep8
This commit fixes the style issues introduced in 3fad0df according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/9512559e-84ef-4e05-a5aa-a5a59ead6e04/
