+++ .deepsource.toml
@@ -12,5 +12,5 @@ enabled = true
   runtime_version = "3.x.x"
 
 [[transformers]]
-name = "flake8"
+name = "black"
 enabled = true
