### 2021-02-18 ee39b5d
Format code with black
This commit fixes the style issues introduced in be46641 according to the output
from black.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/5e9f9428-8fef-4e56-b41d-0166805d0b3b/
