+++ openomics_web/layouts/datatable_view.py
@@ -12,26 +12,30 @@ def DataTableColumnSelect(columns):
     """
     longest_common_prefixes = longest_common_prefix(columns)
 
-    return html.Div([
-        html.Div(['Select the gene id/name column to index by:']),
-        dcc.Dropdown(
-            id='data-table-genes-col-name',
-            options=[{'label': col, 'value': col} for col in columns],
-            style={
-                'width': '100%',
-            },
-            value=columns[0],
-        ),
-        html.Div(['Select the column prefixes to import:']),
-        dcc.Dropdown(
-            id='data-table-columns-select',
-            options=[{'label': col, 'value': col} for col in longest_common_prefixes],
-            style={
-                'width': '100%',
-            },
-            multi=True,
-        )
-    ])
+    return html.Div(
+        [
+            html.Div(["Select the gene id/name column to index by:"]),
+            dcc.Dropdown(
+                id="data-table-genes-col-name",
+                options=[{"label": col, "value": col} for col in columns],
+                style={
+                    "width": "100%",
+                },
+                value=columns[0],
+            ),
+            html.Div(["Select the column prefixes to import:"]),
+            dcc.Dropdown(
+                id="data-table-columns-select",
+                options=[
+                    {"label": col, "value": col} for col in longest_common_prefixes
+                ],
+                style={
+                    "width": "100%",
+                },
+                multi=True,
+            ),
+        ]
+    )
 
 
 def ExpressionDataTable(df):
@@ -44,66 +48,59 @@ def ExpressionDataTable(df):
         children=[
             html.Div(
                 dt.DataTable(
-                    id='expression-datatable',
+                    id="expression-datatable",
                     columns=[{"name": i, "id": i} for i in df.columns],
                     page_current=0,
                     page_size=20,
-                    page_action='custom',
-
-                    filter_action='custom',
-                    filter_query='',
-
-                    sort_action='custom',
-                    sort_mode='multi',
+                    page_action="custom",
+                    filter_action="custom",
+                    filter_query="",
+                    sort_action="custom",
+                    sort_mode="multi",
                     sort_by=[],
-
                     style_as_list_view=True,
                     style_cell={
-                        'overflow': 'hidden',
-                        'textOverflow': 'clip',
-                        'whiteSpace': 'normal'
+                        "overflow": "hidden",
+                        "textOverflow": "clip",
+                        "whiteSpace": "normal",
                     },
-                    style_data={'width': '30px'},
+                    style_data={"width": "30px"},
                     style_data_conditional=[
-                        {'if': {'row_index': 'odd'},
-                         'backgroundColor': 'rgb(248, 248, 248)'
-                         },
+                        {
+                            "if": {"row_index": "odd"},
+                            "backgroundColor": "rgb(248, 248, 248)",
+                        },
                     ],
-                    style_table={"maxHeight": '800px',
-                                 'width': '800px',
-                                 'marginTop': '5px',
-                                 'marginBottom': '10px',
-                                 'overflowX': 'scroll'
-                                 },
-                    style_header={
-                        'backgroundColor': 'white',
-                        'fontWeight': 'bold'
+                    style_table={
+                        "maxHeight": "800px",
+                        "width": "800px",
+                        "marginTop": "5px",
+                        "marginBottom": "10px",
+                        "overflowX": "scroll",
                     },
-
+                    style_header={"backgroundColor": "white", "fontWeight": "bold"},
[…]

[…]
