### 2021-02-18 9ebd3af
Format code with black and yapf
This commit fixes the style issues introduced in 2e3b101 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/b6b53115-5e15-4d13-abdc-321bf21c3a7b/
