+++ openomics/visualization/heatmat.py
@@ -2,7 +2,12 @@ import pandas as pd
 import plotly.graph_objects as go
 
 
-def heatmap(table, file_output=None, title=None, autosize=True, width=800, height=1000):
+def heatmap(table,
+            file_output=None,
+            title=None,
+            autosize=True,
+            width=800,
+            height=1000):
     """
     Args:
         table:
@@ -13,14 +18,16 @@ def heatmap(table, file_output=None, title=None, autosize=True, width=800, heigh
         height:
     """
     if type(table.columns) == pd.MultiIndex:
-        columns = table.columns.to_series().apply(lambda x: '{0}-{1}'.format(*x))
+        columns = table.columns.to_series().apply(
+            lambda x: "{0}-{1}".format(*x))
     else:
         columns = table.columns
     fig = go.Figure(data=go.Heatmap(
         z=table,
         x=columns,
         y=table.index,
-        hoverongaps=False, ))
+        hoverongaps=False,
+    ))
     fig.update_layout(
         title=title,
         autosize=autosize,

[…]
