### 2021-02-18 0e3b947
Change methods not using its bound instance to staticmethods
