+++ openomics/database/base.py
@@ -14,7 +14,14 @@ from openomics.utils.io import get_pkg_data_filename
 class Dataset(object):
     COLUMNS_RENAME_DICT = None  # Needs initialization since subclasses may use this field to rename columns in dataframes.
 
-    def __init__(self, path, file_resources=None, col_rename=None, npartitions=None, verbose=False):
+    def __init__(
+        self,
+        path,
+        file_resources=None,
+        col_rename=None,
+        npartitions=None,
+        verbose=False,
+    ):
         """This is a class used to instantiate a Dataset given a a set of files from either local files or URLs. When
         creating a Dataset class, the `load_dataframe()` function is called where the file_resources are used to
         load (Pandas or Dask) DataFrames, then performs data wrangling to yield a dataframe at `self.data`. This class
@@ -31,9 +38,13 @@ class Dataset(object):
         self.npartitions = npartitions
         self.verbose = verbose
 
-        self.validate_file_resources(path, file_resources, npartitions=npartitions, verbose=verbose)
+        self.validate_file_resources(path,
+                                     file_resources,
+                                     npartitions=npartitions,
+                                     verbose=verbose)
 
-        self.data = self.load_dataframe(file_resources, npartitions=npartitions)
+        self.data = self.load_dataframe(file_resources,
+                                        npartitions=npartitions)
         self.data = self.data.reset_index()
         if col_rename is not None:
             self.data = self.data.rename(columns=col_rename)
@@ -43,7 +54,11 @@ class Dataset(object):
     def info(self):
         print("{}: {}".format(self.name(), self.data.columns.tolist()))
 
-    def validate_file_resources(self, path, file_resources, npartitions=None, verbose=False) -> None:
+    def validate_file_resources(self,
+                                path,
+                                file_resources,
+                                npartitions=None,
+                                verbose=False) -> None:
         """For each file in file_resources, download the file if path+file is a URL
         or load from disk if a local path. Additionally unzip or unrar if the
         file is compressed.
@@ -62,8 +77,9 @@ class Dataset(object):
         """
         if validators.url(path):
             for filename, filepath in copy.copy(file_resources).items():
-                data_file = get_pkg_data_filename(path, filepath,
-                                                  verbose=verbose)  # Download file and replace the file_resource path
+                data_file = get_pkg_data_filename(
+                    path, filepath, verbose=verbose
+                )  # Download file and replace the file_resource path
                 filepath_ext = filetype.guess(data_file)
 
                 # This null if-clause is needed incase when filetype_ext is None, causing the next clauses to fail
@@ -74,25 +90,31 @@ class Dataset(object):
                 elif ".gtf" in filename and npartitions:
                     file_resources[filename] = data_file
 
-                elif filepath_ext.extension == 'gz':
+                elif filepath_ext.extension == "gz":
                     logging.debug(f"Uncompressed gzip file at {data_file}")
-                    file_resources[filename] = gzip.open(data_file, 'rt')
+                    file_resources[filename] = gzip.open(data_file, "rt")
 
-                elif filepath_ext.extension == 'zip':
+                elif filepath_ext.extension == "zip":
                     logging.debug(f"Uncompressed zip file at {data_file}")
-                    zf = zipfile.ZipFile(data_file, 'r')
+                    zf = zipfile.ZipFile(data_file, "r")
 
                     for subfile in zf.infolist():
-                        if os.path.splitext(subfile.filename)[-1] == os.path.splitext(filename)[-1]: # If the file extension matches
-                            file_resources[filename] = zf.open(subfile.filename, mode='r')
[…]
