+++ openomics/database/ontology.py
@@ -11,7 +11,12 @@ from ..utils.df import slice_adj
 class Ontology(Dataset):
     DELIM = "|"
 
-    def __init__(self, path, file_resources=None, col_rename=None, npartitions=0, verbose=False):
+    def __init__(self,
+                 path,
+                 file_resources=None,
+                 col_rename=None,
+                 npartitions=0,
+                 verbose=False):
         """
         Manages dataset input processing from tables and construct an ontology network from .obo file. There ontology
         network is G(V,E) where there exists e_ij for child i to parent j to present "node i is_a node j".
@@ -25,8 +30,13 @@ class Ontology(Dataset):
         """
         self.network, self.node_list = self.load_network(file_resources)
 
-        super(Ontology, self).__init__(path=path, file_resources=file_resources, col_rename=col_rename,
-                                       npartitions=npartitions, verbose=verbose)
+        super(Ontology, self).__init__(
+            path=path,
+            file_resources=file_resources,
+            col_rename=col_rename,
+            npartitions=npartitions,
+            verbose=verbose,
+        )
 
     def load_network(self, file_resources) -> (nx.MultiDiGraph, list):
         raise NotImplementedError
@@ -41,7 +51,8 @@ class Ontology(Dataset):
         if node_list is None or list(node_list) == list(self.node_list):
             return adjacency_matrix
         elif set(node_list) < set(self.node_list):
-            return slice_adj(adjacency_matrix, list(self.node_list), node_list, None)
+            return slice_adj(adjacency_matrix, list(self.node_list), node_list,
+                             None)
         elif not (set(node_list) < set(self.node_list)):
             raise Exception("A node in node_list is not in self.node_list.")
 
@@ -52,8 +63,8 @@ class Ontology(Dataset):
 
     def filter_annotation(self, annotation: pd.Series):
         go_terms = set(self.node_list)
-        filtered_annotation = annotation.map(
-            lambda x: list(set(x) & go_terms) if isinstance(x, list) else [])
+        filtered_annotation = annotation.map(lambda x: list(set(x) & go_terms)
+                                             if isinstance(x, list) else [])
 
         return filtered_annotation
 
@@ -94,12 +105,14 @@ class Ontology(Dataset):
         if not annotation.map(lambda x: isinstance(x, list)).any():
             annotation = annotation.str.split(self.DELIM)
 
-        go_terms_parents = annotation.map(
-            lambda x: list(set(x) & set(leaf_terms)) if isinstance(x, list) else None)
+        go_terms_parents = annotation.map(lambda x: list(
+            set(x) & set(leaf_terms)) if isinstance(x, list) else None)
         return go_terms_parents
 
     @staticmethod
-    def get_node_color(file="~/Bioinformatics_ExternalData/GeneOntology/go_colors_biological.csv"):
+    def get_node_color(
+        file="~/Bioinformatics_ExternalData/GeneOntology/go_colors_biological.csv",
+    ):
         go_colors = pd.read_csv(file)
 
         def selectgo(x):
@@ -109,10 +122,15 @@ class Ontology(Dataset):
             else:
                 return None
 
-        go_colors["node"] = go_colors[[col for col in go_colors.columns if col.isdigit()]].apply(selectgo, axis=1)
-        go_id_colors = go_colors[go_colors["node"].notnull()].set_index("node")["HCL.color"]
-        go_id_colors = go_id_colors[~go_id_colors.index.duplicated(keep="first")]
-        print(go_id_colors.unique().shape, go_colors["HCL.color"].unique().shape)
+        go_colors["node"] = go_colors[[
[…]

[…]
