### 2021-02-18 1e24b1c
Fix dangerous default argument
