+++ openomics/database/disease.py
@@ -12,7 +12,8 @@ class DiseaseAssociation(Dataset):
             file_resources:
             **kwargs:
         """
-        super(DiseaseAssociation, self).__init__(path, file_resources, **kwargs)
+        super(DiseaseAssociation, self).__init__(path, file_resources,
+                                                 **kwargs)
 
     @abstractmethod
     def get_disease_assocs(self, index="gene_name"):
@@ -28,10 +29,16 @@ class OMIM(DiseaseAssociation):
 
 
 class MalaCards(DiseaseAssociation):
-    COLUMNS_RENAME_DICT = {"geneSymbol": "gene_name", "maladyMainName": "disease_associations"}
+    COLUMNS_RENAME_DICT = {
+        "geneSymbol": "gene_name",
+        "maladyMainName": "disease_associations",
+    }
 
-    def __init__(self, path="http://zdzlab.einstein.yu.edu/1/hedd/", file_resources=None,
-                 col_rename=None, **kwargs):
+    def __init__(self,
+                 path="http://zdzlab.einstein.yu.edu/1/hedd/",
+                 file_resources=None,
+                 col_rename=None,
+                 **kwargs):
         """
         Args:
             path:
@@ -43,9 +50,13 @@ class MalaCards(DiseaseAssociation):
             col_rename = COLUMNS_RENAME_DICT
         if file_resources is None:
             file_resources = {}
-            file_resources["MalaCards.csv"] = "download.action.php?filename=DataDownload/MalaCards.csv"
+            file_resources[
+                "MalaCards.csv"] = "download.action.php?filename=DataDownload/MalaCards.csv"
 
-        super(MalaCards, self).__init__(path, file_resources, col_rename=col_rename, **kwargs)
+        super(MalaCards, self).__init__(path,
+                                        file_resources,
+                                        col_rename=col_rename,
+                                        **kwargs)
 
     def load_dataframe(self, file_resources, npartitions=None):
         # type: (dict, int) -> pd.DataFrame
@@ -59,12 +70,18 @@ class MalaCards(DiseaseAssociation):
 
 
 class DisGeNet(DiseaseAssociation):
-    COLUMNS_RENAME_DICT = {"geneSymbol": "gene_name",
-                           "diseaseName": "disease_associations"}
+    COLUMNS_RENAME_DICT = {
+        "geneSymbol": "gene_name",
+        "diseaseName": "disease_associations",
+    }
 
-    def __init__(self, path="https://www.disgenet.org/static/disgenet_ap1/files/downloads/",
-                 file_resources=None, curated=True, col_rename=None,
-                 **kwargs):
+    def __init__(
+            self,
+            path="https://www.disgenet.org/static/disgenet_ap1/files/downloads/",
+            file_resources=None,
+            curated=True,
+            col_rename=None,
+            **kwargs):
         """
         Args:
             path:
@@ -77,11 +94,16 @@ class DisGeNet(DiseaseAssociation):
             col_rename = COLUMNS_RENAME_DICT
         if file_resources is None:
             file_resources = {}
-            file_resources["curated_gene_disease_associations.tsv"] = "curated_gene_disease_associations.tsv.gz"
-            file_resources["all_gene_disease_associations.tsv"] = "all_gene_disease_associations.tsv.gz"
+            file_resources[
+                "curated_gene_disease_associations.tsv"] = "curated_gene_disease_associations.tsv.gz"
+            file_resources[
+                "all_gene_disease_associations.tsv"] = "all_gene_disease_associations.tsv.gz"
 
         self.curated = curated
-        super(DisGeNet, self).__init__(path, file_resources, col_rename=col_rename, **kwargs)
+        super(DisGeNet, self).__init__(path,
+                                       file_resources,
+                                       col_rename=col_rename,
+                                       **kwargs)
[…]

[…]
