### 2021-02-18 521711d
Iterate dictionary directly
