+++ openomics/transcriptomics.py
@@ -43,8 +43,7 @@ class ExpressionData(object):
             data (str, byte-like, pandas.DataFrame): Path or file stream of the table file to import. If a pandas DataFrame is passed, then import this dataframe and skip preprocessing steps.
             transpose (bool): True if given data table has samples or columns and variables for rows. False if the table has samples for row index, and gene names as columns.
             gene_index (str): The column name of gene/transcript/protein to index by.
-            usecols: A regex string to import column names from the table. Columns names
-            imported are string match, separated by "|".
+            usecols: A regex string to import column names from the table. Columns names imported are string match, separated by "|".
             gene_level (str): {"gene", "transcript", "peptide"} Chooses the
                 level of the gene/transcript/peptide of the genes list in this
                 expression data. The expression DataFrame's index will be
