+++ openomics/multiomics.py
@@ -11,7 +11,7 @@ from .clinical import (
 from .genomics import SomaticMutation, CopyNumberVariation, DNAMethylation
 from .imageomics import WholeSlideImage
 from .proteomics import Protein
-from .transcriptomics import MessengerRNA, MicroRNA, LncRNA, ExpressionData
+from .transcriptomics import MessengerRNA, MicroRNA, LncRNA, Expression
 
 
 class MultiOmics:
@@ -28,13 +28,13 @@ class MultiOmics:
         self.data = {}
 
     def add_omic(self,
-                 omic_data: ExpressionData,
+                 omic_data: Expression,
                  initialize_annotations: bool = True):
         """Adds an omic object to the Multiomics such that the samples in omic
         matches the samples existing in the other omics.
 
         Args:
-            omic_data (ExpressionData): The omic to add, e.g., MessengerRNA,
+            omic_data (Expression): The omic to add, e.g., MessengerRNA,
                 MicroRNA, LncRNA, etc.
             initialize_annotations (bool): default True. If true, initializes
                 the annotation dataframe in the omic object
