### 2021-02-18 4fefcbc
Format code with black and yapf
This commit fixes the style issues introduced in 757898d according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/ebc98560-40b2-4a47-ade9-5982cb0f20b0/
