+++ openomics/genomics.py
@@ -3,8 +3,19 @@ from openomics.database import Annotatable
 
 
 class SingleNucleotideVariants(Expression, Annotatable):
-    def __init__(self, data, transpose, gene_index, usecols=None, gene_level=None, sample_level="sample_index",
-                 transform_fn=None, dropna=False, npartitions=None, cohort_name=None):
+    def __init__(
+        self,
+        data,
+        transpose,
+        gene_index,
+        usecols=None,
+        gene_level=None,
+        sample_level="sample_index",
+        transform_fn=None,
+        dropna=False,
+        npartitions=None,
+        cohort_name=None,
+    ):
         """
         Args:
             data:
@@ -18,11 +29,18 @@ class SingleNucleotideVariants(Expression, Annotatable):
             npartitions:
             cohort_name:
         """
-        super(SingleNucleotideVariants, self).__init__(data=data, transpose=transpose, gene_index=gene_index,
-                                                       usecols=usecols,
-                                                       gene_level=gene_level, sample_level=sample_level,
-                                                       transform_fn=transform_fn, dropna=dropna,
-                                                       npartitions=npartitions, cohort_name=cohort_name)
+        super(SingleNucleotideVariants, self).__init__(
+            data=data,
+            transpose=transpose,
+            gene_index=gene_index,
+            usecols=usecols,
+            gene_level=gene_level,
+            sample_level=sample_level,
+            transform_fn=transform_fn,
+            dropna=dropna,
+            npartitions=npartitions,
+            cohort_name=cohort_name,
+        )
 
     @classmethod
     def name(cls):
@@ -30,8 +48,19 @@ class SingleNucleotideVariants(Expression, Annotatable):
 
 
 class SomaticMutation(Expression, Annotatable):
-    def __init__(self, data, transpose, gene_index, usecols=None, gene_level=None, sample_level="sample_index",
-                 transform_fn=None, dropna=False, npartitions=None, cohort_name=None):
+    def __init__(
+        self,
+        data,
+        transpose,
+        gene_index,
+        usecols=None,
+        gene_level=None,
+        sample_level="sample_index",
+        transform_fn=None,
+        dropna=False,
+        npartitions=None,
+        cohort_name=None,
+    ):
         """
         Args:
             data:
@@ -45,10 +74,18 @@ class SomaticMutation(Expression, Annotatable):
             npartitions:
             cohort_name:
         """
-        super(SomaticMutation, self).__init__(data=data, transpose=transpose, gene_index=gene_index, usecols=usecols,
-                                              gene_level=gene_level, sample_level=sample_level,
-                                              transform_fn=transform_fn, dropna=dropna,
-                                              npartitions=npartitions, cohort_name=cohort_name)
+        super(SomaticMutation, self).__init__(
+            data=data,
+            transpose=transpose,
+            gene_index=gene_index,
+            usecols=usecols,
+            gene_level=gene_level,
+            sample_level=sample_level,
+            transform_fn=transform_fn,
+            dropna=dropna,
+            npartitions=npartitions,
+            cohort_name=cohort_name,
+        )
 
     @classmethod
     def name(cls):
@@ -56,8 +93,19 @@ class SomaticMutation(Expression, Annotatable):
 
 
 class DNAMethylation(Expression, Annotatable):
-    def __init__(self, data, transpose, gene_index, usecols=None, gene_level=None, sample_level="sample_index",
-                 transform_fn=None, dropna=False, npartitions=None, cohort_name=None):
+    def __init__(
+        self,
+        data,
+        transpose,
+        gene_index,
+        usecols=None,
+        gene_level=None,
+        sample_level="sample_index",
+        transform_fn=None,
+        dropna=False,
+        npartitions=None,
+        cohort_name=None,
[…]
