### 2021-02-18 f3cbde2
improve speed of automated test by using only a 50 genes in each -omics
