### 2021-02-18 49b5eba
improve speed of automated test by using only a 50 genes in each -omics
