### 2021-02-18 355a4da
improve speed of automated test by using only a 50 genes in each -omics
