### 2021-02-18 b55f442
Format code with black and yapf
This commit fixes the style issues introduced in 355a4da according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/ddc165e8-d9d0-454d-a50d-ec0bf1af0939/
