### 2021-02-18 fb68b2e
improve speed of automated test by using only a 50 genes in each -omics
