### 2021-02-18 ccd0807
Format code with black and yapf
This commit fixes the style issues introduced in fb68b2e according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/2d0f3aed-b620-41b4-997f-2fd6edb01f9c/
