+++ openomics/clinical.py
@@ -10,13 +10,19 @@ NORMAL = "Normal"
 BCR_PATIENT_BARCODE = "bcr_patient_barcode"
 HISTOLOGIC_SUBTYPE = "histologic_subtype"
 PATHOLOGIC_STAGE = "pathologic_stage"
-TUMOR_NORMAL = 'tumor_normal'
-PREDICTED_SUBTYPE = 'predicted_subtype'
+TUMOR_NORMAL = "tumor_normal"
+PREDICTED_SUBTYPE = "predicted_subtype"
+
 
 class ClinicalData:
-    pathologic_stage_map = {'Stage IA': 'Stage I', 'Stage IB': 'Stage I',
-                            'Stage IIA': 'Stage II', 'Stage IIB': 'Stage II',
-                            'Stage IIIA': 'Stage III', 'Stage IIIB': 'Stage III'}
+    pathologic_stage_map = {
+        "Stage IA": "Stage I",
+        "Stage IB": "Stage I",
+        "Stage IIA": "Stage II",
+        "Stage IIB": "Stage II",
+        "Stage IIIA": "Stage III",
+        "Stage IIIB": "Stage III",
+    }
 
     def __init__(self, file_path, patient_id_col, columns=None):
         """This class manages the clinical data tables to handle the phenotype,
@@ -35,20 +41,32 @@ class ClinicalData:
             columns.append(patient_id_col)
 
         if isinstance(file_path, io.StringIO):
-            file_path.seek(0)  # Needed since the file was previous read to extract columns information
-            self.patient = pd.read_table(file_path,
-                                         skiprows=[1, 2],
-                                         na_values=["[Not Available]", "[Unknown]", "[Not Applicable]",
-                                                    "[Discrepancy]"],
-                                         usecols=columns
-                                         )
+            file_path.seek(
+                0
+            )  # Needed since the file was previous read to extract columns information
+            self.patient = pd.read_table(
+                file_path,
+                skiprows=[1, 2],
+                na_values=[
+                    "[Not Available]",
+                    "[Unknown]",
+                    "[Not Applicable]",
+                    "[Discrepancy]",
+                ],
+                usecols=columns,
+            )
         elif type(file_path) == str and os.path.exists(file_path):
-            self.patient = pd.read_table(file_path,
-                                         skiprows=[1, 2],
-                                         na_values=["[Not Available]", "[Unknown]", "[Not Applicable]",
-                                                    "[Discrepancy]"],
-                                         usecols=columns
-                                         )
+            self.patient = pd.read_table(
+                file_path,
+                skiprows=[1, 2],
+                na_values=[
+                    "[Not Available]",
+                    "[Unknown]",
+                    "[Not Applicable]",
+                    "[Discrepancy]",
+                ],
+                usecols=columns,
+            )
         elif isinstance(file_path, (pd.DataFrame, dd.DataFrame)):
             self.patient = file_path
 
@@ -59,11 +77,19 @@ class ClinicalData:
         self.patient.set_index(patient_id_col, inplace=True)
 
         # Rename columns
-        self.patient.rename({"ajcc_pathologic_tumor_stage": PATHOLOGIC_STAGE,
-                             "histological_type": HISTOLOGIC_SUBTYPE,
-                             "histologic_diagnosis.1": HISTOLOGIC_SUBTYPE}, axis=1, inplace=True)
+        self.patient.rename(
+            {
+                "ajcc_pathologic_tumor_stage": PATHOLOGIC_STAGE,
+                "histological_type": HISTOLOGIC_SUBTYPE,
+                "histologic_diagnosis.1": HISTOLOGIC_SUBTYPE,
+            },
+            axis=1,
+            inplace=True,
+        )
 
[…]
