+++ tests/test_multiomics.py
@@ -109,9 +109,9 @@ def test_import_expression_table_size(generate_TCGA_LUAD_MessengerRNA):
     """
     cohort_name = "LUAD"
     luad_data = MultiOmics(cohort_name)
-    luad_data.add_clinical_data(path=os.path.join(
-        cohort_folder_path,
-        "nationwidechildrens.org_clinical_patient_luad.txt"))
+    luad_data.add_clinical_data(
+        path=os.path.join(cohort_folder_path,
+                          "nationwidechildrens.org_clinical_patient_luad.txt"))
     luad_data.add_omic(generate_TCGA_LUAD_MessengerRNA)
     luad_data.build_samples()
     print(luad_data.data.keys())
@@ -135,9 +135,9 @@ def generate_TCGA_LUAD(
     """
     cohort_name = "LUAD"
     luad_data = MultiOmics(cohort_name)
-    luad_data.add_clinical_data(path=os.path.join(
-        cohort_folder_path,
-        "nationwidechildrens.org_clinical_patient_luad.txt"))
+    luad_data.add_clinical_data(
+        path=os.path.join(cohort_folder_path,
+                          "nationwidechildrens.org_clinical_patient_luad.txt"))
     luad_data.add_omic(generate_TCGA_LUAD_MessengerRNA)
     luad_data.add_omic(generate_TCGA_LUAD_MicroRNA)
     luad_data.add_omic(generate_TCGA_LUAD_LncRNA)
