+++ setup.py
@@ -55,7 +55,7 @@ with open('HISTORY.rst') as history_file:
 
 # Package meta-data.
 NAME = 'openomics'
-version='0.8.4'
+version='0.8.5'
 DESCRIPTION = 'OpenOmics provides a bioinformatics API and web-app platform integrate and visualize the multiomics and clinical data.'
 URL = 'https://github.com/BioMeCIS-Lab/OpenOmics'
 EMAIL = 'nhat.tran@mavs.uta.edu'
