### 2021-02-19 #31 @NickleDave
Hey @JonnyTran there's no strict deadline.

Our guide says "aim for one week" but with everyone already stressed out by the pandemic, and you facing even more problems because of the situation in Texas right now, I would not ask you to meet that.

Does three weeks sound do-able to you?
If not we can definitely figure something else out.
I just want to have something in my calendar to make sure I can keep track of reviews.
Please let me know.

Your comment above seems like it covers most of the feedback from the reviews. But when you get a chance and things are a little more back to normal for you, please do also make sure you address any specific comments from @ksielemann too.
