### 2021-02-19 053ac4c
Format code with black and yapf
This commit fixes the style issues introduced in 2dc98fb according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/721db550-c205-49e9-8c17-ed5e65c8d020/
