+++ tests/test_disease.py
@@ -4,12 +4,16 @@ from .test_multiomics import *
 
 @pytest.fixture
 def generate_DisGeNet_ftp():
-    return DisGeNet(path="https://www.disgenet.org/static/disgenet_ap1/files/downloads/", curated=True)
+    return DisGeNet(
+        path="https://www.disgenet.org/static/disgenet_ap1/files/downloads/",
+        curated=True,
+    )
 
 
 @pytest.fixture
 def generate_LncRNADisease_ftp():
-    return LncRNADisease(path="http://www.cuilab.cn/files/images/ldd/", species="Human")
+    return LncRNADisease(path="http://www.cuilab.cn/files/images/ldd/",
+                         species="Human")
 
 
 @pytest.fixture
@@ -27,8 +31,10 @@ def test_import_DisGeNet(generate_DisGeNet_ftp):
     Args:
         generate_DisGeNet_ftp:
     """
-    assert generate_DisGeNet_ftp.data_path == "https://www.disgenet.org/static/disgenet_ap1/files/downloads/"
-    assert not generate_DisGeNet_ftp.get_disease_assocs(index="gene_name").empty
+    assert (generate_DisGeNet_ftp.data_path ==
+            "https://www.disgenet.org/static/disgenet_ap1/files/downloads/")
+    assert not generate_DisGeNet_ftp.get_disease_assocs(
+        index="gene_name").empty
 
 
 def test_annotate_DisGeNet(generate_TCGA_LUAD, generate_DisGeNet_ftp):
@@ -37,8 +43,12 @@ def test_annotate_DisGeNet(generate_TCGA_LUAD, generate_DisGeNet_ftp):
         generate_TCGA_LUAD:
         generate_DisGeNet_ftp:
     """
-    generate_TCGA_LUAD.MessengerRNA.annotate_diseases(generate_DisGeNet_ftp, index="gene_name", )
-    assert {'disease_associations'}.issubset(generate_TCGA_LUAD.MessengerRNA.get_annotations().columns)
+    generate_TCGA_LUAD.MessengerRNA.annotate_diseases(
+        generate_DisGeNet_ftp,
+        index="gene_name",
+    )
+    assert {"disease_associations"}.issubset(
+        generate_TCGA_LUAD.MessengerRNA.get_annotations().columns)
 
 
 def test_import_HMDD(generate_HMDD_ftp):
@@ -56,8 +66,12 @@ def test_annotate_HMDD(generate_TCGA_LUAD, generate_HMDD_ftp):
         generate_TCGA_LUAD:
         generate_HMDD_ftp:
     """
-    generate_TCGA_LUAD.MicroRNA.annotate_diseases(generate_HMDD_ftp, index="gene_name", )
-    assert {'disease_associations'}.issubset(generate_TCGA_LUAD.MicroRNA.get_annotations().columns)
+    generate_TCGA_LUAD.MicroRNA.annotate_diseases(
+        generate_HMDD_ftp,
+        index="gene_name",
+    )
+    assert {"disease_associations"
+            }.issubset(generate_TCGA_LUAD.MicroRNA.get_annotations().columns)
 
 
 def test_LncRNADisease(generate_LncRNADisease_ftp):
@@ -65,18 +79,25 @@ def test_LncRNADisease(generate_LncRNADisease_ftp):
     Args:
         generate_LncRNADisease_ftp:
     """
-    assert generate_LncRNADisease_ftp.data_path == "http://www.cuilab.cn/files/images/ldd/"
-    assert not generate_LncRNADisease_ftp.get_disease_assocs(index="gene_name").empty
+    assert (generate_LncRNADisease_ftp.data_path ==
+            "http://www.cuilab.cn/files/images/ldd/")
[…]
