+++ openomics_web/utils/io.py
@@ -17,24 +17,33 @@ def get_table_columns(list_of_contents, list_of_names):
     content = list_of_contents[0]
     filename = list_of_names[0]
     _, file_extension = os.path.splitext(filename)
-    content_type, content_string = content.split(',')
+    content_type, content_string = content.split(",")
     decoded = base64.b64decode(content_string)
 
     if file_extension == ".tsv":
-        first_line = decoded.decode('utf-8').partition('\n')[0]
-        columns = first_line.split('\t')
+        first_line = decoded.decode("utf-8").partition("\n")[0]
+        columns = first_line.split("\t")
     elif file_extension == ".csv":
-        first_line = decoded.decode('utf-8').partition('\n')[0]
-        columns = first_line.split(',')
+        first_line = decoded.decode("utf-8").partition("\n")[0]
+        columns = first_line.split(",")
     elif file_extension == ".xls" or file_extension == ".xlsx":
-        columns = pd.read_excel(io.BytesIO(decoded), low_memory=True).columns.tolist()
+        columns = pd.read_excel(io.BytesIO(decoded),
+                                low_memory=True).columns.tolist()
     else:
-        columns = pd.read_table(io.StringIO(decoded.decode('utf-8')), low_memory=True).columns.tolist()
+        columns = pd.read_table(io.StringIO(decoded.decode("utf-8")),
+                                low_memory=True).columns.tolist()
     return columns
 
 
-def get_expression_data(list_of_contents, list_of_names, data_type, cohort_name=None, genes_col_name=None,
-                        columns_selected=None, transposed=None):
+def get_expression_data(
+    list_of_contents,
+    list_of_names,
+    data_type,
+    cohort_name=None,
+    genes_col_name=None,
+    columns_selected=None,
+    transposed=None,
+):
     """
     Args:
         list_of_contents:
@@ -60,7 +69,9 @@ def get_expression_data(list_of_contents, list_of_names, data_type, cohort_name=
     if data_type == MicroRNA.name():
         expression_data = MicroRNA(file, transpose=transposed, usecols=columns)
     elif data_type == MessengerRNA.name():
-        expression_data = MessengerRNA(file, transpose=transposed, usecols=columns)
+        expression_data = MessengerRNA(file,
+                                       transpose=transposed,
+                                       usecols=columns)
     elif data_type == LncRNA.name():
         expression_data = LncRNA(file, transpose=transposed, usecols=columns)
     elif data_type == Protein.name():
@@ -71,7 +82,14 @@ def get_expression_data(list_of_contents, list_of_names, data_type, cohort_name=
     return expression_data
 
 
-def get_clinical_data(file_content, file_name, data_type, cohort_name, patient_id_col=None, columns_selected=None):
+def get_clinical_data(
+    file_content,
+    file_name,
+    data_type,
+    cohort_name,
+    patient_id_col=None,
+    columns_selected=None,
+):
     """
     Args:
         file_content:
@@ -81,8 +99,17 @@ def get_clinical_data(file_content, file_name, data_type, cohort_name, patient_i
         patient_id_col:
         columns_selected:
     """
-    file = handle_filestreams([file_content, ], [file_name, ])
-    clinical_data = ClinicalData(file, patient_index=patient_id_col, columns=columns_selected)
+    file = handle_filestreams(
+        [
+            file_content,
+        ],
+        [
+            file_name,
+        ],
+    )
+    clinical_data = ClinicalData(file,
+                                 patient_index=patient_id_col,
+                                 columns=columns_selected)
 
     return clinical_data
 
@@ -99,16 +126,16 @@ def handle_filestreams(list_of_contents, list_of_names):
     else:
[…]

[…]
