+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.8.6
+current_version = 0.8.5
 commit = True
 tag = True
 
