### 2021-02-19 #31 @JonnyTran
Hi @NickleDave. Yes, 3 weeks is plenty of time for me to address the issues - how does Friday, March 12 sounds?
