+++ openomics/database/base.py
@@ -1,11 +1,16 @@
-import os, copy, logging
+import copy
 import difflib
-import filetype, zipfile, gzip, rarfile
+import itertools
+import logging
+import os
+from abc import ABC, abstractmethod
 
-from abc import abstractmethod
-
-import validators
 import dask.dataframe as dd
+import filetype
+import gzip
+import rarfile
+import validators
+import zipfile
 
 from openomics import backend as pd
 from openomics.utils.df import concat_uniques
@@ -14,7 +19,6 @@ from openomics.utils.io import get_pkg_data_filename
 
 class Dataset(object):
     COLUMNS_RENAME_DICT = None  # Needs initialization since subclasses may use this field to rename columns in dataframes.
-    SEQUENCE_COL_NAME = "sequence"
 
     def __init__(
         self,
@@ -157,7 +161,7 @@ class Dataset(object):
     def list_databases():
         return DEFAULT_LIBRARIES
 
-    def get_annotations(self, index, columns, agg="sum"):
+    def get_annotations(self, index, columns, agg="concat", filter_values=None):
         """Returns the Database's DataFrame such that it's indexed by :param
         index:, which then applies a groupby operation and aggregates all other
         columns by concatenating all unique values.
@@ -165,7 +169,9 @@ class Dataset(object):
         Args:
             index (str): The index column name of the DataFrame to join by
             columns ([str]): a list of column names
-            agg (str): Function to aggregate when there is more than one values for each index instance. E.g. ['first', 'last', 'sum', 'mean', 'concat'], default 'concat'.
+            agg (str): Function to aggregate when there is more than one values for each index instance.
+                E.g. ['first', 'last', 'sum', 'mean', 'size', 'concat'], default 'concat'.
+            filter_values (pd.Series): A series of values on the `index` column to filter before performing the groupby-agg operations.
 
         Returns:
             df (DataFrame): A dataframe to be used for annotation
@@ -182,8 +188,11 @@ class Dataset(object):
 
         df = self.data[columns + [index]]
 
-        if index != self.data.index.name and index in self.data.columns:
-            df = df.set_index(index)
+        if filter_values is not None:
+            df = df[df[index].isin(filter_values)]
+
+        # if index != self.data.index.name and index in self.data.columns:
+        #     df = df.set_index(index)
 
         # Groupby index
         groupby = df.groupby(index)
@@ -192,15 +201,21 @@ class Dataset(object):
         if agg == "concat":
             if isinstance(df, pd.DataFrame):
                 aggregated = groupby.agg({col: concat_uniques for col in columns})
-            else:
-                agg_func = dd.Aggregation('custom_agg',
-                                          chunk=lambda x: x,
-                                          agg=concat_uniques)
-                aggregated = groupby.agg({col: agg_func for col in columns})
+
+            elif isinstance(df, dd.DataFrame):
+                collect_concat = dd.Aggregation(
+                    name='collect_concat',
+                    chunk=lambda s1: s1.apply(list),
+                    agg=lambda s2: s2.apply(lambda chunks: filter(
+                        lambda x: False if x == "None" or x is None else True,
+                        set(itertools.chain.from_iterable(chunks)))),
+                    finalize=lambda s3: s3.apply(lambda xx: '|'.join(xx))
+                )
+                aggregated = groupby.agg({col: collect_concat for col in columns})
 
         # Any other aggregation functions
         else:
-            aggregated = groupby.aggregate(agg)
+            aggregated = groupby.agg({col: agg for col in columns})
 
         # if aggregated.index.duplicated().sum() > 0:
         #     raise ValueError("DataFrame must not have duplicates in index")
@@ -229,14 +244,12 @@ class Dataset(object):
         raise NotImplementedError
 
 
[…]
