+++ tests/test_sequences.py
@@ -49,9 +49,9 @@ def test_annotate_GENCODE(generate_TCGA_LUAD, generate_GENCODE):
         generate_TCGA_LUAD:
         generate_GENCODE:
     """
-    generate_TCGA_LUAD.LncRNA.annotate_genomics(generate_GENCODE, index="gene_id", columns=['gene_name'])
+    generate_TCGA_LUAD.LncRNA.annotate_attributes(generate_GENCODE, on="gene_id", columns=['gene_name'])
     assert {'gene_name'}.issubset(
-        generate_TCGA_LUAD.LncRNA.get_annotations().columns)
+        generate_TCGA_LUAD.LncRNA.annotations.columns)
 
 
 def test_annotate_dask_GENCODE(generate_TCGA_LUAD, generate_GENCODE_dask):
@@ -62,11 +62,17 @@ def test_annotate_dask_GENCODE(generate_TCGA_LUAD, generate_GENCODE_dask):
     """
     generate_GENCODE_dask.data = generate_GENCODE_dask.data[generate_GENCODE_dask.data["gene_id"].notnull()]
 
-    generate_TCGA_LUAD.LncRNA.annotate_genomics(generate_GENCODE_dask, index="gene_id", columns=['gene_name'])
-    assert {'gene_name'}.issubset(
-        generate_TCGA_LUAD.LncRNA.get_annotations().columns)
+    generate_TCGA_LUAD.LncRNA.annotate_attributes(generate_GENCODE_dask,
+                                                  on="gene_id",
+                                                  columns=['gene_name', 'transcript_id'],
+                                                  agg="concat")
+    assert {'gene_name', 'transcript_id'}.issubset(
+        generate_TCGA_LUAD.LncRNA.annotations.columns)
 
 
 def test_annotate_sequence_GENCODE(generate_TCGA_LUAD, generate_GENCODE):
-    generate_TCGA_LUAD.LncRNA.annotate_sequences(generate_GENCODE, index="gene_id", omic="LncRNA",
-                                                 agg_sequences="longest")
+    generate_TCGA_LUAD.LncRNA.annotate_sequences(generate_GENCODE,
+                                                 index="gene_id",
+                                                 omic="LncRNA",
+                                                 agg="longest")
+    assert not generate_TCGA_LUAD.LncRNA.annotations["sequence"].empty
