### 2021-02-19 f36a049
Format code with black and yapf
This commit fixes the style issues introduced in 527f034 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/4519b561-479f-427c-99fd-64d1f4ca840d/
