+++ tests/test_annotations.py
@@ -4,12 +4,14 @@ from .test_multiomics import *
 
 @pytest.fixture
 def generate_RNACentral_ftp():
-    return RNAcentral(path="ftp://ftp.ebi.ac.uk/pub/databases/RNAcentral/current_release/")
+    return RNAcentral(
+        path="ftp://ftp.ebi.ac.uk/pub/databases/RNAcentral/current_release/")
 
 
 @pytest.fixture
 def generate_GTEx_expressions():
-    return GTEx(path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/", )
+    return GTEx(
+        path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/", )
 
 
 @pytest.fixture
@@ -22,7 +24,8 @@ def test_import_rnacentral_db(generate_RNACentral_ftp):
     Args:
         generate_RNACentral_ftp:
     """
-    assert generate_RNACentral_ftp.data_path == 'ftp://ftp.ebi.ac.uk/pub/databases/RNAcentral/current_release/'
+    assert (generate_RNACentral_ftp.data_path ==
+            "ftp://ftp.ebi.ac.uk/pub/databases/RNAcentral/current_release/")
 
 
 def test_import_GTEx(generate_GTEx_expressions):
@@ -30,8 +33,10 @@ def test_import_GTEx(generate_GTEx_expressions):
     Args:
         generate_GTEx_expressions:
     """
-    assert generate_GTEx_expressions.data_path == "https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/"
-    assert not generate_GTEx_expressions.get_expressions(index="gene_name").empty
+    assert (generate_GTEx_expressions.data_path ==
+            "https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/")
+    assert not generate_GTEx_expressions.get_expressions(
+        index="gene_name").empty
     assert not generate_GTEx_expressions.get_expressions(index="gene_id").empty
 
 
@@ -40,7 +45,8 @@ def test_import_GeneOntology(generate_GeneOntology):
     Args:
         generate_GeneOntology:
     """
-    assert generate_GeneOntology.data_path == "http://geneontology.org/gene-associations/"
+    assert (generate_GeneOntology.data_path ==
+            "http://geneontology.org/gene-associations/")
 
 
 def test_annotate_rnacentral(generate_TCGA_LUAD, generate_RNACentral_ftp):
@@ -49,18 +55,24 @@ def test_annotate_rnacentral(generate_TCGA_LUAD, generate_RNACentral_ftp):
         generate_TCGA_LUAD:
         generate_RNACentral_ftp:
     """
-    generate_TCGA_LUAD.MicroRNA.annotate_attributes(database=generate_RNACentral_ftp, on="gene_name",
-                                                    columns=['gene_name', 'RNA type'])
-    assert {'RNA type'}.issubset(generate_TCGA_LUAD.MicroRNA.annotations.columns)
+    generate_TCGA_LUAD.MicroRNA.annotate_attributes(
+        database=generate_RNACentral_ftp,
+        on="gene_name",
+        columns=["gene_name", "RNA type"],
+    )
+    assert {"RNA type"
+            }.issubset(generate_TCGA_LUAD.MicroRNA.annotations.columns)
 
 
-def test_annotate_expressions_GTEx(generate_TCGA_LUAD, generate_GTEx_expressions):
+def test_annotate_expressions_GTEx(generate_TCGA_LUAD,
+                                   generate_GTEx_expressions):
     """
     Args:
         generate_TCGA_LUAD:
         generate_GTEx_expressions:
     """
[…]
