+++ setup.py
@@ -9,7 +9,7 @@ from setuptools import setup, find_packages, Command
 class UploadCommand(Command):
     """Support setup.py upload."""
 
-    description = 'Build and publish the package.'
+    description = "Build and publish the package."
     user_options = []
 
     @staticmethod
@@ -19,7 +19,7 @@ class UploadCommand(Command):
         Args:
             s:
         """
-        print('\033[1m{0}\033[0m'.format(s))
+        print("\033[1m{0}\033[0m".format(s))
 
     def initialize_options(self):
         pass
@@ -29,54 +29,64 @@ class UploadCommand(Command):
 
     def run(self):
         try:
-            self.status('Removing previous builds…')
-            rmtree(os.path.join(here, 'dist'))
+            self.status("Removing previous builds…")
+            rmtree(os.path.join(here, "dist"))
         except OSError:
             pass
 
-        self.status('Building Source and Wheel (universal) distribution…')
-        os.system('{0} setup.py sdist bdist_wheel --universal'.format(sys.executable))
+        self.status("Building Source and Wheel (universal) distribution…")
+        os.system("{0} setup.py sdist bdist_wheel --universal".format(
+            sys.executable))
 
-        self.status('Uploading the package to PyPi via Twine…')
-        os.system('twine upload dist/*')
+        self.status("Uploading the package to PyPi via Twine…")
+        os.system("twine upload dist/*")
 
-        self.status('Pushing git tags…')
-        os.system('git tag v{0}'.format(version))
-        os.system('git push --tags')
+        self.status("Pushing git tags…")
+        os.system("git tag v{0}".format(version))
+        os.system("git push --tags")
 
         sys.exit()
 
 
-with open('README.rst') as readme_file:
+with open("README.rst") as readme_file:
     readme = readme_file.read()
 
-with open('HISTORY.rst') as history_file:
+with open("HISTORY.rst") as history_file:
     history = history_file.read()
 
 # Package meta-data.
-NAME = 'openomics'
-version='0.8.6'
-DESCRIPTION = 'OpenOmics provides a bioinformatics API and web-app platform integrate and visualize the multiomics and clinical data.'
-URL = 'https://github.com/BioMeCIS-Lab/OpenOmics'
-EMAIL = 'nhat.tran@mavs.uta.edu'
+NAME = "openomics"
+version = "0.8.6"
+DESCRIPTION = "OpenOmics provides a bioinformatics API and web-app platform integrate and visualize the multiomics and clinical data."
+URL = "https://github.com/BioMeCIS-Lab/OpenOmics"
+EMAIL = "nhat.tran@mavs.uta.edu"
 
-AUTHOR = 'Jonny Tran'
+AUTHOR = "Jonny Tran"
 
 requirements = pd.read_table("./requirements.txt", header=None)[0].tolist()
-setup_requirements = ['pytest-runner', 'twine']
+setup_requirements = ["pytest-runner", "twine"]
 
-test_requirements = ['pytest', ]
+test_requirements = [
+    "pytest",
+]
 
 here = os.path.abspath(os.path.dirname(__file__))
 
 setup(
     name=NAME,
     version=version,
-    packages=find_packages(include=['openomics', 'openomics.database', 'openomics.utils', 'openomics.visualization']),
-    package_dir={NAME: 'openomics',
-                 'openomics_web': 'openomics_web'},
+    packages=find_packages(include=[
+        "openomics",
+        "openomics.database",
+        "openomics.utils",
+        "openomics.visualization",
+    ]),
+    package_dir={
+        NAME: "openomics",
+        "openomics_web": "openomics_web"
+    },
     url=URL,
-    license='MIT license',
+    license="MIT license",
     install_requires=requirements,
[…]
