+++ tests/test_multiomics.py
@@ -1,9 +1,5 @@
 """Tests for `openomics` package."""
 
-import os
-
-import pytest
-
 from openomics import MessengerRNA, MicroRNA, LncRNA, Protein, SomaticMutation
 from openomics import MultiOmics
 
@@ -37,6 +33,8 @@ def generate_TCGA_LUAD_MessengerRNA_dask():
         npartitions=4,
     )
     data.drop_genes(data.expressions.columns[50:])
+    data.initialize_annotations(index="gene_name", gene_list=None)
+
     return data
 
 
