### 2021-02-19 68f7ee2
Format code with black and yapf
This commit fixes the style issues introduced in e2e2821 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/090064fb-7fbd-45ba-9072-372f92f7ca3a/
