### 2021-02-19 4fd3fda
samples only 1% of Database entries for test
