+++ tests/test_sequences.py
@@ -6,21 +6,29 @@ from .test_multiomics import *
 
 @pytest.fixture
 def generate_GENCODE():
-    return GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
-                   file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
-                                   "lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz"}, )
+    gencode = GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
+                      file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
+                                      "lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz"}, )
+
+    gencode.data = gencode.data.sample(frac=0.01)
+    return gencode
 
 
 @pytest.fixture
 def generate_GENCODE_dask():
-    return GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
-                   file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
-                                   "lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz"}, npartitions=8)
+    gencode = GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
+                      file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
+                                      "lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz"}, npartitions=8)
+
+    gencode.data = gencode.data.sample(frac=0.01)
+    return gencode
 
 
 @pytest.fixture
 def generate_MirBase_ftp():
-    return MirBase(path="ftp://mirbase.org/pub/mirbase/CURRENT/")
+    mirbase = MirBase(path="ftp://mirbase.org/pub/mirbase/CURRENT/")
+    mirbase.data = mirbase.data.sample(frac=0.01)
+    return mirbase
 
 
 def test_import_mirbase_db(generate_MirBase_ftp):
@@ -50,7 +58,14 @@ def test_annotate_GENCODE(generate_TCGA_LUAD, generate_GENCODE):
         generate_GENCODE:
     """
     generate_TCGA_LUAD.LncRNA.annotate_attributes(generate_GENCODE, on="gene_id", columns=['gene_name'])
-    assert {'gene_name'}.issubset(
+
+    # Test join on off-index
+    generate_TCGA_LUAD.LncRNA.annotate_attributes(generate_GENCODE,
+                                                  on="gene_name",
+                                                  columns=['transcript_id'],
+                                                  agg="concat")
+
+    assert {'gene_name', 'transcript_id'}.issubset(
         generate_TCGA_LUAD.LncRNA.annotations.columns)
 
 
@@ -83,4 +98,5 @@ def test_annotate_sequence_GENCODE(generate_TCGA_LUAD, generate_GENCODE):
                                                  index="gene_id",
                                                  omic="LncRNA",
                                                  agg="longest")
+
     assert not generate_TCGA_LUAD.LncRNA.annotations["sequence"].empty
