### 2021-02-19 64044ac
Format code with black and yapf
This commit fixes the style issues introduced in 4fd3fda according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/882ba51b-95e3-4c35-8ccc-39df71b54104/
