+++ openomics/database/base.py
@@ -161,7 +161,13 @@ class Dataset(object):
     def list_databases():
         return DEFAULT_LIBRARIES
 
-    def get_annotations(self, index: str, columns: list, agg: str = "concat", filter_values: pd.Series = None):
+    def get_annotations(
+        self,
+        index: str,
+        columns: list,
+        agg: str = "concat",
+        filter_values: pd.Series = None,
+    ):
         """Returns the Database's DataFrame such that it's indexed by :param
         index:, which then applies a groupby operation and aggregates all other
         columns by concatenating all unique values.
@@ -201,18 +207,23 @@ class Dataset(object):
         #  Aggregate by all columns by concatenating unique values
         if agg == "concat":
             if isinstance(df, pd.DataFrame):
-                aggregated = groupby.agg({col: concat_uniques for col in columns})
+                aggregated = groupby.agg(
+                    {col: concat_uniques
+                     for col in columns})
 
             elif isinstance(df, dd.DataFrame):
                 collect_concat = dd.Aggregation(
-                    name='collect_concat',
+                    name="collect_concat",
                     chunk=lambda s1: s1.apply(list),
                     agg=lambda s2: s2.apply(lambda chunks: filter(
                         lambda x: False if x == "None" or x is None else True,
-                        set(itertools.chain.from_iterable(chunks)))),
-                    finalize=lambda s3: s3.apply(lambda xx: '|'.join(xx))
+                        set(itertools.chain.from_iterable(chunks)),
+                    )),
+                    finalize=lambda s3: s3.apply(lambda xx: "|".join(xx)),
                 )
-                aggregated = groupby.agg({col: collect_concat for col in columns})
+                aggregated = groupby.agg(
+                    {col: collect_concat
+                     for col in columns})
 
         # Any other aggregation functions
         else:
@@ -250,6 +261,7 @@ class Annotatable(ABC):
     downloaded from various databases. The database will be imported as attributes
     information to the genes's annotations, or interactions between the genes.
     """
+
     SEQUENCE_COL_NAME = "sequence"
 
     def get_annotations(self):
@@ -279,7 +291,12 @@ class Annotatable(ABC):
         self.annotations = pd.DataFrame(index=gene_list)
         self.annotations.index.name = index
 
-    def annotate_attributes(self, database: Dataset, on, columns, agg="concat", fuzzy_match=False):
+    def annotate_attributes(self,
+                            database: Dataset,
+                            on,
+                            columns,
+                            agg="concat",
+                            fuzzy_match=False):
         """Performs a left outer join between the annotation and Database's
         DataFrame, on the index key. The index argument must be column present
         in both DataFrames. If there exists overlapping columns from the join, then
@@ -294,7 +311,9 @@ class Annotatable(ABC):
             fuzzy_match (bool): default False. Whether to join the annotation by applying a fuzzy match on the index with difflib.get_close_matches(). It is very computationally expensive and thus should only be used sparingly.
         """
         if not hasattr(self, "annotations"):
-            raise Exception(f"Must run .initialize_annotations() on {self.__class__.__name__} first.")
+            raise Exception(
+                f"Must run .initialize_annotations() on {self.__class__.__name__} first."
+            )
 
         if on in self.annotations.columns:
             filter_values = self.annotations[on]
@@ -303,18 +322,24 @@ class Annotatable(ABC):
         else:
             filter_values = None
 
-        database_df = database.get_annotations(on, columns=columns, agg=agg, filter_values=filter_values)
+        database_df = database.get_annotations(on,
+                                               columns=columns,
+                                               agg=agg,
+                                               filter_values=filter_values)
 
         if len(database_df.columns) == 0:
[…]
