+++ tests/test_interaction.py
@@ -25,9 +25,11 @@ def generate_STRING():
 @pytest.fixture
 def generate_MiRTarBase():
     return MiRTarBase(
-        path="/data/datasets/Bioinformatics_ExternalData/miRTarBase/",  # Hard-coded
+        path=
+        "/data/datasets/Bioinformatics_ExternalData/miRTarBase/",  # Hard-coded
         strip_mirna_name=True,
-        filters={"Species (Target Gene)": "Homo sapiens"})
+        filters={"Species (Target Gene)": "Homo sapiens"},
+    )
 
 
 # Test disabled since obtaining MiRTarBase via ftp is unreachable
@@ -37,8 +39,6 @@ def generate_MiRTarBase():
 #         generate_MiRTarBase:
 #     """
 #     assert generate_MiRTarBase.data_path is not None
-
-
 
 
 def test_import_STRING(generate_STRING):
@@ -55,7 +55,8 @@ def test_annotate_STRING(generate_TCGA_LUAD, generate_STRING):
         generate_TCGA_LUAD:
         generate_STRING:
     """
-    generate_TCGA_LUAD.Protein.annotate_sequences(generate_STRING, index="protein_name")
+    generate_TCGA_LUAD.Protein.annotate_sequences(generate_STRING,
+                                                  index="protein_name")
     assert not generate_TCGA_LUAD.Protein.annotations["sequence"].empty
 
 
