### 2021-02-19 b3fd16e
samples only 1% of Database entries for test
