### 2021-02-19 585ff62
Format code with black and yapf
This commit fixes the style issues introduced in b3fd16e according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/f007cdc5-97ac-4778-8886-edd7cea1a746/
