### 2021-02-19 BioMeCIS-Lab/OpenOmics#113 @JonnyTran
Issue has been resolved under Mac OS + Python 3.9 and Python 3.8 when running `pip install openomics`.
