### 2021-02-19 BioMeCIS-Lab/OpenOmics#115 @JonnyTran
Removed the `test_import_MiRTarBase` automated test since MiRTarBase is not obtainable via ftp.
