### 2021-02-19 33fe5b6
samples only 1% of Database entries for test
