+++ setup.py
@@ -40,9 +40,9 @@ class UploadCommand(Command):
         self.status('Uploading the package to PyPi via Twine…')
         os.system('twine upload dist/*')
 
-        self.status('Pushing git tags…')
-        os.system('git tag v{0}'.format(version))
-        os.system('git push --tags')
+        # self.status('Pushing git tags…')
+        # os.system('git tag v{0}'.format(version))
+        # os.system('git push --tags')
 
         sys.exit()
 
@@ -94,15 +94,16 @@ setup(
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
         'Programming Language :: Python :: 3',
-            'Programming Language :: Python :: 3.6',
-            'Programming Language :: Python :: 3.7',
-            'Programming Language :: Python :: 3.8',
-            'Programming Language :: Python :: 3.9',
-        ],
+        'Programming Language :: Python :: 3.6',
+        'Programming Language :: Python :: 3.7',
+        'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
+    ],
     author=AUTHOR,
     author_email=EMAIL,
     description=DESCRIPTION,
     long_description=readme + '\n\n' + history,
+    long_description_content_type="text/markdown",
     cmdclass={
         'upload': UploadCommand,
     },
