### 2021-02-19 92522e8
Format code with black and yapf
This commit fixes the style issues introduced in 33fe5b6 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/87b892bf-8e6b-4cfe-9d96-a1a14800c4ae/
