### 2021-02-19 a8bdd0c
samples only 1% of Database entries for test
