### 2021-02-19 638d842
samples only 1% of Database entries for test
