+++ openomics/__init__.py
@@ -1,12 +1,12 @@
 from __future__ import print_function, division, absolute_import
 
-import imp
+import astropy
+import logging
+import os
 import sys
 
 import dask.dataframe as dd
 import pandas as pd
-
-# -*- coding: utf-8 -*-
 
 """Top-level package for openomics."""
 
@@ -14,8 +14,19 @@ __author__ = """Nhat (Jonny) Tran"""
 __email__ = 'nhat.tran@mavs.uta.edu'
 __version__ = '0.8.5'
 
-__BACKEND__ = "pandas"
-backend = pd
+# Initialize configurations
+this = sys.modules[__name__]
+this.config = {}
+
+# Read configuration from ~/.openomics/conf.json
+
+# Set pandas backend
+if not this.config:
+    this.config["backend"] = pd
+
+    # Set cache download directory
+    this.config["cache_dir"] = astropy.config.get_cache_dir(this.__name__)
+    logging.info("Cache directory is", this.config["cache_dir"])
 
 from . import database, utils
 
@@ -43,15 +54,18 @@ def set_backend(new):
     Args:
         new:
     """
-    global __BACKEND__
-    global backend
     assert new in ["dask", "pandas"]
 
-    __BACKEND__ = new
     if new == "dask":
-        backend = dd
+        this.config["backend"] = dd
     else:
-        backend = pd
+        this.config["backend"] = pd
 
-    for module in sys.modules.values():
-        imp.reload(module)
+
+def set_cache_dir(path, delete_temp=False):
+    if not os.path.exists(path):
+        raise NotADirectoryError(path)
+
+    this.config["cache_dir"] = path
+    astropy.config.set_temp_cache(path=path, delete=delete_temp)
+    logging.info("Cache directory is", this.config["cache_dir"])
