### 2021-02-19 d7ff79e
Format code with black and yapf
This commit fixes the style issues introduced in 638d842 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/97c3987b-7f82-45a4-8911-c48576a5e812/
