### 2021-02-19 e45e215
added openomics configs
