### 2021-02-19 82794c4
added openomics configs
