### 2021-02-19 f918a5f
added openomics configs
