### 2021-02-20 ec93ef6
if not exists, create openomics configs
