### 2021-02-20 5cf92e4
Format code with black and yapf
This commit fixes the style issues introduced in ec93ef6 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/36e373b4-f311-41ef-8486-b965ca7fa8e7/
