+++ openomics/database/annotation.py
@@ -14,8 +14,14 @@ DEFAULT_LIBRARY_PATH = os.path.join(expanduser("~"), ".openomics", "databases")
 import pandas as pd
 import dask.dataframe as dd
 
+
 class TANRIC(Dataset):
-    def __init__(self, path, file_resources=None, col_rename=None, npartitions=0, verbose=False):
+    def __init__(self,
+                 path,
+                 file_resources=None,
+                 col_rename=None,
+                 npartitions=0,
+                 verbose=False):
         """
         Args:
             path:
@@ -24,7 +30,8 @@ class TANRIC(Dataset):
             npartitions:
             verbose:
         """
-        super(TANRIC, self).__init__(path, file_resources, col_rename, npartitions, verbose)
+        super(TANRIC, self).__init__(path, file_resources, col_rename,
+                                     npartitions, verbose)
 
     def load_dataframe(self, file_resources, npartitions=None):
         """
@@ -45,8 +52,10 @@ class TANRIC(Dataset):
             genes_index:
         """
         df = pd.read_table(self.file_resources["TCGA-LUAD-rnaexpr.tsv"])
-        df[genes_index] = df[genes_index].str.replace("[.].*", "")  # Removing .# ENGS gene version number at the end
-        df = df[~df[genes_index].duplicated(keep='first')]  # Remove duplicate genes
+        df[genes_index] = df[genes_index].str.replace(
+            "[.].*", "")  # Removing .# ENGS gene version number at the end
+        df = df[~df[genes_index].duplicated(
+            keep="first")]  # Remove duplicate genes
 
         # Drop NA gene rows
         df.dropna(axis=0, inplace=True)
@@ -58,9 +67,9 @@ class TANRIC(Dataset):
         # Change index string to bcr_sample_barcode standard
         def change_patient_barcode(s):
             if "Normal" in s:
-                return s[s.find('TCGA'):] + "-11A"
+                return s[s.find("TCGA"):] + "-11A"
             elif "Tumor" in s:
-                return s[s.find('TCGA'):] + "-01A"
+                return s[s.find("TCGA"):] + "-01A"
             else:
                 return s
 
@@ -76,8 +85,14 @@ class ProteinAtlas(Dataset):
         "Ensembl": "gene_id",
     }
 
-    def __init__(self, path="https://www.proteinatlas.org/download/", file_resources=None,
-                 col_rename=COLUMNS_RENAME_DICT, npartitions=0, verbose=False):
+    def __init__(
+        self,
+        path="https://www.proteinatlas.org/download/",
+        file_resources=None,
+        col_rename=COLUMNS_RENAME_DICT,
+        npartitions=0,
+        verbose=False,
+    ):
         """
         Args:
             path:
@@ -90,7 +105,8 @@ class ProteinAtlas(Dataset):
             file_resources = {}
             file_resources["proteinatlas.tsv"] = "proteinatlas.tsv.zip"
 
-        super(ProteinAtlas, self).__init__(path, file_resources, col_rename, npartitions, verbose)
+        super(ProteinAtlas, self).__init__(path, file_resources, col_rename,
+                                           npartitions, verbose)
 
     def load_dataframe(self, file_resources, npartitions=None):
         """
@@ -120,19 +136,27 @@ class ProteinAtlas(Dataset):
             expressions (pd.DataFrame):
         """
         columns = "|".join([type, index])
-        expressions = self.data.filter(regex=columns).groupby(
-            index).median()
+        expressions = self.data.filter(regex=columns).groupby(index).median()
         return expressions
 
 
 class RNAcentral(Dataset):
-    COLUMNS_RENAME_DICT = {'ensembl_gene_id': 'gene_id',
-                           'gene symbol': 'gene_name',
-                           'external id': 'transcript_id',
[…]
