### 2021-02-20 6f49f04
if not exists, create openomics configs
