+++ openomics/__init__.py
@@ -15,7 +15,6 @@ __author__ = """Nhat (Jonny) Tran"""
 __email__ = 'nhat.tran@mavs.uta.edu'
 __version__ = '0.8.5'
 
-config_file_path = "~/.openomics/conf.json"
 
 # Initialize configurations
 this = sys.modules[__name__]
@@ -28,25 +27,31 @@ this.config["backend"] = pd
 this.config["cache_dir"] = astropy.config.get_cache_dir(this.__name__)
 logging.info("Cache directory is", this.config["cache_dir"])
 
+user_conf_path = "~/.openomics/conf.txt"
 # Initialize user configuration file at ~/.openomics/conf.json
-if not os.path.exists(config_file_path):
-    if not os.path.exists("~/.openomics/"):
-        os.mkdir("~/.openomics/")
+if not os.path.exists(user_conf_path):
+    if not os.path.exists("~/.openomics"):
+        os.makedirs("~/.openomics")
 
-    if not os.path.exists(config_file_path):
+    if not os.path.exists(user_conf_path):
         base_config = {}
+        base_config['database'] = []
+
         base_config['database'].append({
             'cache_dir': astropy.config.get_cache_dir(this.__name__)
         })
 
-        with open(config_file_path, 'w') as config_file:
+        with open(user_conf_path, 'w') as config_file:
             json.dump(base_config, config_file)
 
 # Read configuration from ~/.openomics/conf.json
-if os.path.exists(config_file_path):
-    user_config = json.load(config_file_path)
+if os.path.exists(user_conf_path):
+    with open(user_conf_path, 'w') as config_file:
+        user_config = json.load(config_file)
+
     if user_config:
-        this.config.update(user_config)
+        for p in user_config['database']:
+            this.config.update(p)
 
 from . import database, utils
 
