### 2021-02-20 7153a41
Format code with black and yapf
This commit fixes the style issues introduced in 6f49f04 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/fa0d5ab9-28e0-485b-b599-3b619099c23b/
