### 2021-02-20 b1936fb
if not exists, create openomics configs
