### 2021-02-20 99f4d0f
Format code with black and yapf
This commit fixes the style issues introduced in b1936fb according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/21a0d615-ea48-4cb7-964f-63e5f328af2d/
