+++ openomics/__init__.py
@@ -8,13 +8,11 @@ import sys
 import astropy
 import dask.dataframe as dd
 import pandas as pd
-
 """Top-level package for openomics."""
 
 __author__ = """Nhat (Jonny) Tran"""
-__email__ = 'nhat.tran@mavs.uta.edu'
-__version__ = '0.8.5'
-
+__email__ = "nhat.tran@mavs.uta.edu"
+__version__ = "0.8.5"
 
 # Initialize configurations
 this = sys.modules[__name__]
@@ -35,43 +33,39 @@ if not os.path.exists(user_conf_path):
 
     if not os.path.exists(user_conf_path):
         base_config = {}
-        base_config['database'] = []
+        base_config["database"] = []
 
-        base_config['database'].append({
-            'cache_dir': astropy.config.get_cache_dir(this.__name__)
-        })
+        base_config["database"].append(
+            {"cache_dir": astropy.config.get_cache_dir(this.__name__)})
 
-        with open(user_conf_path, 'w', encoding='utf-8') as config_file:
+        with open(user_conf_path, "w", encoding="utf-8") as config_file:
             json.dump(base_config, config_file, indent=4)
 
 # Read configuration from ~/.openomics/conf.json
 if os.path.exists(user_conf_path):
-    with open(user_conf_path, 'w') as config_file:
+    with open(user_conf_path, "w") as config_file:
         user_config = json.load(config_file)
 
     if user_config:
-        for p in user_config['database']:
+        for p in user_config["database"]:
             this.config.update(p)
 
 from . import database, utils
 
 from .transcriptomics import (
-    Expression, MessengerRNA, MicroRNA, LncRNA,
+    Expression,
+    MessengerRNA,
+    MicroRNA,
+    LncRNA,
 )
 
-from .genomics import (
-    SomaticMutation, DNAMethylation, CopyNumberVariation
-)
+from .genomics import SomaticMutation, DNAMethylation, CopyNumberVariation
 
-from .proteomics import (
-    Protein
-)
+from .proteomics import Protein
 
 from .clinical import ClinicalData
 
-from .multiomics import (
-    MultiOmics
-)
+from .multiomics import MultiOmics
 
 
 def set_backend(new):
