### 2021-02-20 9373f5d
if not exists, create openomics configs
