### 2021-02-20 2edd566
Format code with black and yapf
This commit fixes the style issues introduced in 9373f5d according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/a4174e49-14fa-4f7e-ae80-d38145d5bf7f/
