### 2021-02-20 bd8245b
if not exists, create openomics configs
