+++ openomics/__init__.py
@@ -41,13 +41,13 @@ if not os.path.exists(user_conf_path):
             'cache_dir': astropy.config.get_cache_dir(this.__name__)
         })
 
-        with open(user_conf_path, 'w', encoding='utf-8') as config_file:
-            json.dump(base_config, fp=user_conf_path, indent=4)
+        with open(user_conf_path, 'w', encoding='utf-8') as file:
+            json.dump(base_config, fp=file, indent=4)
 
 # Read configuration from ~/.openomics/conf.json
 if os.path.exists(user_conf_path):
-    with open(user_conf_path, 'w') as config_file:
-        user_config = json.load(config_file)
+    with open(user_conf_path, 'w', encoding='utf-8') as file:
+        user_config = json.load(fp=file)
 
     if user_config:
         for p in user_config['database']:
