### 2021-02-20 7e34da9
if not exists, create openomics configs
