### 2021-02-20 22391e2
Format code with black and yapf
This commit fixes the style issues introduced in 7e34da9 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/d13e8d58-23ac-4a87-8568-ad0af4fac04f/
