### 2021-02-20 95d83ea
if not exists, create openomics configs
