### 2021-02-20 2412c23
Format code with black and yapf
This commit fixes the style issues introduced in 95d83ea according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/06bbfa5d-d844-49c9-ab3e-65a7edecdec9/
