### 2021-02-20 2f1c099
if not exists, create openomics configs
