### 2021-02-20 8158b4c
Format code with black and yapf
This commit fixes the style issues introduced in 2f1c099 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/1e797a97-8fad-4fdd-9208-96aeb6e8b367/
