### 2021-02-20 358b27c
if not exists, create openomics configs
