+++ openomics/__init__.py
@@ -38,13 +38,12 @@ if not os.path.exists(user_conf_path):
         base_config = {}
         base_config['cache_dir'] = astropy.config.get_cache_dir(this.__name__)
 
-        with open(user_conf_path, 'w+', encoding='utf-8') as file:
+        with open(user_conf_path, 'w', encoding='utf-8') as file:
             json.dump(base_config, fp=file, indent=4)
-            file.close()
 
 # Read configuration from ~/.openomics/conf.json
 if os.path.exists(user_conf_path):
-    with open(user_conf_path, 'a+', encoding='utf-8') as file:
+    with open(user_conf_path, 'a', encoding='utf-8') as file:
         user_config = json.load(fp=file)
 
     if user_config:
