### 2021-02-20 eca2a88
if not exists, create openomics configs
