### 2021-02-20 9fcaab3
if not exists, create openomics configs
