### 2021-02-20 33bafa6
Format code with black and yapf
This commit fixes the style issues introduced in 9fcaab3 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/0c046878-65e5-4c8b-a33d-1da3291f507d/
