### 2021-02-20 502a4e7
if not exists, create openomics configs
