### 2021-02-20 94a9f80
Format code with black and yapf
This commit fixes the style issues introduced in 502a4e7 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/42eea0f3-129b-49f5-acf8-87bf5454543d/
