### 2021-02-20 09895ae
if not exists, create openomics configs
