### 2021-02-20 80982a1
if not exists, create openomics configs
