### 2021-02-20 12c8b28
Format code with black and yapf
This commit fixes the style issues introduced in 80982a1 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/cc7b79f2-b263-4ae0-958a-5c8d725634a6/
