+++ openomics/database/base.py
@@ -83,9 +83,7 @@ class Dataset(object):
         """
         if validators.url(path):
             for filename, filepath in copy.copy(file_resources).items():
-                data_file = get_pkg_data_filename(
-                    path, filepath, verbose=verbose
-                )  # Download file and replace the file_resource path
+                data_file = get_pkg_data_filename(path, filepath)  # Download file and replace the file_resource path
                 filepath_ext = filetype.guess(data_file)
 
                 # This null if-clause is needed incase when filetype_ext is None, causing the next clauses to fail
