### 2021-02-20 4b2be2c
Format code with black and yapf
This commit fixes the style issues introduced in ac7b810 according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/d1f4a4db-4320-4847-a754-9c62eee750da/
