+++ openomics/database/base.py
@@ -83,7 +83,9 @@ class Dataset(object):
         """
         if validators.url(path):
             for filename, filepath in copy.copy(file_resources).items():
-                data_file = get_pkg_data_filename(path, filepath)  # Download file and replace the file_resource path
+                data_file = get_pkg_data_filename(
+                    path, filepath
+                )  # Download file and replace the file_resource path
                 filepath_ext = filetype.guess(data_file)
 
                 # This null if-clause is needed incase when filetype_ext is None, causing the next clauses to fail
@@ -159,7 +161,13 @@ class Dataset(object):
     def list_databases():
         return DEFAULT_LIBRARIES
 
-    def get_annotations(self, index: str, columns: list, agg: str = "concat", filter_values: pd.Series = None):
+    def get_annotations(
+        self,
+        index: str,
+        columns: list,
+        agg: str = "concat",
+        filter_values: pd.Series = None,
+    ):
         """Returns the Database's DataFrame such that it's indexed by :param
         index:, which then applies a groupby operation and aggregates all other
         columns by concatenating all unique values.
@@ -198,18 +206,23 @@ class Dataset(object):
         #  Aggregate by all columns by concatenating unique values
         if agg == "concat":
             if isinstance(df, pd.DataFrame):
-                aggregated = groupby.agg({col: concat_uniques for col in columns})
+                aggregated = groupby.agg(
+                    {col: concat_uniques
+                     for col in columns})
 
             elif isinstance(df, dd.DataFrame):
                 collect_concat = dd.Aggregation(
-                    name='collect_concat',
+                    name="collect_concat",
                     chunk=lambda s1: s1.apply(list),
                     agg=lambda s2: s2.apply(lambda chunks: filter(
                         lambda x: False if x == "None" or x is None else True,
-                        set(itertools.chain.from_iterable(chunks)))),
-                    finalize=lambda s3: s3.apply(lambda xx: '|'.join(xx))
+                        set(itertools.chain.from_iterable(chunks)),
+                    )),
+                    finalize=lambda s3: s3.apply(lambda xx: "|".join(xx)),
                 )
-                aggregated = groupby.agg({col: collect_concat for col in columns})
+                aggregated = groupby.agg(
+                    {col: collect_concat
+                     for col in columns})
 
         # Any other aggregation functions
         else:
@@ -247,6 +260,7 @@ class Annotatable(ABC):
     downloaded from various databases. The database will be imported as attributes
     information to the genes's annotations, or interactions between the genes.
     """
+
     SEQUENCE_COL_NAME = "sequence"
 
     def get_annotations(self):
@@ -276,7 +290,12 @@ class Annotatable(ABC):
         self.annotations = pd.DataFrame(index=gene_list)
         self.annotations.index.name = index
 
-    def annotate_attributes(self, database: Dataset, on, columns, agg="concat", fuzzy_match=False):
+    def annotate_attributes(self,
+                            database: Dataset,
+                            on,
+                            columns,
+                            agg="concat",
+                            fuzzy_match=False):
         """Performs a left outer join between the annotation and Database's
         DataFrame, on the index key. The index argument must be column present
         in both DataFrames. If there exists overlapping columns from the join, then
@@ -291,7 +310,9 @@ class Annotatable(ABC):
             fuzzy_match (bool): default False. Whether to join the annotation by applying a fuzzy match on the index with difflib.get_close_matches(). It is very computationally expensive and thus should only be used sparingly.
         """
         if not hasattr(self, "annotations"):
-            raise Exception(f"Must run .initialize_annotations() on {self.__class__.__name__} first.")
+            raise Exception(
+                f"Must run .initialize_annotations() on {self.__class__.__name__} first."
+            )
 
[…]
