### 2021-02-20 3695218
added mac os to travis builds
