### 2021-02-20 8b26aa8
added mac os to travis builds
