+++ .travis.yml
@@ -19,8 +19,6 @@ install:
     - pip install -r requirements_dev.txt
     - pip install -r requirements.txt
 
-cache: bundler
-
 script:
     - python -m pytest --cov=./
 
