### 2021-02-20 7b7ddfd
added astropy cache to travis builds
