+++ .travis.yml
@@ -12,7 +12,10 @@ os:
 env:
     - CODECOV_TOKEN="0cfdf951-48ac-4a61-818d-639e07fde3b4"
 
-cache: pip
+cache:
+    -   pip: True
+    -   directories:
+            - /home/travis/.astropy/
 
 install:
     - pip install -U tox-travis codecov
@@ -28,7 +31,7 @@ after_success:
 deploy:
     provider: pypi
     distributions: sdist bdist_wheel
-    user: JonnyTran
+    username: JonnyTran
     password:
         secure: m39qXuHs++1ImJ6Ei7dnGgV5lekiAdAZjfOJcenGVmEMVB7yR/pYs+I3ExLEBEvJ7puSpuKQEUELqhWdzEAWd3sAlP8l495W049StXmm7XyvjQuzBrstthOjneb1jjaAnxnI9pPYllxS9n5UOgpHyGGru+V3VTsKM9W9kEoIYtHgfkMTC2+lZlZm7opn+9RnjALxcn5X9qlhsKLyh8GulLuEbuc1s0GM2xqsWkT/cf24x3Y8tXAyWjyuTHAf26n2BgegRHGfVPOgl6A1mRZAG/mzQs/fXOWtiZUqVyephIKc1p5eMh8X+lDqxMNo0WCMw2UqgySWnP7C1PKno7wkAIDpFH/a+LOJIpBITGRc8wM3OkejyQj5w4IgPJQzbWnGAzBuawNnx/8VTAaA2K38Go96jgHLRZ0HOVPZTF8/qXdZ/JKDAhnBx2TL7y/rpFAC2C6XmqiPRzb3ajEv/04GDRIh+XFglRKIpIk8ZgZAeQwNPdBqL3sammceLYwXFicGiEv0ZlY3++f9e5WsDDKUaXQk6hjYtbhBSIz7Oa6HXxw4eLar8o+9WjF9n7Jv3XiURQoLL02nILqUeaZm3Vp1coRV9wyKhFR+7hPTNozkQO2MPVQLxV1beHtlWZTXnjD9JxfiFG1I+BjEnr/hfgUKtjK42iin7Vx62uF/pWJo4f8=
     on:
