### 2021-02-20 8609deb
added astropy cache to travis builds
