### 2021-02-20 BioMeCIS-Lab/OpenOmics#117 @JonnyTran
Running build at Travis.org temporarily at https://travis-ci.org/github/BioMeCIS-Lab/OpenOmics/builds
