### 2021-02-20 e9d5ed7
added astropy cache to travis builds
