### 2021-02-26 80c52e3
added GitHub Actions CI
