+++ .github/workflows/build.yaml
@@ -1,0 +1,43 @@
+name: OpenOmics
+
+on: [ push ]
+
+jobs:
+    build:
+
+        runs-on: ${{ matrix.os }}
+        strategy:
+            matrix:
+                os: [ ubuntu-latest, macos-latest, windows-latest ]
+                python-version: [ 3.5, 3.6, 3.7, 3.8, 3.9 ]
+
+        steps:
+            -   uses: actions/checkout@v2
+            -   name: Set up Python ${{ matrix.python-version }}
+                uses: actions/setup-python@v2
+                with:
+                    python-version: ${{ matrix.python-version }}
+            -   name: Cache pip
+                uses: actions/cache@v2
+                with:
+                    # This path is specific to Ubuntu
+                    path: ~/.cache/pip
+                    # Look to see if there is a cache hit for the corresponding requirements file
+                    key: ${{ runner.os }}-pip-${{ hashFiles('requirements.txt') }}
+                    restore-keys: |
+                        ${{ runner.os }}-pip-
+                        ${{ runner.os }}-
+            -   name: Install dependencies
+                run: |
+                    python -m pip install --upgrade pip
+                    if [ -f requirements_dev.txt ]; then pip install -r requirements_dev.txt; fi
+                    if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+            -   name: Lint with flake8
+                run: |
+                    # stop the build if there are Python syntax errors or undefined names
+                    flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+                    # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+                    flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+            -   name: Test with pytest
+                run: |
+                    pytest ./
