### 2021-02-26 66b8cfe
added GitHub Actions CI
