### 2021-02-26 7717242
added GitHub Actions CI
