+++ .github/workflows/python-package.yml
@@ -1,6 +1,10 @@
 name: OpenOmics
 
-on: [ push ]
+on:
+    push:
+        branches: [ master ]
+    pull_request:
+        branches: [ master ]
 
 jobs:
     build:
