### 2021-02-26 BioMeCIS-Lab/OpenOmics#117 @JonnyTran
Added GitHub Actions CI at https://github.com/BioMeCIS-Lab/OpenOmics/commit/66b8cfea0c60c008ea4f8250304193783839ae72

This will run pytests with Python 3.5, 3.7, 3.8, 3.9 as well as across Mac OS, Windows, and Linux.

Will close this issue once almost all builds are without errors.
