+++ openomics/database/interaction.py
@@ -46,8 +46,8 @@ class Interactions(Dataset):
         self.network = self.load_network(file_resources=file_resources, source_col_name=source_col_name,
                                          target_col_name=target_col_name,
                                          edge_attr=edge_attr, directed=directed, filters=filters)
-        assert isinstance(self.network, nx.Graph), f"load_network must return a NetworkX Graph or DiGraph but instead " \
-                                                   f"returns {self.network}"
+        assert isinstance(self.network, nx.Graph)
+
         self.network.name = self.name()
 
         if self.network is None:
@@ -117,7 +117,7 @@ class Interactions(Dataset):
 
         for key, values in filters.items():
             if key not in df.columns:
-                logging.info(f"Filter key `{key}` must be in one of {df.columns}")
+                logging.info("Filter key `", key, "` must be in one of ", df.columns)
                 continue
             n_rows = df.shape[0]
 
@@ -131,9 +131,9 @@ class Interactions(Dataset):
             else:
                 df = df[df[key] == values]
 
-            logging.info(f"INFO: Removed {n_rows - df.shape[0]} rows with `{key}` != {values}")
+            logging.info("INFO: Removed ", n_rows - df.shape[0], " rows with `", key, "` != ", values)
 
-        assert df.shape[0] > 0, f"ERROR: Dataframe is empty because of filter: {filters}"
+        assert df.shape[0] > 0, "ERROR: Dataframe is empty because of filter: {filters}"
         return df
 
 
@@ -316,7 +316,7 @@ class LncBase(Interactions, Dataset):
         if edge_attr is None:
             edge_attr = ["tissue", "positive_negative"]
         df = pd.read_table(file_resources["LncBasev2_download.csv"], low_memory=True)
-        logging.info(f"{self.name()}, {df.columns.tolist()}")
+        logging.info(self.name(), df.columns.tolist())
         df.replace({"species": {"Homo Sapiens": "Homo sapiens", "Mus Musculus": "Mus musculus"}}, inplace=True)
 
         df = self.filter_values(df, filters)
@@ -350,7 +350,7 @@ class LncReg(Interactions):
 
     def load_network(self, file_resources, source_col_name, target_col_name, edge_attr, directed, filters):
         df = pd.read_excel(self.file_resources["data.xlsx"])
-        logging.info(f"{self.name()}, {df.columns.tolist()}")
+        logging.info(self.name(), df.columns.tolist())
 
         df = df[df["species"] == "Homo sapiens"]
         df.loc[df["B_category"] == "miRNA", "B_name_in_paper"] = df[df["B_category"] == "miRNA"][
@@ -388,7 +388,7 @@ class lncRInter(Interactions):
 
     def load_network(self, file_resources, source_col_name, target_col_name, edge_attr, directed, filters):
         lncRInter_df = pd.read_table(file_resources["human_interactions.txt"])
-        logging.info(f"{self.name()}, {lncRInter_df.columns.tolist()}")
+        logging.info(self.name(), lncRInter_df.columns.tolist())
 
         lncRInter_df = self.filter_values(lncRInter_df, filters)
         # Data cleaning
@@ -458,7 +458,7 @@ class LncRNA2Target(Interactions):
                                      edge_attr=None, directed=True, filters=None):
         table = pd.read_table(file_resources["lncRNA_target_from_high_throughput_experiments.txt"], sep="\t")
         table = self.filter_values(table, filters)
-        logging.info(f"{self.name()}, {table.columns.tolist()}")
+        logging.info(self.name(), table.columns.tolist())
 
         table["lncrna_symbol"] = table["lncrna_symbol"].str.upper()
[…]
