### 2021-02-26 0746d16
Format code with black and yapf
This commit fixes the style issues introduced in b06f77c according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/dfd1c669-c4e3-4fca-865f-f47c126c696c/
