+++ openomics/database/interaction.py
@@ -10,8 +10,20 @@ from openomics.database.sequence import SequenceDataset
 
 
 class Interactions(Dataset):
-    def __init__(self, path, file_resources, source_col_name=None, target_col_name=None, source_index=None,
-                 target_index=None, edge_attr=None, filters=None, directed=True, relabel_nodes=None, verbose=False, ):
+    def __init__(
+        self,
+        path,
+        file_resources,
+        source_col_name=None,
+        target_col_name=None,
+        source_index=None,
+        target_index=None,
+        edge_attr=None,
+        filters=None,
+        directed=True,
+        relabel_nodes=None,
+        verbose=False,
+    ):
         """
         This is an abstract class used to instantiate a database given a folder containing various file resources. When creating a Database class, the load_data function is called where the file resources are load as a DataFrame and performs necessary processings. This class provides an interface for RNA classes to annotate various genomic annotation, functional annotation, sequences, and disease associations.
         Args:
@@ -43,19 +55,26 @@ class Interactions(Dataset):
         self.file_resources = file_resources
         self.source_index = source_index
         self.target_index = target_index
-        self.network = self.load_network(file_resources=file_resources, source_col_name=source_col_name,
-                                         target_col_name=target_col_name,
-                                         edge_attr=edge_attr, directed=directed, filters=filters)
+        self.network = self.load_network(
+            file_resources=file_resources,
+            source_col_name=source_col_name,
+            target_col_name=target_col_name,
+            edge_attr=edge_attr,
+            directed=directed,
+            filters=filters,
+        )
         assert isinstance(self.network, nx.Graph)
 
         self.network.name = self.name()
 
         if self.network is None:
             raise Exception(
-                "Make sure load_network() returns a Networkx Graph and is called with super().__init__() in the constructor.")
+                "Make sure load_network() returns a Networkx Graph and is called with super().__init__() in the constructor."
+            )
 
         if relabel_nodes is not None:
-            self.network = nx.relabel_nodes(self.network, mapping=relabel_nodes)
+            self.network = nx.relabel_nodes(self.network,
+                                            mapping=relabel_nodes)
 
         self.verbose = verbose
         self.info() if verbose else None
@@ -68,7 +87,15 @@ class Interactions(Dataset):
         return cls.__name__
 
     @abstractmethod
-    def load_network(self, file_resources, source_col_name, target_col_name, edge_attr, directed, filters):
+    def load_network(
+        self,
+        file_resources,
+        source_col_name,
+        target_col_name,
+        edge_attr,
+        directed,
+        filters,
+    ):
         """
         Handles data processing from `file_resources` to a Pandas DataFrame which contain edgelist data, then constructs
         and return a NetworkX Graph.
@@ -100,8 +127,8 @@ class Interactions(Dataset):
         """
         if not hasattr(self, "network"):
             raise Exception(
-                "{} does not have network interaction data yet. Must run load_network() and assign self.network field first.".format(
-                    self.name()))
+                "{} does not have network interaction data yet. Must run load_network() and assign self.network field first."
+                .format(self.name()))
 
         if nodelist is None:
             return self.network.edges(data=data)
@@ -117,13 +144,15 @@ class Interactions(Dataset):
 
         for key, values in filters.items():
             if key not in df.columns:
-                logging.info("Filter key `", key, "` must be in one of ", df.columns)
+                logging.info("Filter key `", key, "` must be in one of ",
+                             df.columns)
                 continue
             n_rows = df.shape[0]
 
[…]

[…]
