### 2021-02-26 bfefbeb
Format code with black and yapf
This commit fixes the style issues introduced in eb2a19f according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/87a65acb-818e-453f-9518-0de14248fbff/
