+++ openomics_web/app.py
@@ -4,12 +4,23 @@ from dash.dependencies import Input, Output, State
 
 from openomics import MultiOmics
 from openomics_web.layouts import app_layout
-from openomics_web.layouts.clinical_view import ClinicalDataColumnSelect, ClinicalDataTable
-from openomics_web.layouts.datatable_view import ExpressionDataTable, DataTableColumnSelect, split_filter_part
+from openomics_web.layouts.clinical_view import (
+    ClinicalDataColumnSelect,
+    ClinicalDataTable,
+)
+from openomics_web.layouts.datatable_view import (
+    ExpressionDataTable,
+    DataTableColumnSelect,
+    split_filter_part,
+)
 from openomics_web.server import server
-from openomics_web.utils.io import get_table_columns, get_expression_data, get_clinical_data
+from openomics_web.utils.io import (
+    get_table_columns,
+    get_expression_data,
+    get_clinical_data,
+)
 
-external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
+external_stylesheets = ["https://codepen.io/chriddyp/pen/bWLwgP.css"]
 
 # running directly with Python
 app = dash.Dash(__name__,
@@ -21,15 +32,19 @@ app.layout = app_layout.app_main()
 user_multiomics = MultiOmics(cohort_name="TEST", )
 
 
-@app.callback([
-    Output('data-table-column-select', 'children'),
-    Output('upload-data-table', 'children')
-], [
-    Input('upload-data-table', 'contents'),
-    Input('upload-data-table', 'filename')
-], [
-    State('data-table-type', 'value'),
-])
+@app.callback(
+    [
+        Output("data-table-column-select", "children"),
+        Output("upload-data-table", "children"),
+    ],
+    [
+        Input("upload-data-table", "contents"),
+        Input("upload-data-table", "filename")
+    ],
+    [
+        State("data-table-type", "value"),
+    ],
+)
 def update_datatable_metadata(
     list_of_contents,
     list_of_names,
@@ -42,31 +57,41 @@ def update_datatable_metadata(
         data_type:
     """
     if list_of_contents is None:
-        return None, ['Drag and Drop or ', html.A('Select Files')]
+        return None, ["Drag and Drop or ", html.A("Select Files")]
 
     try:
         columns = get_table_columns(list_of_contents, list_of_names)
 
     except Exception as e:
         print(e)
-        return None, 'There was an error processing this file.'
+        return None, "There was an error processing this file."
 
     return DataTableColumnSelect(columns), "Uploaded {}".format(list_of_names)
 
 
-@app.callback(Output('output-data-upload', 'children'),
-              [Input('upload-data-table-submit', 'n_clicks')], [
-                  State('data-table-cohort', 'value'),
-                  State('data-table-type', 'value'),
-                  State('upload-data-table', 'contents'),
-                  State('upload-data-table', 'filename'),
-                  State('data-table-genes-col-name', 'value'),
-                  State('data-table-columns-select', 'value'),
-                  State('data-table-transpose', 'value')
-])
-def import_datatable_upload(n_clicks, cohort_name, data_type, list_of_contents,
-                            list_of_names, genes_col_name, columns_select,
-                            transposed):
+@app.callback(
+    Output("output-data-upload", "children"),
+    [Input("upload-data-table-submit", "n_clicks")],
+    [
+        State("data-table-cohort", "value"),
+        State("data-table-type", "value"),
+        State("upload-data-table", "contents"),
+        State("upload-data-table", "filename"),
+        State("data-table-genes-col-name", "value"),
+        State("data-table-columns-select", "value"),
[…]
