+++ .github/workflows/python-package.yml
@@ -37,7 +37,7 @@ jobs:
                     restore-keys: |
                         ${{ runner.os }}-pip-
 
-            -   name: Install dependencies
+            -   name: Install development dependencies
                 run: |
                     python -m pip install --upgrade pip
                     if [ -f requirements_dev.txt ]; then pip install -r requirements_dev.txt; fi
@@ -50,6 +50,18 @@ jobs:
                     # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
                     flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
 
+            -   name: Install requirement dependencies
+                    run: |
+                        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+
             -   name: Test with pytest
                 run: |
                     pytest ./
+
+            -   name: Build and publish
+                    env:
+                        TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
+                        TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
+                    run: |
+                        python setup.py sdist bdist_wheel
+                        twine upload dist/*
