+++ .github/workflows/python-package.yml
@@ -51,17 +51,17 @@ jobs:
                     flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
 
             -   name: Install requirement dependencies
-                    run: |
-                        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+                run: |
+                    if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
             -   name: Test with pytest
                 run: |
                     pytest ./
 
             -   name: Build and publish
-                    env:
-                        TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
-                        TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-                    run: |
-                        python setup.py sdist bdist_wheel
-                        twine upload dist/*
+                env:
+                    TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
+                    TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
+                run: |
+                    python setup.py sdist bdist_wheel
+                    twine upload dist/*
