+++ .github/workflows/python-package.yml
@@ -13,7 +13,7 @@ jobs:
         strategy:
             matrix:
                 os: [ ubuntu-latest, macos-latest, windows-latest ]
-                python-version: [ 3.5, 3.6, 3.7, 3.8, 3.9 ]
+                python-version: [ 3.6, 3.7, 3.8, 3.9 ]
                 include:
                     -   os: ubuntu-latest
                         path: ~/.cache/pip
