+++ .github/workflows/python-package.yml
@@ -67,10 +67,10 @@ jobs:
                 run: |
                     pytest ./
 
-            -   name: Build and publish
-                env:
-                    TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
-                    TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-                run: |
-                    python setup.py sdist bdist_wheel
-                    twine upload dist/*
+#            -   name: Build and publish
+#                env:
+#                    TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
+#                    TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
+#                run: |
+#                    python setup.py sdist bdist_wheel
+#                    twine upload dist/*
