+++ .github/workflows/python-package.yml
@@ -65,7 +65,19 @@ jobs:
 
             -   name: Test with pytest
                 run: |
-                    pytest ./
+                    pytest ./ --cov=./ --cov-report=xml
+            -   name: Upload coverage to Codecov
+                uses: codecov/codecov-action@v1
+                with:
+                    token: ${{ secrets.CODECOV_TOKEN }}
+                    files: ./coverage1.xml,./coverage2.xml
+                    directory: ./coverage/reports/
+                    flags: unittests
+                    env_vars: OS,PYTHON
+                    name: codecov-umbrella
+                    fail_ci_if_error: false
+                    path_to_write_report: ./coverage/codecov_report.txt
+                    verbose: true
 
 #            -   name: Build and publish
 #                env:
