### 2021-02-27 fc77bcb
removed if clause for windows
