+++ .github/workflows/python-package.yml
@@ -40,7 +40,7 @@ jobs:
                     restore-keys: |
                         ${{ runner.os }}-pip-
 
-            -   name: Cache database files
+            -   name: Cache download files
                 uses: actions/cache@v2
                 with:
                     path: ${{ matrix.download_cache }}
@@ -49,8 +49,8 @@ jobs:
             -   name: Install development dependencies
                 run: |
                     python -m pip install --upgrade pip
-                    if [ -f requirements_dev.txt ]; then pip install -r requirements_dev.txt; fi
-                    if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+                    pip install -r requirements_dev.txt
+                    pip install -r requirements.txt
 
             -   name: Lint with flake8
                 run: |
