### 2021-02-27 c1997fe
added cache download files with  folder hash
