+++ .github/workflows/install-linux-packages.yml
@@ -1,0 +1,17 @@
+name: Install Linux packages
+
+on:
+    push:
+        branches: '*'
+
+jobs:
+    build:
+        runs-on: ubuntu-latest
+
+        steps:
+            -   name: Checkout
+                uses: actions/checkout@v2
+
+            -   name: Install package
+                run: |
+                    sudo apt-get -y install unrar
