+++ .github/workflows/python-package.yml
@@ -45,7 +45,8 @@ jobs:
                 with:
                     path: ${{ matrix.download_cache }}
                     key: ${{ hashFiles('**/contents') }}
-                    restore-keys: ${{ hashFiles('*/url') }}
+                    restore-keys: |
+                        ${{ hashFiles('*/url') }}
 
             #            -   name: Cache download files with Folder Hash
             #                uses: theowenyoung/folder-hash@v2
