+++ .github/workflows/python-package.yml
@@ -44,7 +44,7 @@ jobs:
                 uses: actions/cache@v2
                 with:
                     path: ${{ matrix.download_cache }}
-                    key: ${{ hashFiles('**/contents') }}
+                    key: ${{ hashFiles('*/contents') }}
                     restore-keys: |
                         ${{ hashFiles('*/url') }}
 
