+++ .github/workflows/python-package.yml
@@ -44,7 +44,7 @@ jobs:
                 uses: actions/cache@v2
                 with:
                     path: ${{ matrix.download_cache }}
-                    key: ${{ hashFiles('*/contents') }}
+                    key: ${{ hashFiles('*') }}
                     restore-keys: |
                         ${{ hashFiles('*/url') }}
 
