+++ .github/workflows/python-package.yml
@@ -46,7 +46,7 @@ jobs:
                     path: ${{ matrix.download_cache }}
                     key: ${{ hashFiles('*') }}
                     restore-keys: |
-                        ${{ hashFiles('*/url') }}
+                        ${{ hashFiles('*/*/url') }}
 
             #            -   name: Cache download files with Folder Hash
             #                uses: theowenyoung/folder-hash@v2
