+++ .github/workflows/python-package.yml
@@ -17,13 +17,13 @@ jobs:
                 include:
                     -   os: ubuntu-latest
                         pip_cache: ~/.cache/pip
-                        download_cache: ~/.openomics/cache
+                        download_cache: ~/.astropy/cache/download/url/
                     -   os: macos-latest
                         pip_cache: ~/Library/Caches/pip
-                        download_cache: ~/.openomics/cache
+                        download_cache: ~/.astropy/cache/download/url/
                     -   os: windows-latest
                         pip_cache: ~\AppData\Local\pip\Cache
-                        download_cache: ~\.openomics\cache
+                        download_cache: ~\.astropy\cache\download\url\
 
         steps:
             -   uses: actions/checkout@v2
