+++ .github/workflows/python-package.yml
@@ -44,9 +44,9 @@ jobs:
                 uses: actions/cache@v2
                 with:
                     path: ${{ matrix.download_cache }}
-                    key: ${{ hashFiles('*') }}
+                    key: ${{ hashFiles('*/contents') }}
                     restore-keys: |
-                        ${{ hashFiles('*') }}
+                        ${{ hashFiles('*/url') }}
 
             #            -   name: Cache download files with Folder Hash
             #                uses: theowenyoung/folder-hash@v2
