+++ .github/workflows/python-package.yml
@@ -39,14 +39,15 @@ jobs:
                     key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}
                     restore-keys: |
                         ${{ runner.os }}-pip-
+                        ${{ runner.os }}-
 
             -   name: Cache download files
                 uses: actions/cache@v2
                 with:
                     path: ${{ matrix.download_cache }}
-                    key: ${{ hashFiles('*/contents') }}
+                    key: ${{ hashFiles('*') }}
                     restore-keys: |
-                        ${{ hashFiles('*/url') }}
+                        ${{ hashFiles('*') }}
 
             #            -   name: Cache download files with Folder Hash
             #                uses: theowenyoung/folder-hash@v2
