+++ requirements.txt
@@ -27,3 +27,5 @@ rarfile
 filetype
 goatools
 obonet
+furo
+myst-parser
