+++ docs/index.rst
@@ -1,6 +1,9 @@
 Welcome to OpenOmics's documentation!
 ======================================
 
+OpenOmics provides a bioinformatics API and web-app platform integrate, analyze, and visualize the multi-omics and clinical data.
+
+
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
