+++ README.md
@@ -8,10 +8,13 @@
 [![OpenOmics](https://github.com/BioMeCIS-Lab/OpenOmics/actions/workflows/python-package.yml/badge.svg?branch=master)](https://github.com/BioMeCIS-Lab/OpenOmics/actions/workflows/python-package.yml)
 [![codecov](https://codecov.io/gh/BioMeCIS-Lab/OpenOmics/branch/master/graph/badge.svg?token=WAN3PJwM42)](https://codecov.io/gh/BioMeCIS-Lab/OpenOmics)
 
-This Python package provide a series of tools to integrate and query the genomics, transcriptomics, proteomics, and clinical data (aka multi-omics data).
-With scalable data-frame manipulation tools, OpenOmics facilitates the common coding tasks when preparing data for bioinformatics analysis.
+This Python package provide a series of tools to integrate and query the genomics, transcriptomics, proteomics, and
+clinical data (aka multi-omics data). With scalable data-frame manipulation tools, OpenOmics facilitates the common
+coding tasks when preparing data for RNA-seq bioinformatics analysis.
 
-Documentation ([Latest](https://openomics.readthedocs.io/en/latest/) | [Stable](https://openomics.readthedocs.io/en/stable/)) | [OpenOmics at a glance](https://openomics.readthedocs.io/en/stable/usage.html)
+Documentation ([Latest](https://openomics.readthedocs.io/en/latest/)
+| [Stable](https://openomics.readthedocs.io/en/stable/))
+| [OpenOmics at a glance](https://openomics.readthedocs.io/en/stable/usage.html)
 
 ## Features
 OpenOmics assist in integration of heterogeneous multi-omics bioinformatics data. The library provides a Python API as well as an interactive Dash web interface.
@@ -21,16 +24,21 @@ It features support for:
 
 OpenOmics also has an efficient data pipeline that bridges the popular data manipulation Pandas library and Dask distributed processing to address the following use cases:
 
-- Provides a standard pipeline for dataset indexing, table joining and querying, which are transparent and customizable for end-users.
-- Efficient disk storage for large multi-omics dataset with Parquet data structures.
-- Multiple data types that supports both interactions and sequence data, and allows users to export to NetworkX graphs or down-stream machine learning.
-- An easy-to-use API that works seamlessly with external Galaxy tool interface or the built-in Dash web interface (WIP).
+- Providing a standard pipeline for dataset indexing, table joining and querying, which are transparent and customizable
+  for end-users.
+- Providing Efficient disk storage for large multi-omics dataset with Parquet data structures.
+- Integrating various data types including interactions and sequence data, then exporting to NetworkX graphs or data
+  generators for down-stream machine learning.
+- Accessible by both developers and scientists with a Python API that works seamlessly with external Galaxy tool
+  interface or the built-in Dash web interface (WIP).
 
 <br/>
 
 ## Installation via pip:
 
-    pip install openomics
+```
+$ pip install openomics
+```
 
 <br/>
 
