+++ docs/readme.rst
@@ -1,1 +1,47 @@
-.. include:: ../README.md
+=========
+OpenOmics
+=========
+
+.. image:: https://github.com/BioMeCIS-Lab/OpenOmics/raw/master/openomics_web/assets/openomics_logo.png
+
+.. image:: https://img.shields.io/pypi/v/OpenOmics.svg
+        :target: https://pypi.python.org/pypi/OpenOmics
+
+.. image:: https://img.shields.io/travis/BioMeCIS-Lab/OpenOmics.svg
+        :target: https://travis-ci.org/BioMeCIS-Lab/OpenOmics
+
+.. image:: https://codecov.io/gh/BioMeCIS-Lab/OpenOmics/branch/master/graph/badge.svg
+        :target: https://codecov.io/gh/BioMeCIS-Lab/OpenOmics
+
+.. image:: https://readthedocs.org/projects/openomics/badge/?version=latest
+        :target: https://openomics.readthedocs.io/en/latest/?badge=latest
+        :alt: Documentation Status
+
+.. image:: https://pyup.io/repos/github/BioMeCIS-Lab/OpenOmics/shield.svg
+     :target: https://pyup.io/repos/github/BioMeCIS-Lab/OpenOmics/
+     :alt: Updates
+
+
+
+
+* Free software: MIT license
+* Documentation: https://openomics.readthedocs.io.
+
+
+Features
+--------
+
+* Provides a bioinformatics workflow to generate integrative results from multi-omics data.
+* Facilitates integration of various bio-databases, multi-omics expression, genomics, and clinical data.
+* Highly flexible to different data types and missing data.
+* Provides researchers with means to consistently store and explore their experimental datasets.
+* Enables scalable performance with parallel computing, while easily configurable to deploy on both single machine and a cluster.
+
+Credits
+-------
+
+This package was created with Cookiecutter_ and the `pyOpenSci/cookiecutter-pyopensci`_ project template, based off `audreyr/cookiecutter-pypackage`_.
+
+.. _Cookiecutter: https://github.com/audreyr/cookiecutter
+.. _`pyOpenSci/cookiecutter-pyopensci`: https://github.com/pyOpenSci/cookiecutter-pyopensci
+.. _`audreyr/cookiecutter-pypackage`: https://github.com/audreyr/cookiecutter-pypackage
