+++ setup.py
@@ -47,7 +47,7 @@ class UploadCommand(Command):
         sys.exit()
 
 
-with open('README.rst') as readme_file:
+with open('docs/readme.rst') as readme_file:
     readme = readme_file.read()
 
 with open('HISTORY.rst') as history_file:
