+++ docs/installation.md
@@ -1,0 +1,28 @@
+# Installation
+
+## Stable release
+
+To install OpenOmics, run this command in your terminal:
+
+    $ pip install openomics
+
+This is the preferred method to install openOmics, as it will always install the most recent stable release.
+
+If you don't have `pip` installed, it is recommended to use [Anaconda python distribution](https://www.anaconda.com/products/individual) can guide
+you through the process.
+
+
+## From sources
+
+The sources for OpenOmics can be downloaded from the [Github repo](https://github.com/BioMeCIS-Lab/OpenOmics.
+
+You can either clone the public repository:
+
+    $ git clone https://github.com/BioMeCIS-Lab/OpenOmics
+
+Once you have a copy of the source, you can install it with:
+
+    $ cd OpenOmics
+    $ pip install ./ -e
+
+The versioned release history can be found at [OpenOmics/releases](https://github.com/BioMeCIS-Lab/OpenOmics/releases).
