+++ docs/installation.rst
@@ -1,45 +1,0 @@
-.. highlight:: shell
-
-============
-Installation
-============
-
-
-Stable release
---------------
-
-To install OpenOmics, run this command in your terminal:
-
-.. code-block:: console
-
-    $ pip install openomics
-
-This is the preferred method to install openOmics, as it will always install the most recent stable release.
-
-If you don't have `pip`_ installed, this `Python installation guide`_ can guide
-you through the process.
-
-.. _pip: https://pip.pypa.io
-.. _Python installation guide: http://docs.python-guide.org/en/latest/starting/installation/
-
-
-From sources
-------------
-
-The sources for OpenOmics can be downloaded from the `Github repo`_.
-
-You can either clone the public repository:
-
-.. code-block:: console
-
-    $ git clone https://github.com/JonnyTran/OpenOmics
-
-Once you have a copy of the source, you can install it with:
-
-.. code-block:: console
-
-    $ pip install .openomics/ -e
-
-
-.. _Github repo: https://github.com/BioMeCIS-Lab/OpenOmics
-.. _tarball: https://github.com/JonnyTran/BioMeCIS-Lab/tarball/master
