+++ docs/installation.md
@@ -6,10 +6,11 @@ To install OpenOmics, run this command in your terminal:
 
     $ pip install openomics
 
-This is the preferred method to install openOmics, as it will always install the most recent stable release.
+This is the preferred method to install OpenOmics, as it will always install the most recent stable release.
 
-If you don't have `pip` installed, it is recommended to use [Anaconda python distribution](https://www.anaconda.com/products/individual) can guide
-you through the process.
+If you don't have `pip` installed, it is recommended to install
+the [Anaconda Python distribution](https://www.anaconda.com/products/individual) to get started with Python 3.7, 3.8, or
+3.9 on either Mac OS, Linux, or Windows.
 
 
 ## From sources
