+++ CONTRIBUTING.rst
@@ -1,135 +1,0 @@
-.. highlight:: shell
-
-============
-Contributing
-============
-
-Contributions are welcome, and they are greatly appreciated! Every little bit
-helps, and credit will always be given.
-
-You can contribute in many ways:
-
-Types of Contributions
-----------------------
-* Implement general purpose utilities for importing various fasta, gtf and sequencing files.
-* Organize the library of genomics, functional ontologies, interactions, and sequence databases.
-* Documentation standards for the openomics API.
-
-Report Bugs
-~~~~~~~~~~~
-
-Report bugs at https://github.com/BioMeCIS-Lab/openomics/issues.
-
-If you are reporting a bug, please include:
-
-* Your operating system name and version.
-* Any details about your local setup that might be helpful in troubleshooting.
-* Detailed steps to reproduce the bug.
-
-Fix Bugs
-~~~~~~~~
-
-Look through the GitHub issues for bugs. Anything tagged with "bug" and "help
-wanted" is open to whoever wants to implement it.
-
-Implement Features
-~~~~~~~~~~~~~~~~~~
-
-Look through the GitHub issues for features. Anything tagged with "enhancement"
-and "help wanted" is open to whoever wants to implement it.
-
-Write Documentation
-~~~~~~~~~~~~~~~~~~~
-
-openOmics could always use more documentation, whether as part of the
-official openOmics docs, in docstrings, or even on the web in blog posts,
-articles, and such.
-
-Submit Feedback
-~~~~~~~~~~~~~~~
-
-The best way to send feedback is to file an issue at https://github.com/BioMeCIS-Lab/openomics/issues.
-
-If you are proposing a feature:
-
-* Explain in detail how it would work.
-* Keep the scope as narrow as possible, to make it easier to implement.
-* Remember that this is a volunteer-driven project, and that contributions
-  are welcome :)
-
-Get Started!
-------------
-
-Ready to contribute? Here's how to set up `openomics` for local development.
-
-1. Fork the `openomics` repo on GitHub.
-2. Clone your fork locally and work on the develop branch::
-
-    $ git clone git@github.com:your_name_here/openomics.git
-    $ git checkout develop
-
-3. Install your local copy into a virtualenv. Assuming you have virtualenvwrapper installed, this is how you set up your fork for local development::
-
-    $ mkvirtualenv openomics
-    $ cd openomics/
-    $ python setup.py develop
-
-4. Create a branch for local development::
-
-    $ git checkout -b name-of-your-bugfix-or-feature
-
-   Now you can make your changes locally.
-
-5. When you're done making changes, check that your changes pass flake8 and the
-   tests, including testing other Python versions with tox::
-
-    $ flake8 openomics tests
-    $ python setup.py test or py.test
-    $ tox
-
-   To get flake8 and tox, just pip install them into your virtualenv.
-
-6. Commit your changes and push your branch to GitHub::
-
-    $ git add .
-    $ git commit -m "Your detailed description of your changes."
-    $ git push develop name-of-your-bugfix-or-feature
-
-7. Submit a pull request through the GitHub website to the develop branch. Once major features are tested, we can create another pull-request to the master branch.
-
-Pull Request Guidelines
------------------------
-
-Before you submit a pull request, check that it meets these guidelines:
-
-1. The pull request should include tests. Run tests by with `pytest tests/` and make sure tests are 100% passing.
-2. If the pull request adds functionality, the docs should be updated. Put
-   your new functionality into a function with a docstring, and add the
-   feature to the list in README.rst.
-3. The pull request should work for Python 3.6 or higher, and for PyPi. Check
-   https://travis-ci.org/BioMeCIS-Lab/openomics/pull_requests
[…]
