+++ CONTRIBUTORS.rst
@@ -1,13 +1,0 @@
-=======
-Credits
-=======
-
-Development Lead
-----------------
-
-* Nhat (Jonny) Tran <nhat.tran@mavs.uta.edu>
-
-Contributors
-------------
-
-None yet. Why not be the first?
