+++ docs/usage/Importing MultiOmics data.md
@@ -1,22 +1,13 @@
-=====
-Usage
-=====
+# Usage
 
-To use openomics in a project
------------------------------
-
-.. code:: python
-
+# To use openomics in a project
    import openomics
 
+## Import GTEx Tissue-specific gene expression dataset (directly from URL)
 
-Import GTEx Tissue-specific gene expression dataset (directly from URL)
------------------------------------------------------------------------
 We start with tissue-specific gene expression data set from The Genotype-Tissue Expression (GTEx) project. First, we load the GTEx database from the path https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/, which automatically download and parse the gene_median_tpm files to create gene expression matrix. The matrix can be accessed at `gtex.data`.
 
-
-.. code:: python
-
+```
    from openomics.database import GTEx
 
    gtex = GTEx(path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/")
@@ -27,17 +18,24 @@ We start with tissue-specific gene expression data set from The Genotype-Tissue 
    gene_name = pd.Index(gtex_transcripts["gene_name"].unique())
 
    gtex_transcripts.head()
+```
 
-Import GENCODE human release 32
--------------------------------
-Next, we can annotate the genes in our GTEx expression dataset with genomics annotation from GENCODE. In this example, we use the URL path prefix "ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/" which specifies the species and release version. We also pass a dictionary `file_resources`, with key-value pairs where the key is name of file and value is the suffix of the file download URL.
+## Import GENCODE human release 32
 
-For example, file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz"} will download file located at `ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/gencode.v32.long_noncoding_RNAs.gtf.gz` to process the `long_noncoding_RNAs.gtf` file.
+Next, we can annotate the genes in our GTEx expression dataset with genomics annotation from GENCODE. In this example,
+we use the URL path prefix "ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/" which specifies the
+species and release version. We also pass a dictionary `file_resources`, with key-value pairs where the key is name of
+file and value is the suffix of the file download URL.
 
-Here, we loads both "long_noncoding_RNAs.gtf" and "basic.annotation.gtf" which builds a dataframe of combined annotations for both lncRNAs and mRNAs. You can specify different annotation files options from GENCODE by modifying the `file_resources` dict argument.
+For example, file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz"} will download file
+located at `ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/gencode.v32.long_noncoding_RNAs.gtf.gz`
+to process the `long_noncoding_RNAs.gtf` file.
 
-.. code:: python
+Here, we loads both "long_noncoding_RNAs.gtf" and "basic.annotation.gtf" which builds a dataframe of combined
+annotations for both lncRNAs and mRNAs. You can specify different annotation files options from GENCODE by modifying
+the `file_resources` dict argument.
 
+```
    from openomics.database import GENCODE, EnsemblGenes
 
    gencode = ENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
[…]
