+++ requirements.txt
@@ -27,4 +27,5 @@ rarfile
 filetype
 goatools
 obonet
-
+furo
+myst-parser
