+++ docs/usage.md
@@ -7,17 +7,18 @@
 
 We start with tissue-specific gene expression data set from The Genotype-Tissue Expression (GTEx) project. First, we load the GTEx database from the path https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/, which automatically download and parse the gene_median_tpm files to create gene expression matrix. The matrix can be accessed at `gtex.data`.
 
-```
-   from openomics.database import GTEx
+```python
+import pandas as pd
+from openomics.database import GTEx
 
-   gtex = GTEx(path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/")
+gtex = GTEx(path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/")
 
-   gtex_transcripts = gtex.data
+gtex_transcripts = gtex.data
 
-   gene_id = pd.Index(gtex_transcripts["gene_id"].unique())
-   gene_name = pd.Index(gtex_transcripts["gene_name"].unique())
+gene_id = pd.Index(gtex_transcripts["gene_id"].unique())
+gene_name = pd.Index(gtex_transcripts["gene_name"].unique())
 
-   gtex_transcripts.head()
+gtex_transcripts.head()
 ```
 
 ## Import GENCODE human release 32
@@ -35,18 +36,18 @@ Here, we loads both "long_noncoding_RNAs.gtf" and "basic.annotation.gtf" which b
 annotations for both lncRNAs and mRNAs. You can specify different annotation files options from GENCODE by modifying
 the `file_resources` dict argument.
 
-```
-   from openomics.database import GENCODE, EnsemblGenes
+```python
+from openomics.database import GENCODE, EnsemblGenes
 
-   gencode = ENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
-                     file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
-                                     "basic.annotation.gtf": "gencode.v32.basic.annotation.gtf.gz",
-                                     "lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz",
-                                     "transcripts.fa": "gencode.v32.transcripts.fa.gz"},
-                     remove_version_num=True)
+gencode = GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
+                 file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
+                                 "basic.annotation.gtf": "gencode.v32.basic.annotation.gtf.gz",
+                                 "lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz",
+                                 "transcripts.fa": "gencode.v32.transcripts.fa.gz"},
+                 remove_version_num=True)
 
-   # We also loads Ensembl genes to get list of miRNA gene IDs
-   ensembl = EnsemblGenes(biomart='hsapiens_gene_ensembl', npartitions=8, )
+# We also loads Ensembl genes to get list of miRNA gene IDs
+ensembl = EnsemblGenes(biomart='hsapiens_gene_ensembl', npartitions=8, )
 ```
 
 ## Load single omics expressions for MessengerRNA, MicroRNA, LncRNA
@@ -55,25 +56,25 @@ We instantiate the MessengerRNA, MicroRNA and LncRNA -omics expression data from
[…]
