+++ docs/index.rst
@@ -36,7 +36,6 @@ Table of Content
 ----------------
 .. toctree::
    :maxdepth: 2
-   :caption:
 
    readme
    installation
