+++ docs/usage/preprocess-downstream-analysis.md
@@ -1,0 +1,193 @@
+# Preparing data for downstream analyses
+
+## To prepare the data for classification
+
+
+```python
+# This function selects only patients with patholotic stages "Stage I" and "Stage II"
+X_multiomics, y = luad_data.load_dataframe(modalities=["MessengerRNA", "MicroRNA", "LncRNA"], target=['pathologic_stage'],
+                                     pathologic_stages=['Stage I', 'Stage II'])
+print(X_multiomics['MessengerRNA'].shape, X_multiomics['MicroRNA'].shape, X_multiomics['LncRNA'].shape, y.shape)
+```
+
+    (336, 20472) (336, 1870) (336, 12727) (336, 1)
+
+
+```python
+y
+```
+
+
+<div>
+<table border="1" class="dataframe">
+  <thead>
+    <tr style="text-align: right;">
+      <th></th>
+      <th>pathologic_stage</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <th>TCGA-05-4390-01A</th>
+      <td>Stage I</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-4405-01A</th>
+      <td>Stage I</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-4410-01A</th>
+      <td>Stage I</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-4417-01A</th>
+      <td>Stage I</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-4424-01A</th>
+      <td>Stage II</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-4427-01A</th>
+      <td>Stage II</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-4433-01A</th>
+      <td>Stage I</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-5423-01A</th>
+      <td>Stage II</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-5425-01A</th>
+      <td>Stage II</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-5428-01A</th>
+      <td>Stage II</td>
+    </tr>
+    <tr>
+      <th>TCGA-05-5715-01A</th>
+      <td>Stage I</td>
+    </tr>
+    <tr>
+      <th>TCGA-38-4631-01A</th>
+      <td>Stage I</td>
+    </tr>
+    <tr>
+      <th>TCGA-38-7271-01A</th>
+      <td>Stage I</td>
+    </tr>
+    <tr>
+      <th>TCGA-38-A44F-01A</th>
+      <td>Stage I</td>
+    </tr>
+    <tr>
+      <th>TCGA-44-2655-11A</th>
+      <td>Stage I</td>
+    </tr>
+  </tbody>
+</table>
+<p>336 rows × 1 columns</p>
+</div>
+
+
+
+## Log2 transform the mRNA, microRNA, and lncRNA expression values
+
+
+```python
+def expression_val_transform(x):
+    return np.log2(x+1)
+X_multiomics['MessengerRNA'] = X_multiomics['MessengerRNA'].applymap(expression_val_transform)
+X_multiomics['MicroRNA'] = X_multiomics['MicroRNA'].applymap(expression_val_transform)
+# X_multiomics['LncRNA'] = X_multiomics['LncRNA'].applymap(expression_val_transform)
+```
+
[…]
