+++ README.md
@@ -40,533 +40,21 @@ OpenOmics also has an efficient data pipeline that bridges the popular data mani
 $ pip install openomics
 ```
 
-<br/>
-
-# How to use OpenOmics:
-
-
 ## Importing the openomics library
 
 ```python
 from openomics import MultiOmics
 ```
 
-Import TCGA LUAD data included in tests dataset (preprocessed from TCGA-Assembler). It is located at [tests/data/TCGA_LUAD](https://github.com/BioMeCIS-Lab/OpenOmics/tree/master/tests/data/TCGA_LUAD).
-
-```python
-folder_path = "tests/data/TCGA_LUAD/"
-```
-
-Load the multiomics: Gene Expression, MicroRNA expression lncRNA expression, Copy Number Variation, Somatic Mutation, DNA Methylation, and Protein Expression data
-
-```python
-from openomics import MessengerRNA, MicroRNA, LncRNA, SomaticMutation, Protein
-
-# Load each expression dataframe
-mRNA = MessengerRNA(data=folder_path + "LUAD__geneExp.txt",
-                    transpose=True, usecols="GeneSymbol|TCGA", gene_index="GeneSymbol", gene_level="gene_name")
-miRNA = MicroRNA(data=folder_path + "LUAD__miRNAExp__RPM.txt",
-                 transpose=True, usecols="GeneSymbol|TCGA", gene_index="GeneSymbol", gene_level="transcript_name")
-lncRNA = LncRNA(data=folder_path + "TCGA-rnaexpr.tsv",
-                transpose=True, usecols="Gene_ID|TCGA", gene_index="Gene_ID", gene_level="gene_id")
-som = SomaticMutation(data=folder_path + "LUAD__somaticMutation_geneLevel.txt",
-                      transpose=True, usecols="GeneSymbol|TCGA", gene_index="gene_name")
-pro = Protein(data=folder_path + "protein_RPPA.txt",
-              transpose=True, usecols="GeneSymbol|TCGA", gene_index="GeneSymbol", gene_level="protein_name")
-
-# Create an integrated MultiOmics dataset
-luad_data = MultiOmics(cohort_name="LUAD")
-luad_data.add_clinical_data(
-    clinical=folder_path + "nationwidechildrens.org_clinical_patient_luad.txt")
-
-luad_data.add_omic(mRNA)
-luad_data.add_omic(miRNA)
-luad_data.add_omic(lncRNA)
-luad_data.add_omic(som)
-luad_data.add_omic(pro)
-
-luad_data.build_samples()
-```
-
-Each data is stored as a Pandas DataFrame. Below are all the data imported for TCGA LUAD. For each, the first number represents the number of samples, the second number is the number of features.
-
-    PATIENTS (522, 5)
-    SAMPLES (1160, 6)
-    DRUGS (461, 4)
-    MessengerRNA (576, 20472)
-    SomaticMutation (587, 21070)
-    MicroRNA (494, 1870)
-    LncRNA (546, 12727)
-    Protein (364, 154)
-
-## Annotate LncRNAs with GENCODE genomic annotations
-
-```python
-# Import GENCODE database (from URL)
-from openomics.database import GENCODE
-
-gencode = GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
[…]

[…]
