+++ docs/conf.py
@@ -120,6 +120,7 @@ html_theme = 'furo'
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
 
+html_favicon = '../openomics_web/assets/favicon.png'
 
 # -- Options for HTMLHelp output ---------------------------------------
 
