+++ docs/api/openomics.set_cache_dir.rst
@@ -1,0 +1,6 @@
+set_cache_dir
+=============
+
+.. currentmodule:: openomics
+
+.. autofunction:: set_cache_dir
