+++ docs/api/openomics.Protein.rst
@@ -1,0 +1,19 @@
+Protein
+=======
+
+.. currentmodule:: openomics
+
+.. autoclass:: Protein
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~Protein.name
+      ~Protein.process_HPRD_PPI_network
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: name
+   .. automethod:: process_HPRD_PPI_network
