+++ docs/api/openomics.LncRNA.rst
@@ -1,0 +1,17 @@
+LncRNA
+======
+
+.. currentmodule:: openomics
+
+.. autoclass:: LncRNA
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~LncRNA.name
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: name
