+++ docs/api/openomics.Expression.rst
@@ -1,0 +1,47 @@
+Expression
+==========
+
+.. currentmodule:: openomics
+
+.. autoclass:: Expression
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~Expression.features
+      ~Expression.gene_index
+      ~Expression.samples
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~Expression.drop_genes
+      ~Expression.drop_samples
+      ~Expression.get_genes_list
+      ~Expression.get_samples_list
+      ~Expression.load_dataframe
+      ~Expression.load_dataframe_glob
+      ~Expression.name
+      ~Expression.preprocess_table
+      ~Expression.set_genes_index
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: features
+   .. autoattribute:: gene_index
+   .. autoattribute:: samples
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: drop_genes
+   .. automethod:: drop_samples
+   .. automethod:: get_genes_list
+   .. automethod:: get_samples_list
+   .. automethod:: load_dataframe
+   .. automethod:: load_dataframe_glob
+   .. automethod:: name
+   .. automethod:: preprocess_table
+   .. automethod:: set_genes_index
