+++ docs/api/openomics.SomaticMutation.rst
@@ -1,0 +1,17 @@
+SomaticMutation
+===============
+
+.. currentmodule:: openomics
+
+.. autoclass:: SomaticMutation
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~SomaticMutation.name
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: name
