+++ docs/api/openomics.CopyNumberVariation.rst
@@ -1,0 +1,17 @@
+CopyNumberVariation
+===================
+
+.. currentmodule:: openomics
+
+.. autoclass:: CopyNumberVariation
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~CopyNumberVariation.name
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: name
