+++ docs/api/openomics.set_backend.rst
@@ -1,0 +1,6 @@
+set_backend
+===========
+
+.. currentmodule:: openomics
+
+.. autofunction:: set_backend
