+++ docs/API/_autosummary/openomics.multiomics.rst
@@ -1,0 +1,39 @@
+﻿openomics.multiomics
+====================
+
+.. automodule:: openomics.multiomics
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      ClinicalData
+      CopyNumberVariation
+      DNAMethylation
+      Expression
+      LncRNA
+      MessengerRNA
+      MicroRNA
+      MultiOmics
+      Protein
+      SomaticMutation
+      WholeSlideImage
+   
+   
+
+   
+   
+   
+
+
+
