+++ requirements.txt
@@ -30,3 +30,7 @@ obonet
 furo
 myst-parser
 sphinx_autodoc_typehints
+sphinx-inline-tabs
+sphinx-autobuild
+sphinx-copybutton
+sphinx-automodapi
