+++ .readthedocs.yml
@@ -3,7 +3,7 @@ version: 2
 python:
    version: 3.6
    install:
-      - requirements: requirements.txt
+       -   requirements: requirements_dev.txt
       - method: pip
         path: .
         extra_requirements:
