+++ docs/conf.py
@@ -40,7 +40,6 @@ extensions = [
     'myst_parser',
 
     # User Usability
-    'sphinx_autobuild',
     'sphinx_copybutton',
     'sphinx_inline_tabs',
 
@@ -48,7 +47,8 @@ extensions = [
     'sphinx.ext.intersphinx',
     "sphinx.ext.viewcode",
     'sphinx_automodapi.automodapi',
-    'sphinx_automodapi.smart_resolver',
+    # 'sphinx_automodapi.smart_resolver',
+    'sphinx.ext.graphviz',
     'sphinx.ext.inheritance_diagram',
 ]
 
@@ -76,9 +76,9 @@ source_suffix = {
 master_doc = 'index'
 
 # General information about the project.
-project = u'OpenOmics'
-copyright = u"2019, Nhat Tran"
-author = u"Jonny Tran"
+project = 'OpenOmics'
+copyright = "2019, Nhat Tran"
+author = "Jonny Tran"
 
 # The version info for the project you're documenting, acts as replacement
 # for |version| and |release|, also used in various other places throughout
@@ -88,6 +88,7 @@ author = u"Jonny Tran"
 version = openomics.__version__
 # The full version, including alpha/beta/rc tags.
 release = openomics.__version__
+default_role = "obj"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -124,7 +125,7 @@ html_theme = 'furo'
 # documentation.
 
 html_theme_options = {
-    "sidebar_hide_name": True,
+    "sidebar_hide_name": False,
     "light_css_variables": {
         # "color-brand-primary": "#ff5c84",
         # "color-brand-content": "#ff5c84",
@@ -139,6 +140,7 @@ html_favicon = '_static/favicon.png'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+graphviz_dot = "/usr/bin/dot"
 
 # -- Options for HTMLHelp output ---------------------------------------
 
