+++ docs/api/openomics.database.TargetScan.rst
@@ -1,0 +1,21 @@
+TargetScan
+==========
+
+.. currentmodule:: openomics.database
+
+.. autoclass:: TargetScan
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~TargetScan.load_network
+      ~TargetScan.process_interactions_table
+      ~TargetScan.process_miR_family_info_table
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_network
+   .. automethod:: process_interactions_table
+   .. automethod:: process_miR_family_info_table
