+++ docs/modules/modules.rst
@@ -1,7 +1,0 @@
-openomics
-=========
-
-.. toctree::
-   :maxdepth: 4
-
-   openomics
