+++ docs/modules/openomics.utils.rst
@@ -1,38 +1,0 @@
-openomics.utils package
-=======================
-
-Submodules
-----------
-
-openomics.utils.GTF module
---------------------------
-
-.. automodule:: openomics.utils.read_gtf
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-openomics.utils.df module
--------------------------
-
-.. automodule:: openomics.utils.df
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-openomics.utils.io module
--------------------------
-
-.. automodule:: openomics.utils.io
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: openomics.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
