+++ docs/api/openomics.database.sequence.SequenceDataset.rst
@@ -1,0 +1,21 @@
+SequenceDataset
+===============
+
+.. currentmodule:: openomics.database.sequence
+
+.. autoclass:: SequenceDataset
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~SequenceDataset.get_aggregator
+      ~SequenceDataset.get_sequences
+      ~SequenceDataset.read_fasta
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: get_aggregator
+   .. automethod:: get_sequences
+   .. automethod:: read_fasta
