+++ docs/api/openomics.database.STRING.rst
@@ -1,0 +1,29 @@
+STRING
+======
+
+.. currentmodule:: openomics.database
+
+.. autoclass:: STRING
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~STRING.COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~STRING.get_sequences
+      ~STRING.load_network
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: get_sequences
+   .. automethod:: load_network
