+++ docs/modules/openomics.database.sequence.md
@@ -1,0 +1,5 @@
+# openomics.Database.Sequence
+
+```{eval-rst}
+.. automodapi:: openomics.database.sequence
+```
