+++ docs/conf.py
@@ -88,7 +88,6 @@ author = "Jonny Tran"
 version = openomics.__version__
 # The full version, including alpha/beta/rc tags.
 release = openomics.__version__
-default_role = "obj"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
