+++ docs/api/openomics.database.Annotatable.rst
@@ -1,0 +1,47 @@
+Annotatable
+===========
+
+.. currentmodule:: openomics.database
+
+.. autoclass:: Annotatable
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~Annotatable.DISEASE_ASSOCIATIONS_COL
+      ~Annotatable.SEQUENCE_COL_NAME
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~Annotatable.annotate_attributes
+      ~Annotatable.annotate_diseases
+      ~Annotatable.annotate_expressions
+      ~Annotatable.annotate_interactions
+      ~Annotatable.annotate_sequences
+      ~Annotatable.get_annotation_expressions
+      ~Annotatable.get_annotations
+      ~Annotatable.get_rename_dict
+      ~Annotatable.initialize_annotations
+      ~Annotatable.set_index
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: DISEASE_ASSOCIATIONS_COL
+   .. autoattribute:: SEQUENCE_COL_NAME
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: annotate_attributes
+   .. automethod:: annotate_diseases
+   .. automethod:: annotate_expressions
+   .. automethod:: annotate_interactions
+   .. automethod:: annotate_sequences
+   .. automethod:: get_annotation_expressions
+   .. automethod:: get_annotations
+   .. automethod:: get_rename_dict
+   .. automethod:: initialize_annotations
+   .. automethod:: set_index
