+++ docs/api/openomics.database.LncRNADisease.rst
@@ -1,0 +1,27 @@
+LncRNADisease
+=============
+
+.. currentmodule:: openomics.database
+
+.. autoclass:: LncRNADisease
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~LncRNADisease.COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~LncRNADisease.load_dataframe
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_dataframe
