+++ docs/modules/openomics.database.rst
@@ -1,38 +1,0 @@
-openomics.database package
-==========================
-
-Submodules
-----------
-
-openomics.database.annotation module
-------------------------------------
-
-.. automodule:: openomics.database.annotation
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-openomics.database.disease module
----------------------------------
-
-.. automodule:: openomics.database.disease
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-openomics.database.interaction module
--------------------------------------
-
-.. automodule:: openomics.database.interaction
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: openomics.database
-   :members:
-   :undoc-members:
-   :show-inheritance:
