+++ docs/api/openomics.database.LncBase.rst
@@ -1,0 +1,19 @@
+LncBase
+=======
+
+.. currentmodule:: openomics.database
+
+.. autoclass:: LncBase
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~LncBase.get_rename_dict
+      ~LncBase.load_network
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: get_rename_dict
+   .. automethod:: load_network
