+++ docs/modules/openomics.multiomics.md
@@ -1,0 +1,5 @@
+# openomics.MultiOmics
+
+```{eval-rst}
+.. automodapi:: openomics
+```
