+++ docs/api/openomics.database.BioGRID.rst
@@ -1,0 +1,17 @@
+BioGRID
+=======
+
+.. currentmodule:: openomics.database
+
+.. autoclass:: BioGRID
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~BioGRID.load_network
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_network
