+++ docs/api/openomics.database.GeneMania.rst
@@ -1,0 +1,17 @@
+GeneMania
+=========
+
+.. currentmodule:: openomics.database
+
+.. autoclass:: GeneMania
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~GeneMania.load_network
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_network
