+++ docs/api/openomics.database.MalaCards.rst
@@ -1,0 +1,27 @@
+MalaCards
+=========
+
+.. currentmodule:: openomics.database
+
+.. autoclass:: MalaCards
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~MalaCards.COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~MalaCards.load_dataframe
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_dataframe
