+++ docs/modules/openomics.database.annotation.md
@@ -1,0 +1,5 @@
+# openomics.Database.Annotation
+
+```{eval-rst}
+.. automodapi:: openomics.database.annotation
+```
