+++ docs/api/openomics.database.interaction.MiRTarBase.rst
@@ -1,0 +1,17 @@
+MiRTarBase
+==========
+
+.. currentmodule:: openomics.database.interaction
+
+.. autoclass:: MiRTarBase
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~MiRTarBase.load_network
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_network
