+++ docs/api/openomics.database.annotation.BioMartManager.rst
@@ -1,0 +1,21 @@
+BioMartManager
+==============
+
+.. currentmodule:: openomics.database.annotation
+
+.. autoclass:: BioMartManager
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~BioMartManager.cache_dataset
+      ~BioMartManager.query_biomart
+      ~BioMartManager.retrieve_dataset
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: cache_dataset
+   .. automethod:: query_biomart
+   .. automethod:: retrieve_dataset
