+++ docs/api/openomics.database.annotation.EnsemblGeneSequences.rst
@@ -1,0 +1,7 @@
+EnsemblGeneSequences
+====================
+
+.. currentmodule:: openomics.database.annotation
+
+.. autoclass:: EnsemblGeneSequences
+   :show-inheritance:
