+++ docs/api/openomics.database.ontology.Ontology.rst
@@ -1,0 +1,43 @@
+Ontology
+========
+
+.. currentmodule:: openomics.database.ontology
+
+.. autoclass:: Ontology
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~Ontology.DELIM
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~Ontology.filter_annotation
+      ~Ontology.filter_network
+      ~Ontology.get_adjacency_matrix
+      ~Ontology.get_child_nodes
+      ~Ontology.get_dfs_paths
+      ~Ontology.get_node_color
+      ~Ontology.get_root_nodes
+      ~Ontology.load_network
+      ~Ontology.remove_predecessor_terms
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: DELIM
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: filter_annotation
+   .. automethod:: filter_network
+   .. automethod:: get_adjacency_matrix
+   .. automethod:: get_child_nodes
+   .. automethod:: get_dfs_paths
+   .. automethod:: get_node_color
+   .. automethod:: get_root_nodes
+   .. automethod:: load_network
+   .. automethod:: remove_predecessor_terms
