+++ docs/api/openomics.database.annotation.NONCODE.rst
@@ -1,0 +1,17 @@
+NONCODE
+=======
+
+.. currentmodule:: openomics.database.annotation
+
+.. autoclass:: NONCODE
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~NONCODE.load_dataframe
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_dataframe
