+++ docs/api/openomics.database.annotation.EnsemblSNP.rst
@@ -1,0 +1,7 @@
+EnsemblSNP
+==========
+
+.. currentmodule:: openomics.database.annotation
+
+.. autoclass:: EnsemblSNP
+   :show-inheritance:
