+++ docs/api/openomics.database.annotation.TANRIC.rst
@@ -1,0 +1,19 @@
+TANRIC
+======
+
+.. currentmodule:: openomics.database.annotation
+
+.. autoclass:: TANRIC
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~TANRIC.get_expressions
+      ~TANRIC.load_dataframe
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: get_expressions
+   .. automethod:: load_dataframe
