+++ docs/api/openomics.database.annotation.GTEx.rst
@@ -1,0 +1,27 @@
+GTEx
+====
+
+.. currentmodule:: openomics.database.annotation
+
+.. autoclass:: GTEx
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~GTEx.COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~GTEx.load_dataframe
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_dataframe
