+++ docs/api/openomics.database.annotation.EnsemblGenes.rst
@@ -1,0 +1,31 @@
+EnsemblGenes
+============
+
+.. currentmodule:: openomics.database.annotation
+
+.. autoclass:: EnsemblGenes
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~EnsemblGenes.COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~EnsemblGenes.get_functional_annotations
+      ~EnsemblGenes.get_rename_dict
+      ~EnsemblGenes.load_data
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: get_functional_annotations
+   .. automethod:: get_rename_dict
+   .. automethod:: load_data
