+++ docs/api/openomics.database.disease.DisGeNet.rst
@@ -1,0 +1,27 @@
+DisGeNet
+========
+
+.. currentmodule:: openomics.database.disease
+
+.. autoclass:: DisGeNet
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~DisGeNet.COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~DisGeNet.load_dataframe
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_dataframe
