+++ docs/api/openomics.database.annotation.EnsemblSomaticVariation.rst
@@ -1,0 +1,7 @@
+EnsemblSomaticVariation
+=======================
+
+.. currentmodule:: openomics.database.annotation
+
+.. autoclass:: EnsemblSomaticVariation
+   :show-inheritance:
