+++ docs/api/openomics.database.disease.HMDD.rst
@@ -1,0 +1,27 @@
+HMDD
+====
+
+.. currentmodule:: openomics.database.disease
+
+.. autoclass:: HMDD
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~HMDD.COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~HMDD.load_dataframe
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: load_dataframe
