+++ docs/api/openomics.database.ontology.write_taxonomy.rst
@@ -1,0 +1,6 @@
+write_taxonomy
+==============
+
+.. currentmodule:: openomics.database.ontology
+
+.. autofunction:: write_taxonomy
