+++ docs/api/openomics.database.ontology.flatten.rst
@@ -1,0 +1,6 @@
+flatten
+=======
+
+.. currentmodule:: openomics.database.ontology
+
+.. autofunction:: flatten
