+++ docs/api/openomics.database.ontology.HumanPhenotypeOntology.rst
@@ -1,0 +1,7 @@
+HumanPhenotypeOntology
+======================
+
+.. currentmodule:: openomics.database.ontology
+
+.. autoclass:: HumanPhenotypeOntology
+   :show-inheritance:
