+++ docs/api/openomics.database.annotation.EnsemblTranscriptSequences.rst
@@ -1,0 +1,7 @@
+EnsemblTranscriptSequences
+==========================
+
+.. currentmodule:: openomics.database.annotation
+
+.. autoclass:: EnsemblTranscriptSequences
+   :show-inheritance:
