+++ .readthedocs.yml
@@ -1,13 +1,13 @@
-formats: [ htmlzip ]
+formats: [htmlzip]
 version: 2
 python:
-    version: 3.6
-    install:
-        -   requirements: requirements_dev.txt
-        -   method: pip
-            path: .
-            extra_requirements:
-                - docs
-        -   method: setuptools
-            path: .
-    system_packages: true
+   version: 3.6
+   install:
+      - requirements: requirements_dev.txt
+      - method: pip
+        path: .
+        extra_requirements:
+            - docs
+      - method: setuptools
+        path: .
+   system_packages: true
