+++ openomics/clinical.py
@@ -16,8 +16,9 @@ NORMAL = "Normal"
 
 
 class ClinicalData:
-    """This class manages the clinical data tables to handle the phenotype,
-    treatment, and sample data associated to a patient.
+    """This class manages the clinical data tables to handle the patient's
+    phenotype data, as well as the treatment, and sample data associated to each
+    patient.
     """
 
     pathologic_stage_map = {'Stage IA': 'Stage I', 'Stage IB': 'Stage I',
@@ -73,9 +74,7 @@ class ClinicalData:
 
     @classmethod
     def name(self):
-        """
-        Returns the name of the class, i.e. 'ClinicalData'
-        """
+        """Returns the name of the class, i.e. 'ClinicalData'"""
         return self.__class__.__name__
 
     def build_clinical_samples(self, all_samples, index="bcr_patient_barcode"):
