+++ openomics/database/base.py
@@ -1,3 +1,4 @@
+from typing import Dict, Tuple, List
 import copy
 import difflib
 import gzip
@@ -19,6 +20,14 @@ from openomics.utils.io import get_pkg_data_filename
 
 
 class Dataset(object):
+    """This is a class used to instantiate a Dataset given a a set of files from
+    either local files or URLs. When creating a Dataset class, the
+    `load_dataframe()` function is called where the file_resources are used to
+    load (Pandas or Dask) DataFrames, then performs data wrangling to yield a
+    dataframe at `self.data` . This class also provides an interface for -omics
+    tables, e.g. `ExpressionData` , to annotate various annotations,
+    expressions, sequences, and disease associations.
+    """
     COLUMNS_RENAME_DICT = None  # Needs initialization since subclasses may use this field to rename columns in dataframes.
 
     def __init__(
@@ -29,17 +38,20 @@ class Dataset(object):
         npartitions=None,
         verbose=False,
     ):
-        """This is a class used to instantiate a Dataset given a a set of files from either local files or URLs. When
-        creating a Dataset class, the `load_dataframe()` function is called where the file_resources are used to
-        load (Pandas or Dask) DataFrames, then performs data wrangling to yield a dataframe at `self.data`. This class
-        also provides an interface for -omics tables, e.g. `ExpressionData`, to annotate various annotations, expressions,
-        sequences, and disease associations.
-
+        """
         Args:
-            path: The folder or url path containing the data file resources. If url path, the files will be downloaded and cached to the user's home folder (at ~/.astropy/).
-            file_resources: Used to list required files for preprocessing of the database. A dictionary where keys are required filenames and value are file paths. If None, then the class constructor should automatically build the required file resources dict.
-            col_rename (dict): default None, A dictionary to rename columns in the data table. If None, then automatically load defaults.
-            npartitions (int): [0-n], default 0 If 0, then uses a Pandas DataFrame, if >1, then creates an off-memory Dask DataFrame with n partitions
+            path: The folder or url path containing the data file resources. If
+                url path, the files will be downloaded and cached to the user's
+                home folder (at ~/.astropy/).
+            file_resources: Used to list required files for preprocessing of the
+                database. A dictionary where keys are required filenames and
+                value are file paths. If None, then the class constructor should
+                automatically build the required file resources dict.
+            col_rename (dict): default None, A dictionary to rename columns in
+                the data table. If None, then automatically load defaults.
+            npartitions (int): [0-n], default 0 If 0, then uses a Pandas
+                DataFrame, if >1, then creates an off-memory Dask DataFrame with
+                n partitions
             verbose (bool): Default False.
         """
         self.npartitions = npartitions
@@ -66,20 +78,21 @@ class Dataset(object):
                                 file_resources,
                                 npartitions=None,
                                 verbose=False) -> None:
-        """For each file in file_resources, download the file if path+file is a URL
-        or load from disk if a local path. Additionally unzip or unrar if the
-        file is compressed.
+        """For each file in file_resources, download the file if path+file is a
+        URL or load from disk if a local path. Additionally unzip or unrar if
+        the file is compressed.
 
         Args:
             path (str): The folder or url path containing the data file
-                resources. If a url path, the files will be downloaded and cached
-                to the user's home folder (at ~/.astropy/).
+                resources. If a url path, the files will be downloaded and
+                cached to the user's home folder (at ~/.astropy/).
             file_resources (dict): default None, Used to list required files for
                 preprocessing of the database. A dictionary where keys are
                 required filenames and value are file paths. If None, then the
[…]
