+++ docs/modules/openomics.database.disease.md
@@ -1,4 +1,4 @@
-# openomics.Database.Disease
+# Disease databases
 
 ```{eval-rst}
 .. automodapi:: openomics.database.disease
