+++ docs/modules/openomics.database.annotation.md
@@ -1,4 +1,4 @@
-# openomics.Database.Annotation
+# Genomic databases
 
 ```{eval-rst}
 .. automodapi:: openomics.database.annotation
