+++ docs/index.md
@@ -2,7 +2,6 @@
 
 [![PyPI version](https://badge.fury.io/py/openomics.svg)](https://badge.fury.io/py/openomics)
 [![Documentation Status](https://readthedocs.org/projects/openomics/badge/?version=latest)](https://openomics.readthedocs.io/en/latest/?badge=latest)
-[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![OpenOmics](https://github.com/BioMeCIS-Lab/OpenOmics/actions/workflows/python-package.yml/badge.svg?branch=master)](https://github.com/BioMeCIS-Lab/OpenOmics/actions/workflows/python-package.yml)
 [![codecov](https://codecov.io/gh/JonnyTran/OpenOmics/branch/master/graph/badge.svg)](https://codecov.io/gh/JonnyTran/OpenOmics)
 
@@ -25,7 +24,7 @@ coding tasks when preparing data for RNA-seq bioinformatics analysis.
 ## Table of Content
 
 ```{toctree}
-:maxdepth: 1
+:maxdepth: 2
 :caption: Using OpenOmics
 :name: mastertoc
 installation
@@ -36,17 +35,21 @@ usage/network-data
 usage/preprocess-downstream-analysis
 ```
 
+
 ```{toctree}
 :maxdepth: 1
-:caption: Reference and Contributing
+:caption: Contributing and releases
 contributing
+history
 ```
 
+
 ```{toctree}
 :maxdepth: 1
-:caption: API
+:caption: API Reference
 
 modules/openomics.multiomics
+modules/openomics.annotate
 modules/openomics.database.annotation
 modules/openomics.database.sequence
 modules/openomics.database.interaction
