+++ docs/modules/openomics.database.md
@@ -1,5 +1,0 @@
-# openomics.Database
-
-```{eval-rst}
-.. automodapi:: openomics.database
-```
