+++ docs/modules/openomics.database.interaction.md
@@ -1,4 +1,4 @@
-# openomics.Database.Interaction
+# Interaction databases
 
 ```{eval-rst}
 .. automodapi:: openomics.database.interaction
