+++ docs/history.md
@@ -1,0 +1,19 @@
+# Release history
+
+## 0.8.5 (Pending)
+- Build web-app dashboard interface for importing user multiomics data files
+
+
+## 0.8.4 (2020-01-07)
+- Replace Pandas Dataframe with Dask DataFrame for parallel processings
+
+
+## 0.7.2 (2019-09-01)
+- Added compatibility for Python 2.7
+- Refactored ClinicalData
+- Built working documentations with Sphinx on readthedocs
+- Added pytests for MultiOmicsData
+- First successful build on Travis CI on Python 3.4-3.7, 2.7
+
+## 0.6.0 (2019-08-31)
+- First release on PyPI.
