+++ docs/modules/openomics.database.ontology.md
@@ -1,4 +1,4 @@
-# openomics.Database.Ontology
+# Ontology databases
 
 ```{eval-rst}
 .. automodapi:: openomics.database.ontology
