+++ docs/modules/openomics.annotate.md
@@ -1,0 +1,5 @@
+# Annotation interfaces
+
+```{eval-rst}
+.. automodapi:: openomics.database.base
+```
