+++ docs/api/openomics.database.base.Dataset.rst
@@ -1,0 +1,41 @@
+Dataset
+=======
+
+.. currentmodule:: openomics.database.base
+
+.. autoclass:: Dataset
+   :show-inheritance:
+
+   .. rubric:: Attributes Summary
+
+   .. autosummary::
+
+      ~Dataset.COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~Dataset.close
+      ~Dataset.get_annotations
+      ~Dataset.get_expressions
+      ~Dataset.info
+      ~Dataset.list_databases
+      ~Dataset.load_dataframe
+      ~Dataset.name
+      ~Dataset.validate_file_resources
+
+   .. rubric:: Attributes Documentation
+
+   .. autoattribute:: COLUMNS_RENAME_DICT
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: close
+   .. automethod:: get_annotations
+   .. automethod:: get_expressions
+   .. automethod:: info
+   .. automethod:: list_databases
+   .. automethod:: load_dataframe
+   .. automethod:: name
+   .. automethod:: validate_file_resources
