+++ docs/api/openomics.multiomics.MultiOmics.rst
@@ -1,0 +1,35 @@
+MultiOmics
+==========
+
+.. currentmodule:: openomics.multiomics
+
+.. autoclass:: MultiOmics
+   :show-inheritance:
+
+   .. rubric:: Methods Summary
+
+   .. autosummary::
+
+      ~MultiOmics.add_clinical_data
+      ~MultiOmics.add_omic
+      ~MultiOmics.annotate_patients
+      ~MultiOmics.build_samples
+      ~MultiOmics.get_omics_list
+      ~MultiOmics.get_patients_clinical
+      ~MultiOmics.load_data
+      ~MultiOmics.match_samples
+      ~MultiOmics.print_sample_sizes
+      ~MultiOmics.remove_duplicate_genes
+
+   .. rubric:: Methods Documentation
+
+   .. automethod:: add_clinical_data
+   .. automethod:: add_omic
+   .. automethod:: annotate_patients
+   .. automethod:: build_samples
+   .. automethod:: get_omics_list
+   .. automethod:: get_patients_clinical
+   .. automethod:: load_data
+   .. automethod:: match_samples
+   .. automethod:: print_sample_sizes
+   .. automethod:: remove_duplicate_genes
