+++ docs/contributing.md
@@ -14,7 +14,7 @@ You can contribute in many ways:
 
 ## Report Bugs
 
-Report bugs at https://github.com/BioMeCIS-Lab/openomics/issues.
+Report bugs at [openomics/issues](https://github.com/BioMeCIS-Lab/openomics/issues).
 
 If you are reporting a bug, please include:
 
@@ -40,7 +40,7 @@ articles, and such.
 
 ## Submit Feedback
 
-The best way to send feedback is to file an issue at https://github.com/BioMeCIS-Lab/openomics/issues.
+The best way to send feedback is to file an issue at [openomics/issues](https://github.com/BioMeCIS-Lab/openomics/issues).
 
 If you are proposing a feature:
 
@@ -54,36 +54,39 @@ If you are proposing a feature:
 Ready to contribute? Here's how to set up `openomics` for local development.
 
 1. Fork the `openomics` repo on GitHub.
-2. Clone your fork locally and work on the develop branch::
+2. Clone your fork locally and work on the develop branch:
+```
+$ git clone git@github.com:your_name_here/openomics.git
+$ git checkout develop
+```
 
-    $ git clone git@github.com:your_name_here/openomics.git
-    $ git checkout develop
+3. Install your local copy into a virtualenv. Assuming you have virtualenvwrapper installed, this is how you set up your fork for local development:
+```
+$ mkvirtualenv openomics
+$ cd openomics/
+$ python setup.py develop
+```
 
-3. Install your local copy into a virtualenv. Assuming you have virtualenvwrapper installed, this is how you set up your fork for local development::
-
-    $ mkvirtualenv openomics
-    $ cd openomics/
-    $ python setup.py develop
-
-4. Create a branch for local development::
-
-    $ git checkout -b name-of-your-bugfix-or-feature
-
+4. Create a branch for local development:
+```
+$ git checkout -b name-of-your-bugfix-or-feature
+```
    Now you can make your changes locally.
 
 5. When you're done making changes, check that your changes pass flake8 and the
-   tests, including testing other Python versions with tox::
-
-    $ flake8 openomics tests
-   $ python setup.py test or py.test $ tox
+   tests, including testing other Python versions with tox:
+```
+$ flake8 openomics tests
+$ python setup.py test or py.test $ tox
+```
 
    To get flake8 and tox, just pip install them into your virtualenv.
 
-6. Commit your changes and push your branch to GitHub::
-
-   $ git add . $ git commit -m "Your detailed description of your changes."
-   $ git push develop name-of-your-bugfix-or-feature
-
+6. Commit your changes and push your branch to GitHub:
+```
+$ git add . $ git commit -m "Your detailed description of your changes."
+$ git push develop name-of-your-bugfix-or-feature
+```
 7. Submit a pull request through the GitHub website to the develop branch. Once major features are tested, we can create
    another pull-request to the **master** branch.
 
@@ -93,9 +96,9 @@ Before you submit a pull request, check that it meets these guidelines:
 
 1. The pull request should include tests. Run tests by with `pytest ./` and make sure tests are 100% passing.
 2. If the pull request adds functionality, the docs should be updated. Put your new functionality into a function with a
-   docstring, and add the feature to the list in https://github.com/BioMeCIS-Lab/OpenOmics/blob/master/docs/index.md.
+   docstring, and add the feature to the list in [docs/history.md](https://github.com/BioMeCIS-Lab/OpenOmics/blob/master/docs/history.md).
 3. The pull request should work for Python 3.6 or higher, and for PyPi. Check
-   https://github.com/BioMeCIS-Lab/OpenOmics/actions/workflows/python-package.yml
[…]
