+++ docs/api/openomics.database.BioGRID.rst
@@ -1,17 +1,0 @@
-BioGRID
-=======
-
-.. currentmodule:: openomics.database
-
-.. autoclass:: BioGRID
-   :show-inheritance:
-
-   .. rubric:: Methods Summary
-
-   .. autosummary::
-
-      ~BioGRID.load_network
-
-   .. rubric:: Methods Documentation
-
-   .. automethod:: load_network
