+++ docs/usage/import-your-dataset.md
@@ -1,5 +1,10 @@
 # Loading a multi-omics dataset
 
+Suppose you have your own -omics dataset(s) and you'd like to load them. One of OpenOmics's primary goal is to
+encapsulate the data import process with one line of code along with a few parameters. Given any processed single-omic
+dataset, the library loads the data as a tabular structure where rows correspond to observation samples and columns
+correspond to measurements of different biomolecules.
+
 Import TCGA LUAD data included in tests dataset (preprocessed from TCGA-Assembler). It is located at [tests/data/TCGA_LUAD](https://github.com/BioMeCIS-Lab/OpenOmics/tree/master/tests/data/TCGA_LUAD).
 
 ```{code-block} python
