+++ docs/api/openomics.database.GENCODE.rst
@@ -1,23 +1,0 @@
-GENCODE
-=======
-
-.. currentmodule:: openomics.database
-
-.. autoclass:: GENCODE
-   :show-inheritance:
-
-   .. rubric:: Methods Summary
-
-   .. autosummary::
-
-      ~GENCODE.get_rename_dict
-      ~GENCODE.get_sequences
-      ~GENCODE.load_dataframe
-      ~GENCODE.read_fasta
-
-   .. rubric:: Methods Documentation
-
-   .. automethod:: get_rename_dict
-   .. automethod:: get_sequences
-   .. automethod:: load_dataframe
-   .. automethod:: read_fasta

[…]
