+++ openomics/transcriptomics.py
@@ -1,4 +1,7 @@
+from typing import Union, Dict
+import logging
 import io
+import re
 import os
 from glob import glob
 from typing import Union
@@ -38,7 +41,8 @@ class Expression(object):
         transform_fn=None,
         dropna=False,
         npartitions=None,
-        cohort_name=None,
+        prefix=None,
+        **kwargs,
     ):
         """This class handles importing of any quantitative omics data that is
         in a table format (e.g. csv, tsv, excel). Pandas will load the DataFrame
@@ -75,16 +79,11 @@ class Expression(object):
             npartitions (int): [0-n], default 0 If 0, then uses a Pandas
                 DataFrame, if >1, then creates an off-memory Dask DataFrame with
                 n partitions
-            cohort_name (str): The unique cohort code name string.
         """
-        self.cohort_name = cohort_name
         self.gene_level = gene_level
         self.sample_level = sample_level
 
-        df = self.load_dataframe(data,
-                                 transpose=transpose,
-                                 usecols=usecols,
-                                 gene_index=gene_index)
+        df = self.load_dataframe(data, transpose=transpose, usecols=usecols, gene_index=gene_index, dropna=dropna,)
         self.expressions = self.preprocess_table(
             df,
             usecols=usecols,
@@ -113,59 +112,46 @@ class Expression(object):
     def gene_index(self):
         return self.expressions.columns.name
 
-    def load_dataframe(self, data: Union[str, pd.DataFrame, dd.DataFrame], transpose: bool, usecols: str,
-                       gene_index: str):
+    def load_dataframe(self, data: Union[str, pd.DataFrame, dd.DataFrame, io.StringIO], transpose: bool, usecols: str,
+                       gene_index: str, dropna: bool):
         """Reading table data inputs to create a DataFrame.
 
         Args:
-            data: either a file path, a glob file path (e.g. "table- *.tsv"), a
+            data: either a file path, a glob file path (e.g. "table-*.tsv"), a
                 pandas.DataFrame, or a dask DataFrame.
             transpose (bool): True if table oriented with samples columns, else
                 False.
             usecols (str): A regex string to select columns. Default None.
             gene_index (str):
+            dropna (bool): Whether to drop rows with null values
         """
         if isinstance(data, pd.DataFrame):
             df = data
-            df = df.reset_index()
+
         elif isinstance(data, dd.DataFrame):
             df = data
-        elif "*" in data:
-            df = self.load_dataframe_glob(data, usecols, gene_index, transpose)
+
+        elif isinstance(data, str) and "*" in data:
+            df = self.load_dataframe_glob(globstring=data, usecols=usecols, gene_index=gene_index, transpose=transpose,
+                                          dropna=dropna)
+            # print(df.head())
+
         elif isinstance(data, io.StringIO):
-            data.seek(
-                0
-            )  # Needed since the file was previous read to extract columns information
+            # Needed since the file was previous read to extract columns information
+            data.seek(0)
             df = pd.read_table(data)
+
         elif isinstance(data, str) and os.path.isfile(data):
             df = pd.read_table(data, sep=None, engine="python")
+
         else:
             raise IOError(data)
 
         # TODO implement handling for multiple file ByteIO
         return df
 
-    def load_dataframe_glob(self, globstring: str, usecols: str, genes_index: str, transpose: bool):
-        # type: (str, str, str, bool) -> dd.DataFrame
-        """
-        Args:
-            globstring (str):
-            usecols (str):
-            genes_index (str):
-            transpose (bool):
-        """
-        lazy_dataframes = []
-        for file_path in glob(globstring):
-            df = delayed(pd.read_table)(file_path, )
[…]
