+++ docs/api/openomics.database.lncRNome.rst
@@ -1,19 +1,0 @@
-lncRNome
-========
-
-.. currentmodule:: openomics.database
-
-.. autoclass:: lncRNome
-   :show-inheritance:
-
-   .. rubric:: Methods Summary
-
-   .. autosummary::
-
-      ~lncRNome.load_dataframe
-      ~lncRNome.load_network
-
-   .. rubric:: Methods Documentation
-
-   .. automethod:: load_dataframe
-   .. automethod:: load_network
