+++ docs/api/openomics.database.GeneOntology.rst
@@ -1,37 +1,0 @@
-GeneOntology
-============
-
-.. currentmodule:: openomics.database
-
-.. autoclass:: GeneOntology
-   :show-inheritance:
-
-   .. rubric:: Attributes Summary
-
-   .. autosummary::
-
-      ~GeneOntology.COLUMNS_RENAME_DICT
-
-   .. rubric:: Methods Summary
-
-   .. autosummary::
-
-      ~GeneOntology.add_predecessor_terms
-      ~GeneOntology.filter_network
-      ~GeneOntology.get_predecessor_terms
-      ~GeneOntology.info
-      ~GeneOntology.load_dataframe
-      ~GeneOntology.load_network
-
-   .. rubric:: Attributes Documentation
-
-   .. autoattribute:: COLUMNS_RENAME_DICT
-
-   .. rubric:: Methods Documentation
-
-   .. automethod:: add_predecessor_terms
-   .. automethod:: filter_network
-   .. automethod:: get_predecessor_terms
-   .. automethod:: info
-   .. automethod:: load_dataframe
-   .. automethod:: load_network
