+++ docs/api/openomics.database.LncRNA2Target.rst
@@ -1,21 +1,0 @@
-LncRNA2Target
-=============
-
-.. currentmodule:: openomics.database
-
-.. autoclass:: LncRNA2Target
-   :show-inheritance:
-
-   .. rubric:: Methods Summary
-
-   .. autosummary::
-
-      ~LncRNA2Target.load_network
-      ~LncRNA2Target.load_network_high_throughput
-      ~LncRNA2Target.load_network_low_throughput
-
-   .. rubric:: Methods Documentation
-
-   .. automethod:: load_network
-   .. automethod:: load_network_high_throughput
-   .. automethod:: load_network_low_throughput
