+++ docs/api/openomics.database.GTEx.rst
@@ -1,27 +1,0 @@
-GTEx
-====
-
-.. currentmodule:: openomics.database
-
-.. autoclass:: GTEx
-   :show-inheritance:
-
-   .. rubric:: Attributes Summary
-
-   .. autosummary::
-
-      ~GTEx.COLUMNS_RENAME_DICT
-
-   .. rubric:: Methods Summary
-
-   .. autosummary::
-
-      ~GTEx.load_dataframe
-
-   .. rubric:: Attributes Documentation
-
-   .. autoattribute:: COLUMNS_RENAME_DICT
-
-   .. rubric:: Methods Documentation
-
-   .. automethod:: load_dataframe

[…]
