+++ docs/usage/getting-started.md
@@ -4,48 +4,57 @@ Welcome! This tutorial highlights the OpenOmics API’s core features; for in-de
 
 ## Loading a single-omics dataframe
 
-Suppose you have your own -omics dataset(s) and you'd like to load them as dataframes. Depending on whether your data is
-stored locally, retrieved via an online database, or was already a dataframe, you can instantiate a data object from one
-of the class in the `openomics.multiomics` submodule.
+Suppose you have a single-omics dataset and would like to load them as a dataframe.
 
-````{tab} From local file
-If the dataset is a local file in a tabular format, OpenOmics can help you load them to Pandas dataframes. As an example, we downloaded the `TGCA_LUAD` dataset from [tests/data/TCGA_LUAD](https://github.com/BioMeCIS-Lab/OpenOmics/tree/master/tests/data/TCGA_LUAD).
+As an example, we use the `TGCA` Lung Adenocarcinoma dataset from [tests/data/TCGA_LUAD](https://github.com/BioMeCIS-Lab/OpenOmics/tree/master/tests/data/TCGA_LUAD). Data tables are tab-delimited and have the following format:
+
+| GeneSymbol      | EntrezID  | TCGA-05-4244-01A-01R-1107-07 | TCGA-05-4249-01A-01R-1107-07 | ...  |
+| --------------- | --------- | ---------------------------- | ---------------------------- | ---- |
+| ENSG00000223972 | 100133144 | 10.8123                      | 3.7927                       | ...  |
+| ⋮ | ⋮ | ⋮ | ⋮ |
+
+Depending on whether your data table is stored locally as a single file, splitted into multiple files, or was already a dataframe, you can load it using the class {class}`openomics.transcriptomics.Expression` or any of its subclasses.
+
+````{tab} From a single file
+If the dataset is a local file in a tabular format, OpenOmics can help you load them to Pandas dataframe.
 
 ```{code-block} python
-from openomics.multiomics import  MessengerRNA
+from openomics.multiomics import MessengerRNA
 
-# Load each expression dataframe
-mrna = MessengerRNA("TCGA_LUAD/LUAD__geneExp.txt",
-                    transpose=True,
-                    usecols="GeneSymbol|TCGA",
-                    gene_index="GeneSymbol")
+mrna = MessengerRNA(
+    data="https://raw.githubusercontent.com/BioMeCIS-Lab/OpenOmics/master/tests/data/TCGA_LUAD/LUAD__geneExp.txt",
+    transpose=True,
+    usecols="GeneSymbol|TCGA", # A regex that matches all column name with either "GeneSymbol" or "TCGA substring
+    gene_index="GeneSymbol", # This column contains the gene index
+    )
 ```
 
 One thing to pay attention is that the raw data file given is column-oriented where columns corresponds to samples, so we have use the argument `transpose=True` to convert to row-oriented.
-
 > MessengerRNA (576, 20472)
-
 ````
 
 ````{tab} From multiple files (glob)
-If your dataset is large, it may be broken up into multiple files with a similar file name prefix. Assuming all the files have the same tabular format, OpenOmics can load all files and contruct
+If your dataset is large, it may be broken up into multiple files with a similar file name prefix/suffix. Assuming all the files have similar tabular format, OpenOmics can load all files and contruct an integrated data table using the memory-efficient Dask dataframe.
 
 ```python
-from openomics.multiomics import Expression # a superclass of MessengerRNA
+from openomics.multiomics import MessengerRNA
 
-rnaseq = Expression("TCGA_LUAD/LUAD__*",
[…]
