+++ openomics/clinical.py
@@ -1,9 +1,11 @@
+import validators
 import io
 import os
 from typing import List, Union
 
 import dask.dataframe as dd
 import pandas as pd
+from .utils import get_pkg_data_filename
 
 BCR_PATIENT_BARCODE_COL = "bcr_patient_barcode"
 HISTOLOGIC_SUBTYPE_COL = "histologic_subtype"
@@ -59,6 +61,12 @@ class ClinicalData:
         elif isinstance(file_path, (pd.DataFrame, dd.DataFrame)):
             self.patient = file_path
 
+        elif isinstance(file_path, str) and validators.url(file_path):
+            dataurl, filename = os.path.split(file_path)
+            file = get_pkg_data_filename(dataurl + "/", filename)
+            self.patient = pd.read_table(file)
+
+
         else:
             raise IOError(file_path)
 
@@ -89,13 +97,15 @@ class ClinicalData:
         self.samples.index.name = index
         self.samples.index = self.samples.index.str[:-4]  # Cut sample barcode for TCGA
 
-        no_samples = self.samples.shape[0]
+        num_samples = self.samples.shape[0]
 
         # Merge patients clinical data with patient barcode as index
         # target = pd.merge(target, self.patient,
         #                      how="left", left_on="patient_barcode", right_on="patient_barcode")
+
+
         self.samples = self.samples.join(self.patient, on=index, how="left", rsuffix="_")
-        if self.samples.shape[0] != no_samples:
+        if self.samples.shape[0] != num_samples:
             raise Exception("Clinical data merging has wrong number of samples")
 
         # self.samples.dropna(axis=0, subset=["bcr_patient_barcode"], inplace=True) # Remove samples without clinical data
