+++ docs/usage/getting-started.md
@@ -1,6 +1,6 @@
 # Getting started
 
-Welcome! This tutorial highlights the OpenOmics API’s core features; for in-depth details and conceptual design, see the links within, or the documentation index which has links to use cases, and API reference sections.
+Welcome! This tutorial highlights the OpenOmics API’s core features; for in-depth details and conceptual guides, see the links within, or the documentation index which has links to use cases, and API reference sections.
 
 ## Loading a single-omics dataframe
 
@@ -8,9 +8,9 @@ Suppose you have a single-omics dataset and would like to load them as a datafra
 
 As an example, we use the `TGCA` Lung Adenocarcinoma dataset from [tests/data/TCGA_LUAD](https://github.com/BioMeCIS-Lab/OpenOmics/tree/master/tests/data/TCGA_LUAD). Data tables are tab-delimited and have the following format:
 
-| GeneSymbol      | EntrezID  | TCGA-05-4244-01A-01R-1107-07 | TCGA-05-4249-01A-01R-1107-07 | ...  |
-| --------------- | --------- | ---------------------------- | ---------------------------- | ---- |
-| ENSG00000223972 | 100133144 | 10.8123                      | 3.7927                       | ...  |
+| GeneSymbol | EntrezID  | TCGA-05-4244-01A-01R-1107-07 | TCGA-05-4249-01A-01R-1107-07 | ...  |
+| ---------- | --------- | ---------------------------- | ---------------------------- | ---- |
+| A1BG       | 100133144 | 10.8123                      | 3.7927                       | ...  |
 | ⋮ | ⋮ | ⋮ | ⋮ |
 
 Depending on whether your data table is stored locally as a single file, splitted into multiple files, or was already a dataframe, you can load it using the class {class}`openomics.transcriptomics.Expression` or any of its subclasses.
@@ -39,7 +39,7 @@ If your dataset is large, it may be broken up into multiple files with a similar
 ```python
 from openomics.multiomics import MessengerRNA
 
-mrna = MessengerRNA("TCGA_LUAD/LUAD__*",
+mrna = MessengerRNA("TCGA_LUAD/LUAD__*", # Files must be stored locally
                     transpose=True,
                     usecols="GeneSymbol|TCGA",
                     gene_index="GeneSymbol")
@@ -68,82 +68,336 @@ df = pd.DataFrame(data={"ENSG00000194717": np.random.rand(5),
 mrna = MessengerRNA(df, transpose=False, sample_level="sample_id")
 ```
 ````
+---
+To access the {class}`DataFrame`, simply use {obj}`mrna.expressions`:
+```python
+print(mrna.expressions)
+```
+<div>
+<style scoped>
+    .dataframe tbody tr th:only-of-type {
+        vertical-align: middle;
+    }
 
-To access the
+    .dataframe tbody tr th {
+        vertical-align: top;
+    }
+
+    .dataframe thead th {
+        text-align: right;
+    }
+
+</style>
+<table border="1" class="dataframe">
+  <thead>
+    <tr style="text-align: right;">
+      <th>GeneSymbol</th>
+      <th>A1BG</th>
+      <th>A1BG-AS1</th>
+      <th>A1CF</th>
+      <th>A2M</th>
+    </tr>
+    <tr>
+      <th>sample_index</th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <th>TCGA-05-4244-01A-01R-1107-07</th>
+      <td>26.0302</td>
+      <td>36.7711</td>
[…]
