+++ .github/workflows/python-package.yml
@@ -12,7 +12,7 @@ jobs:
         runs-on: ${{ matrix.os }}
         strategy:
             matrix:
-                os: [ ubuntu-latest, macos-latest, windows-latest ]
+                os: [ ubuntu-latest, macos-latest ]
                 python-version: [ 3.6, 3.7, 3.8, 3.9 ]
                 include:
                     -   os: ubuntu-latest
