### 2021-03-12 fb0a948
Format code with black and yapf
This commit fixes the style issues introduced in 2f8a5ad according to the output
from black and yapf.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/79606691-5c43-4e84-9476-85f984c4d9f6/
