+++ openomics_web/layouts/clinical_view.py
@@ -10,25 +10,31 @@ def ClinicalDataColumnSelect(columns):
         columns:
     """
     return html.Div([
-        html.Div(['Select the gene id/name column to index by:']),
+        html.Div(["Select the gene id/name column to index by:"]),
         dcc.Dropdown(
-            id='clinical-patient-col-name',
-            options=[{'label': col, 'value': col} for col in columns],
+            id="clinical-patient-col-name",
+            options=[{
+                "label": col,
+                "value": col
+            } for col in columns],
             style={
-                'width': '100%',
+                "width": "100%",
             },
             value=columns[0],
         ),
-        html.Div(['Select the column prefixes to import:']),
+        html.Div(["Select the column prefixes to import:"]),
         dcc.Dropdown(
-            id='clinical-data-columns-select',
-            options=[{'label': col, 'value': col} for col in columns],
+            id="clinical-data-columns-select",
+            options=[{
+                "label": col,
+                "value": col
+            } for col in columns],
             style={
-                'width': '100%',
+                "width": "100%",
             },
             value=columns,
             multi=True,
-        )
+        ),
     ])
 
 
@@ -41,33 +47,43 @@ def ClinicalDataTable(df: pd.DataFrame):
     print("df.reset_index()", df.reset_index().columns)
 
     return dt.DataTable(
-        id='clinical-datatable',
-        columns=[{'name': i, 'id': i, 'deletable': True} for i in df.columns],
-        data=df.reset_index().to_dict('records'),
+        id="clinical-datatable",
+        columns=[{
+            "name": i,
+            "id": i,
+            "deletable": True
+        } for i in df.columns],
+        data=df.reset_index().to_dict("records"),
         style_as_list_view=True,
         # fixed_columns={'headers': True, 'data': 1},
         style_cell={
-            'textAlign': 'left',
-            'minWidth': '180px', 'width': '180px', 'maxWidth': '180px',
+            "textAlign": "left",
+            "minWidth": "180px",
+            "width": "180px",
+            "maxWidth": "180px",
         },
         style_data={
-            'whiteSpace': 'normal',
-            'height': 'auto'
+            "whiteSpace": "normal",
+            "height": "auto"
         },
         style_data_conditional=[
-            {'if': {'row_index': 'odd'},
-             'backgroundColor': 'rgb(248, 248, 248)'
-             },
+            {
+                "if": {
+                    "row_index": "odd"
+                },
+                "backgroundColor": "rgb(248, 248, 248)"
+            },
         ],
-        style_table={"maxHeight": '1200px',
-                     'width': '1000px',
-                     'marginTop': '5px',
-                     'marginBottom': '10px',
-                     'overflowX': 'scroll'
-                     },
+        style_table={
+            "maxHeight": "1200px",
+            "width": "1000px",
+            "marginTop": "5px",
+            "marginBottom": "10px",
+            "overflowX": "scroll",
+        },
         style_header={
-            'backgroundColor': 'white',
-            'fontWeight': 'bold'
+            "backgroundColor": "white",
+            "fontWeight": "bold"
         },
         virtualization=True,
         filter_action="native",
