+++ docs/conf.py
@@ -21,7 +21,7 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath('..'))
+sys.path.insert(0, os.path.abspath(".."))
 
 import openomics
 
@@ -32,24 +32,22 @@ import openomics
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
 extensions = [
-    'sphinx.ext.autodoc',
+    "sphinx.ext.autodoc",
     # 'sphinx.ext.autosummary',
-    'sphinx_autodoc_typehints',
-    'sphinx.ext.linkcode',
-    'sphinx.ext.napoleon',
-    'myst_parser',
-
+    "sphinx_autodoc_typehints",
+    "sphinx.ext.linkcode",
+    "sphinx.ext.napoleon",
+    "myst_parser",
     # User Usability
-    'sphinx_copybutton',
-    'sphinx_inline_tabs',
-
-    'sphinx.ext.coverage',
-    'sphinx.ext.intersphinx',
+    "sphinx_copybutton",
+    "sphinx_inline_tabs",
+    "sphinx.ext.coverage",
+    "sphinx.ext.intersphinx",
     "sphinx.ext.viewcode",
-    'sphinx_automodapi.automodapi',
-    'sphinx_automodapi.smart_resolver',
-    'sphinx.ext.graphviz',
-    'sphinx.ext.inheritance_diagram',
+    "sphinx_automodapi.automodapi",
+    "sphinx_automodapi.smart_resolver",
+    "sphinx.ext.graphviz",
+    "sphinx.ext.inheritance_diagram",
 ]
 
 autosummary_generate = False
@@ -59,23 +57,23 @@ napoleon_use_param = True
 napoleon_use_ivar = True
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
-exclude_patterns = ['_build', '_templates']
+templates_path = ["_templates"]
+exclude_patterns = ["_build", "_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
 source_suffix = {
-    '.rst': 'restructuredtext',
-    '.txt': 'markdown',
-    '.md': 'markdown',
+    ".rst": "restructuredtext",
+    ".txt": "markdown",
+    ".md": "markdown",
 }
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # General information about the project.
-project = 'OpenOmics'
+project = "OpenOmics"
 copyright = "2019, Nhat Tran"
 author = "Jonny Tran"
 
@@ -98,10 +96,10 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
+pygments_style = "sphinx"
 
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 # todo_include_todos = True
@@ -116,7 +114,7 @@ myst_heading_anchors = 3
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'furo'
+html_theme = "furo"
 
 # Theme options are theme-specific and customize the look and feel of a
 # theme further.  For a list of options available for each theme, see the
[…]
