+++ openomics/multicohorts.py
@@ -1,2 +1,1 @@
-
 # TODO Multicohort data integration features Coming soon!
