### 2021-03-12 #31 @JonnyTran
Hi guys @NickleDave  @ksielemann @gawbul.  It's been awhile, but I finally managed to address *most* of the comments. Yay!

---
For @ksielemann comments:

> Download of test data

I implemented a way to load the Expression data files directly from URL, see <https://openomics.readthedocs.io/en/latest/usage/getting-started.html#creating-a-multi-omics-dataset>

> choose the directory in which the files should be downloaded?

There is now a function to do so, see <https://openomics.readthedocs.io/en/latest/usage/annotate-external-databases.html#setting-the-cache-download-directory>

---
For @gawbul comments:

> Some tests are available and run as part of the Travis CI pipeline, though coverage isn't amazing and would benefit from additional work.

I've set up GitHub Actions at <https://github.com/BioMeCIS-Lab/OpenOmics/actions/workflows/python-package.yml> to replace Travis CI for better pricing plans. The automated test suite currently targets Mac OS and Linux for Python 3.6-3.9. A few tests are failing (due to unavailability of some FTP servers), but I believe you should have the same problems on your Mac OS + Python 3.9 anymore.

> I found the documentation on Read the Docs to be lacking

I did a complete revamp of ReadTheDocs documentation site, especially vignettes and usage guide at <https://openomics.readthedocs.io/en/latest/usage/getting-started.html>. The structure of API references is in place, although more in-depth usage guides should be written.

The readme.md file is also edited to reflect guidelines from <https://www.writethedocs.org/guide/writing/beginners-guide-to-docs/#readme>

---
For the JOSS submission, my manuscript can be compiled from <https://github.com/BioMeCIS-Lab/OpenOmics/tree/master/inst>

Is there anything else I might be missing?
