+++ openomics/utils/io.py
@@ -15,7 +15,7 @@ from requests.adapters import HTTPAdapter, Retry
 import openomics
 
 
-@astropy.config.set_temp_cache(openomics.config["cache_dir"])
+# @astropy.config.set_temp_cache(openomics.config["cache_dir"])
 def get_pkg_data_filename(dataurl, file):
     """Downloads a remote file given the url, then caches it to the user's home
     folder.
