+++ docs/misc/faq.md
@@ -1,1 +1,9 @@
 # Frequently Asked Questions (FAQ)
+
+### How to change the download directory where database files are cached?
+
+You can make the setting at `openomics.set_cache_dir(path="PATH/OF/YOUR/CHOICE/")` which will change the default
+cache location. However, this setting doesn't persist across user sessions. You can also change the default cache
+directory at a `~/.openomics/conf.json` to persist the setting for each user.
+
+
