### 2021-03-15 #31 @ksielemann
First of all, thank you @JonnyTran for addressing the comments above! Some of my review comments are embedded in the specific points of the Package Review form. I believe that these points were not yet addressed (or did I miss something?).
