### 2021-03-17 #31 @NickleDave
Thank you for your quick reply @ksielemann
Yes, I see that some of your comments were embedded in the form.

I don't mean to make more work for you, but could I ask you to raise separate issues on the OpenOmics repository for any comments that you feel have not yet been addressed?

This is the usual approach that JOSS reviews use (to raise issues with details on the repo, and then link to them / summarize on the "review issue"), I think to avoid situations like this. We should probably have very clear instructions suggesting the same approach in our guidelines--our fault.

@gawbul I think that you did raise issues for some of your comments. Can you please also check whether there are any that remain to be addressed?
