### 2021-03-17 #31 @NickleDave
No need to apologize @ksielemann , definitely my fault for not being clearer about process.
Thank you for taking time to open issues. That will help.

Sorry @JonnyTran for adding more to your plate.
I just want to make sure it's very clear what review criteria have been met, according to reviewers.

If we need to, we can discuss further here, and you can link to specific issues.

I will check back by Friday at the latest.
Again, as far as JOSS goes, if the manuscript compiles and you have a DOI for the version that we approve, then I think you are good to go.
I will make sure of that when we reach that point.
