+++ docs/modules/openomics.annotate.md
@@ -4,5 +4,5 @@
 .. automodapi:: openomics.database.base
     :no-heading:
     :no-inheritance-diagram:
-    :skip: ABC
+    :skip: ABC, Dict, List, Tuple
 ```
