+++ tests/test_clinical.py
@@ -1,5 +1,4 @@
 import os
-
 import pytest
 
 from openomics import ClinicalData
@@ -9,6 +8,7 @@ cohort_folder_path = "tests/data/TCGA_LUAD"
 
 @pytest.fixture
 def generate_TCGA_clinical():
+    print(os.path.join(cohort_folder_path, "nationwidechildrens.org_clinical_patient_luad.txt"))
     clinical = ClinicalData(
         file_path=os.path.join(cohort_folder_path, "nationwidechildrens.org_clinical_patient_luad.txt"),
         patient_index="bcr_patient_barcode")
