+++ openomics/genomics.py
@@ -19,10 +19,9 @@ class SingleNucleotideVariants(Expression, Annotatable):
             cohort_name:
         """
         super(SingleNucleotideVariants, self).__init__(data=data, transpose=transpose, gene_index=gene_index,
-                                                       usecols=usecols,
-                                                       gene_level=gene_level, sample_level=sample_level,
-                                                       transform_fn=transform_fn, dropna=dropna,
-                                                       npartitions=npartitions, cohort_name=cohort_name)
+                                                       usecols=usecols, gene_level=gene_level,
+                                                       sample_level=sample_level, transform_fn=transform_fn,
+                                                       dropna=dropna, npartitions=npartitions, cohort_name=cohort_name)
 
     @classmethod
     def name(cls):
@@ -47,8 +46,8 @@ class SomaticMutation(Expression, Annotatable):
         """
         super(SomaticMutation, self).__init__(data=data, transpose=transpose, gene_index=gene_index, usecols=usecols,
                                               gene_level=gene_level, sample_level=sample_level,
-                                              transform_fn=transform_fn, dropna=dropna,
-                                              npartitions=npartitions, cohort_name=cohort_name)
+                                              transform_fn=transform_fn, dropna=dropna, npartitions=npartitions,
+                                              cohort_name=cohort_name)
 
     @classmethod
     def name(cls):
@@ -73,8 +72,8 @@ class DNAMethylation(Expression, Annotatable):
         """
         super(DNAMethylation, self).__init__(data=data, transpose=transpose, gene_index=gene_index, usecols=usecols,
                                              gene_level=gene_level, sample_level=sample_level,
-                                             transform_fn=transform_fn, dropna=dropna,
-                                             npartitions=npartitions, cohort_name=cohort_name)
+                                             transform_fn=transform_fn, dropna=dropna, npartitions=npartitions,
+                                             cohort_name=cohort_name)
 
     @classmethod
     def name(cls):
@@ -98,10 +97,9 @@ class CopyNumberVariation(Expression, Annotatable):
             cohort_name:
         """
         super(CopyNumberVariation, self).__init__(data=data, transpose=transpose, gene_index=gene_index,
-                                                  usecols=usecols,
-                                                  gene_level=gene_level, sample_level=sample_level,
-                                                  transform_fn=transform_fn, dropna=dropna,
-                                                  npartitions=npartitions, cohort_name=cohort_name)
+                                                  usecols=usecols, gene_level=gene_level, sample_level=sample_level,
+                                                  transform_fn=transform_fn, dropna=dropna, npartitions=npartitions,
+                                                  cohort_name=cohort_name)
 
     @classmethod
     def name(cls):
