+++ tests/test_clinical.py
@@ -3,13 +3,12 @@ import pytest
 
 from openomics import ClinicalData
 
-cohort_folder_path = "tests/data/TCGA_LUAD"
+cohort_folder_path = "tests/data/TCGA_LUAD/"
 
 
 @pytest.fixture
 def generate_TCGA_clinical():
-    print(os.path.join(cohort_folder_path, "nationwidechildrens.org_clinical_patient_luad.txt"))
     clinical = ClinicalData(
-        file_path=os.path.join(cohort_folder_path, "nationwidechildrens.org_clinical_patient_luad.txt"),
+        file=os.path.join(cohort_folder_path, "nationwidechildrens.org_clinical_patient_luad.txt"),
         patient_index="bcr_patient_barcode")
     return clinical
