+++ docs/modules/openomics.multiomics.md
@@ -1,6 +1,7 @@
 # MultiOmics data
 
 ```{eval-rst}
-.. automodapi::
-openomics
+.. automodapi:: openomics
+    :classes-only:
+    :skip: set_backend, set_cache_dir
 ```
