+++ openomics-0.8.6/openomics/visualization/__init__.py
@@ -1,0 +1,1 @@
+from .heatmat import heatmap
