+++ openomics-0.8.6/openomics/multiomics.py
@@ -1,0 +1,304 @@
+import logging
+from typing import List, Dict, Union
+
+import pandas as pd
+
+import openomics
+from .clinical import (
+    ClinicalData,
+    HISTOLOGIC_SUBTYPE_COL,
+    PATHOLOGIC_STAGE_COL,
+    TUMOR_NORMAL_COL,
+    PREDICTED_SUBTYPE_COL,
+)
+from .genomics import SomaticMutation, CopyNumberVariation, DNAMethylation
+from .imageomics import WholeSlideImage
+from .proteomics import Protein
+from .transcriptomics import MessengerRNA, MicroRNA, LncRNA, Expression
+
+
+class MultiOmics:
+    """A data object which holds multiple -omics data for a single clinical cohort."""
+    def __init__(self, cohort_name, omics_data=None):
+        """
+
+        Args:
+            cohort_name (str): the clinical cohort name
+        """
+        self._cohort_name = cohort_name
+        self._omics = []
+
+        # This is a data dictionary accessor to retrieve individual -omic data
+        self.data = {}
+
+        if omics_data:
+            for omics in omics_data:
+                self.add_omic(omics)
+
+    def add_omic(self,
+                 omic_data: Expression,
+                 initialize_annotations: bool = True):
+        """Adds an omic object to the Multiomics such that the samples in omic
+        matches the samples existing in the other omics.
+
+        Args:
+            omic_data (Expression): The omic to add, e.g., MessengerRNA,
+                MicroRNA, LncRNA, etc.
+            initialize_annotations (bool): default True. If true, initializes
+                the annotation dataframe in the omic object
+        """
+        self.__dict__[omic_data.name()] = omic_data
+
+        if omic_data.name not in self._omics:
+            self._omics.append(omic_data.name())
+
+        # dictionary as data accessor to the expression data
+        self.data[omic_data.name()] = omic_data.expressions
+
+        # Initialize annotation
+        if initialize_annotations:
+            omic_data.initialize_annotations(index=omic_data.gene_index,
+                                             gene_list=None)
+
+        logging.info(
+            omic_data.name(),
+            self.data[omic_data.name()].shape if hasattr(
+                self.data[omic_data.name()], "shape") else ": None",
+            ", indexed by:",
+            omic_data.annotations.index.name,
+        )
+
+    def add_clinical_data(self, clinical: openomics.clinical.ClinicalData, **kwargs):
+        """ Add a ClinicalData instance to the MultiOmics instance.
+
+        Args:
+            clinical (openomics.clinical.ClinicalData):
+        """
+        if not isinstance(clinical, ClinicalData):
+            raise Exception("Must pass a ClinicalData in, not a file path.")
+
+        self.clinical = clinical
+
+        self.data["PATIENTS"] = self.clinical.patient
+        if hasattr(self.clinical, "biospecimen"):
+            self.data["BIOSPECIMENS"] = self.clinical.biospecimen
+        if hasattr(self.clinical, "drugs"):
+            self.data["DRUGS"] = self.clinical.drugs
+
+        self.build_samples(**kwargs)
+
+    def get_omics_list(self):
+        return self._omics
+
+    def __getitem__(self, item:str):
+        """This function allows the MultiOmicData class objects to access
+        individual omics by a dictionary lookup, e.g. openomics["MicroRNA"]
+
+        Args:
+            item (str): a string of the class name
+        """
+        if item.lower() == MessengerRNA.name().lower():
+            return self.__getattribute__(MessengerRNA.name())
+
+        elif item.lower() == MicroRNA.name().lower():
+            return self.__getattribute__(MicroRNA.name())
+
+        elif item.lower() == LncRNA.name().lower():
[…]
