### 2021-03-18 dcf0db3
Bump version: 0.8.6 → 0.8.7
