### 2021-03-18 1d50e03
Bump version: 0.8.6 → 0.8.7
