+++ openomics/__init__.py
@@ -13,7 +13,7 @@ import pandas as pd
 
 __author__ = """Nhat (Jonny) Tran"""
 __email__ = 'nhat.tran@mavs.uta.edu'
-__version__ = '0.8.7'
+__version__ = '0.8.8'
 
 
 # Initialize configurations
