+++ openomics-0.8.6/openomics/utils/__init__.py
@@ -1,2 +1,0 @@
-from .io import get_pkg_data_filename
-from .read_gtf import read_gtf
