+++ openomics-0.8.6/openomics/multicohorts.py
@@ -1,2 +1,0 @@
-
-# TODO Multicohort data integration features Coming soon!

[…]
