+++ openomics/utils/io.py
@@ -34,7 +34,7 @@ def get_pkg_data_filename(dataurl, file):
         dataurl = dataurl + "/"
 
     try:
-        logging.info("Fetching file from: {}{}, saving to {}".format(dataurl, file, openomics.config['cache_dir']))
+        logging.debug("Fetching file from: {}{}, saving to {}".format(dataurl, file, openomics.config['cache_dir']))
 
         with data.conf.set_temp("dataurl", dataurl), data.conf.set_temp("remote_timeout", 30):
             return data.get_pkg_data_filename(file, package="openomics.database", show_progress=True)
