+++ docs/usage/preprocess-downstream-analysis.md
@@ -15,7 +15,7 @@ print(X_multiomics['MessengerRNA'].shape,
 
 
 ```python
-y
+print(y)
 ```
 
 
