+++ README.md
@@ -8,7 +8,7 @@
 [![OpenOmics](https://github.com/BioMeCIS-Lab/OpenOmics/actions/workflows/python-package.yml/badge.svg?branch=master)](https://github.com/BioMeCIS-Lab/OpenOmics/actions/workflows/python-package.yml)
 [![codecov](https://codecov.io/gh/BioMeCIS-Lab/OpenOmics/branch/master/graph/badge.svg?token=WAN3PJwM42)](https://codecov.io/gh/BioMeCIS-Lab/OpenOmics)
 
-This Python package provide a series of tools to integrate and query the genomics, transcriptomics, proteomics, and
+This Python package provide a series of tools to integrate and explore the genomics, transcriptomics, proteomics, and
 clinical data (aka multi-omics data). With scalable data-frame manipulation tools, OpenOmics facilitates the common
 data wrangling tasks when preparing data for RNA-seq bioinformatics analysis.
 
