### 2021-03-19 #31 @ksielemann
I just closed my last issue with this comment:
_'I think it is really important for future users that the usage guide works without errors. Otherwise, the user might get frustrated and refrains from using the library. So the usage guide should be updated according to the functionalities and current version of the package. But I think this can also happen while the package is further developed.'_

So from my side, all my comments are sufficiently addressed now!
