### 2021-03-19 #31 @NickleDave
Great thank you @ksielemann glad to hear it.
Looks like the additional issues were all easily addressed or fixed already.
Sorry, I didn't mean to ask you to do extra work, just wanted to make sure we were all on the same page about requested revisions. Thank you again 🙏
