### 2021-03-19 #31 @NickleDave
@gawbul just want to check back -- can you please confirm whether your comments have been addressed?

Looks like the corresponding issues were:
https://github.com/BioMeCIS-Lab/OpenOmics/issues/119
https://github.com/BioMeCIS-Lab/OpenOmics/issues/115
https://github.com/BioMeCIS-Lab/OpenOmics/issues/114
https://github.com/BioMeCIS-Lab/OpenOmics/issues/113
