### 2021-04-22 BioMeCIS-Lab/OpenOmics#115 @gawbul
Tests all pass, though a few warnings:

```
$ OpenOmics [master] python -m pytest
============================================================================================================ test session starts =============================================================================================================
platform darwin -- Python 3.8.9, pytest-6.2.3, py-1.10.0, pluggy-0.13.1
rootdir: /Users/stephenmoss/Dropbox/Code/OpenOmics, configfile: setup.cfg
plugins: dash-1.20.0
collected 38 items

tests/test_annotations.py .........                                                                                                                                                                                                    [ 23%]
tests/test_disease.py ..........                                                                                                                                                                                                       [ 50%]
tests/test_interaction.py .......                                                                                                                                                                                                      [ 68%]
tests/test_multiomics.py ...                                                                                                                                                                                                           [ 76%]
tests/test_sequences.py .........                                                                                                                                                                                                      [100%]

============================================================================================================== warnings summary ==============================================================================================================
../../../.pyenv/versions/3.8.9/lib/python3.8/site-packages/_pytest/config/__init__.py:1233
  /Users/stephenmoss/.pyenv/versions/3.8.9/lib/python3.8/site-packages/_pytest/config/__init__.py:1233: PytestConfigWarning: Unknown config option: collect_ignore

    self._warn_or_fail_if_strict(f"Unknown config option: {key}\n")

openomics/database/ontology.py:235
  /Users/stephenmoss/Dropbox/Code/OpenOmics/openomics/database/ontology.py:235: DeprecationWarning: invalid escape sequence \|
    and annotation.str.contains("\||;", regex=True).any()):

tests/test_disease.py::test_import_HMDD
tests/test_disease.py::test_annotate_HMDD
  /Users/stephenmoss/.pyenv/versions/3.8.9/lib/python3.8/encodings/unicode_escape.py:26: DeprecationWarning: invalid escape sequence '\ '
    return codecs.unicode_escape_decode(input, self.errors)[0]

tests/test_disease.py::test_import_HMDD
tests/test_disease.py::test_annotate_HMDD
  /Users/stephenmoss/.pyenv/versions/3.8.9/lib/python3.8/encodings/unicode_escape.py:26: DeprecationWarning: invalid escape sequence '\s'
    return codecs.unicode_escape_decode(input, self.errors)[0]

tests/test_interaction.py::test_import_LncRNA2Target
tests/test_interaction.py::test_get_interactions_lnc2target
  /Users/stephenmoss/Dropbox/Code/OpenOmics/openomics/database/interaction.py:479: FutureWarning: Your version of xlrd is 1.2.0. In xlrd >= 2.0, only the xls format is supported. As a result, the openpyxl engine will be used if it is installed and the engine argument is not specified. Install openpyxl instead.
    table = pd.read_excel(file_resources["lncRNA_target_from_low_throughput_experiments.xlsx"])

-- Docs: https://docs.pytest.org/en/stable/warnings.html
================================================================================================= 38 passed, 8 warnings in 603.78s (0:10:03) =================================================================================================
```
