### 2021-04-22 BioMeCIS-Lab/OpenOmics#119 @gawbul
I did notice that this section of the docs https://openomics.readthedocs.io/en/latest/usage/getting-started.html#creating-a-multi-omics-dataset was missing the following:

```
from openomics import MultiOmics
```

Otherwise, the following statement fails:

```
luad_data = MultiOmics(cohort_name="LUAD", omics_data=[mRNA, mRNA, lncRNA, som, pro])
```
