### 2021-04-22 #31 @gawbul
All looks good 👍

I made a quick comment here regarding an issue with the docs https://github.com/BioMeCIS-Lab/OpenOmics/issues/119#issuecomment-825137257, but otherwise, I'm happy everything has been addressed ☺️
