### 2021-04-22 BioMeCIS-Lab/OpenOmics#119 @JonnyTran
Thanks for pointing this out. I will fix that.
