### 2021-04-27 #31 @NickleDave
Excellent, thank you so much @gawbul and @ksielemann for your very thorough reviews

----------------------------------------------
🎉 `openomics` has been approved by pyOpenSci! Thank you @JonnyTran for submitting

- [ ] Add the badge for pyOpenSci peer-review to the README.md of `openomics`. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/31)`
- [ ] Add `openomics` to the pyOpenSci website. @JonnyTran , please open a PR to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/packages.yml): to add your package and name to the list of contributors
- [ ] @JonnyTran @gawbul @ksielemann if you have time and are open to being listed on our website, please add yourselves to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/contributors.yml) via a PR so we can list you on our website as contributors!

Since this package is going to move on to JOSS for review, you'll also want to do the following:
- [ ] Tag and create a new release through GitHub, to produce a new version and DOI listed on Zenodo
- [ ] Submit to JOSS using the Zenodo DOI. We will tag it for expedited review.
@JonnyTran feel free to tag me (@NickleDave) and reference this issue with a link when you submit to JOSS just so I can stay up to date with the process there.

All -- if you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and our documentation in the [contributing-guide](https://www.pyopensci.org/contributing-guide). We have also been updating our documentation to improve the process so all feedback is appreciated!
