### 2021-04-30 32c9852
updated documentations for Database classes
