### 2021-04-30 09d0518
updated documentations for Database classes
