+++ openomics/database/interaction.py
@@ -140,6 +140,15 @@ class Interactions(Database):
 
 
 class GeneMania(Interactions):
+    """Loads the GeneMania database from  .
+
+    Default path: local_directory .
+    Default file_resources: {
+        "COMBINED.DEFAULT_NETWORKS.BP_COMBINING.txt": "COMBINED.DEFAULT_NETWORKS.BP_COMBINING.txt",
+        "identifier_mappings.txt": "identifier_mappings.txt",
+    }
+    """
+
     def __init__(self, path, file_resources=None, source_col_name="Gene_A", target_col_name="Gene_B",
                  source_index="gene_name", target_index="gene_name",
                  edge_attr=None, filters=None, directed=True, relabel_nodes=None):
@@ -175,6 +184,14 @@ class GeneMania(Interactions):
 
 
 class BioGRID(Interactions):
+    """Loads the BioGRID database from https://thebiogrid.org .
+
+    Default path: "https://downloads.thebiogrid.org/Download/BioGRID/Latest-Release/" .
+    Default file_resources: {
+        "BIOGRID-ALL-LATEST.tab2.zip": "BIOGRID-ALL-LATEST.tab2.zip",
+    }
+    """
+
     def __init__(self, path="https://downloads.thebiogrid.org/Download/BioGRID/Latest-Release/",
                  file_resources=None, source_col_name="Official Symbol Interactor A",
                  target_col_name="Official Symbol Interactor B",
@@ -208,6 +225,16 @@ class BioGRID(Interactions):
 
 
 class STRING(Interactions, SequenceDatabase):
+    """Loads the STRING database from https://string-db.org/ .
+
+    Default path: "https://stringdb-static.org/download/" .
+    Default file_resources: {
+        "protein.actions.txt": f"protein.actions.v11.0/{species_id}.protein.actions.v11.0.txt.gz",
+        "protein.links.txt": f"protein.links.v11.0/{species_id}.protein.links.v11.0.txt.gz",
+        "protein.info.txt": f"protein.info.v11.0/{species_id}.protein.info.v11.0.txt.gz",
+        "protein.sequences.fa": f"protein.sequences.v11.0/{species_id}.protein.sequences.v11.0.fa.gz"
+    }
+    """
     COLUMNS_RENAME_DICT = {
         "protein_external_id": "protein_id",
         "preferred_name": "protein_name",
@@ -288,6 +315,14 @@ class STRING(Interactions, SequenceDatabase):
 
 
 class LncBase(Interactions, Database):
+    """Loads the LncBase database from http://carolina.imis.athena-innovation.gr/diana_tools/web/index.php?r=lncbasev2%2Findex .
+
+    Default path: local_directory .
+    Default file_resources: {
+        "LncBasev2_download.csv": "LncBasev2_download.csv"",
+    }
+    """
+
     def __init__(self, path, file_resources=None, strip_mirna_name=False,
                  source_col_name="mirna", target_col_name="geneId",
                  source_index="transcript_name", target_index="gene_id",
@@ -335,7 +370,15 @@ class LncBase(Interactions, Database):
 
 
 class LncReg(Interactions):
+    """Loads the  database from  .
 
+    Default path:  .
+    Default file_resources: {
+        "": "",
+        "": "",
+        "": "",
+    }
+    """
     def __init__(self, path, file_resources,
                  source_col_name='A_name_in_paper', target_col_name='B_name_in_paper',
                  source_index="transcript_name", target_index="gene_name",
@@ -370,7 +413,15 @@ class LncReg(Interactions):
 
 
 class lncRInter(Interactions):
+    """Loads the  database from  .
 
+    Default path:  .
+    Default file_resources: {
+        "": "",
+        "": "",
[…]
