### 2021-04-30 67feaec
updated documentations for Database classes
