### 2021-05-04 #31 @NickleDave
Thank you for taking care of those final to-dos @JonnyTran

Just checking, are you about to submit to JOSS? Based on your last couple of commits I'd guess yes.
 Please do let us know and/or reference this issue on the JOSS review when you do.

@lwasser I have edited the first comment to reflect those changes -- not sure if there's more I need to resolve what I was assigned through GitHub. I will close this once the JOSS review is initiated

@ksielemann @gawbul would you be okay with me adding you as contributors to the pyOpenSci site?
I can request your review on the PR when I do so
