### 2021-05-04 #31 @JonnyTran
Hi @NickleDave,

Yes I've submitted to JOSS about 4 days ago, mentioning this was approved by pyOpenSci. I will reference this issue once the review process starts on JOSS's github.

Thanks for the updates!
