### 2021-05-04 #31 @NickleDave
Ah great -- I didn't realize how the process worked, I looked for an issue on their repo but didn't see it.

Thank you for letting me know! Please just let me know if you need anything from us for review at JOSS.

Going to go ahead and close.
Yay!!! congrats on completing review and officially becoming part of pyOpenSci!
Will tweet about `openomics` later!
Thank you again @ksielemann and @gawbul for your great reviews
