### 2021-05-04 #31 @lwasser
oh all - So the JOSS review process is simple in that they accept our review  as theirs! @arfon please note that this package was submitted to the JOSS review process. Can we please fast track it given we have reviewed here on the pyopensci side of things. Please let us know what you need. @JonnyTran is there an open issue in JOSS right now? can you kindly reference it here if you haven't already.

I normally keep these reviews  open here until the JOSS part is finished. When it is they will ask you to add the JOSS badge on your readme as well. thank you all for this!
