### 2021-05-04 #31 @JonnyTran
Hi @lwasser , there isn't a JOSS issue yet, but I will tag this pyOpenSci issue once it is opened.
