### 2021-05-05 #31 @ksielemann
>
> @ksielemann @gawbul would you be okay with me adding you as contributors to the pyOpenSci site?
> I can request your review on the PR when I do so

Sure, I am okay with this! Should I add myself to the file you mentioned above or do you prefer to add me to the contributor site yourself?
