### 2021-05-05 #31 @NickleDave
> Sure, I am okay with this! Should I add myself to the file you mentioned above or do you prefer to add me to the contributor site yourself?
@ksielemann it would be excellent if you could add yourself with a PR

> Sorry for the delay folks. Things are now moving in [openjournals/joss-reviews#3249](https://github.com/openjournals/joss-reviews/issues/3249)
Great thank you @arfon !!!
