### 2021-05-05 c4b02eb
added missing DOIs
