+++ inst/paper.bib
@@ -2,7 +2,8 @@
     title = {The Galaxy platform for accessible, reproducible and collaborative biomedical analyses: 2020 update},
     author = {Jalili, Vahid and Afgan, Enis and Gu, Qiang and Clements, Dave and Blankenberg, Daniel and Goecks, Jeremy and Taylor, James and Nekrutenko, Anton},
     journal = {Nucleic Acids Research},
-    year = {2020}
+    year = {2020},
+    doi = {10.1093/nar/gkaa554},
 }
 
 @techreport{hagberg2008exploring,
@@ -24,13 +25,14 @@
 }
 
 @article{hassan2020integration,
-  title={Integration of Transcriptome and Metabolome Provides Unique Insights to Pathways Associated With Obese Breast Cancer Patients},
-  author={Hassan, Mohammed A and Al-Sakkaf, Kaltoom and Shait Mohammed, Mohammed Razeeth and Dallol, Ashraf and Al-Maghrabi, Jaudah and Aldahlawi, Alia and Ashoor, Sawsan and Maamra, Mabrouka and Ragoussis, Jiannis and Wu, Wei and others},
-  journal={Frontiers in Oncology},
-  volume={10},
-  pages={804},
-  year={2020},
-  publisher={Frontiers}
+    title = {Integration of Transcriptome and Metabolome Provides Unique Insights to Pathways Associated With Obese Breast Cancer Patients},
+    author = {Hassan, Mohammed A and Al-Sakkaf, Kaltoom and Shait Mohammed, Mohammed Razeeth and Dallol, Ashraf and Al-Maghrabi, Jaudah and Aldahlawi, Alia and Ashoor, Sawsan and Maamra, Mabrouka and Ragoussis, Jiannis and Wu, Wei and others},
+    journal = {Frontiers in Oncology},
+    volume = {10},
+    pages = {804},
+    year = {2020},
+    publisher = {Frontiers},
+    doi = {10.3389/fonc.2020.00804}
 }
 
 @article{stuart2019integrative,
@@ -45,14 +47,15 @@
 }
 
 @article{cervera2019anduril,
-  title={Anduril 2: upgraded large-scale data integration framework},
-  author={Cervera, Alejandra and Rantanen, Ville and Ovaska, Kristian and Laakso, Marko and Nu{\~n}ez-Fontarnau, Javier and Alkodsi, Amjad and Casado, Julia and Facciotto, Chiara and H{\"a}kkinen, Antti and Louhimo, Riku and others},
-  journal={Bioinformatics},
-  volume={35},
-  number={19},
-  pages={3815--3817},
-  year={2019},
-  publisher={Oxford University Press}
+    title = {Anduril 2: upgraded large-scale data integration framework},
+    author = {Cervera, Alejandra and Rantanen, Ville and Ovaska, Kristian and Laakso, Marko and Nu{\~n}ez-Fontarnau, Javier and Alkodsi, Amjad and Casado, Julia and Facciotto, Chiara and H{\"a}kkinen, Antti and Louhimo, Riku and others},
+    journal = {Bioinformatics},
+    volume = {35},
+    number = {19},
+    pages = {3815--3817},
+    year = {2019},
+    publisher = {Oxford University Press},
+    doi = {10.1093/bioinformatics/btz133}
 }
 
 @article{lappalainen2013transcriptome,
@@ -96,7 +99,8 @@
     number = {11},
     pages = {e1005752},
     year = {2017},
-    publisher = {Public Library of Science}
+    publisher = {Public Library of Science},
+    doi = {10.1371/journal.pcbi.1005752},
 }
 
 @article{boekel2015multi,
@@ -129,7 +133,8 @@
     number = {9},
     pages = {65},
     year = {2010},
[…]
