+++ inst/paper.bib
@@ -14,14 +14,26 @@
 }
 
 @article{cancer2014comprehensive,
-  title={Comprehensive molecular profiling of lung adenocarcinoma},
-  author={Cancer Genome Atlas Research Network and others},
-  journal={Nature},
-  volume={511},
-  number={7511},
-  pages={543--550},
-  year={2014},
-  publisher={Nature Publishing Group}
+    title = {Comprehensive molecular profiling of lung adenocarcinoma},
+    author = {Cancer Genome Atlas Research Network and others},
+    journal = {Nature},
+    volume = {511},
+    number = {7511},
+    pages = {543--550},
+    year = {2014},
+    publisher = {Nature Publishing Group}
+}
+
+@article{wolf2018scanpy,
+    title = {SCANPY: large-scale single-cell gene expression data analysis},
+    author = {Wolf, F Alexander and Angerer, Philipp and Theis, Fabian J},
+    journal = {Genome biology},
+    volume = {19},
+    number = {1},
+    pages = {1--5},
+    year = {2018},
+    publisher = {BioMed Central},
+    doi = {10.1186/s13059-017-1382-0}
 }
 
 @article{hassan2020integration,
@@ -88,7 +100,8 @@
     number = {1},
     pages = {154--163},
     year = {2016},
-    publisher = {ASBMB}
+    publisher = {ASBMB},
+    doi = {10.1074/mcp.M115.052381}
 }
 
 @article{rohart2017mixomics,
@@ -111,7 +124,8 @@
     number = {2},
     pages = {137--139},
     year = {2015},
-    publisher = {Nature Publishing Group}
+    publisher = {Nature Publishing Group},
+    doi = {10.1038/nbt.3134}
 }
 
 @article{sangaralingam2019multi,
@@ -122,7 +136,8 @@
     number = {1},
     pages = {130--143},
     year = {2019},
-    publisher = {Oxford University Press}
+    publisher = {Oxford University Press},
+    doi = {10.1093/bib/bbx080},
 }
 
 @article{ovaska2010large,
@@ -303,7 +318,7 @@
     pages = {gkw943},
     year = {2016},
     publisher = {Oxford University Press},
-    doi = {.1093/nar/gkw943},
+    doi = {10.1093/nar/gkw943},
 }
 
 
