### 2021-05-08 983c2d0
Minor tweaks to paper.
