+++ inst/paper.bib
@@ -6,6 +6,28 @@
     doi = {10.1093/nar/gkaa554},
 }
 
+@software{reback2020pandas,
+    author = {The pandas development team},
+    title = {pandas-dev/pandas: Pandas},
+    month = feb,
+    year = 2020,
+    publisher = {Zenodo},
+    version = {latest},
+    doi = {10.5281/zenodo.3509134},
+    url = {https://doi.org/10.5281/zenodo.3509134}
+}
+
+@InProceedings{matthew_rocklin-proc-scipy-2015,
+    author = { Matthew Rocklin },
+    title = { Dask: Parallel Computation with Blocked algorithms and Task Scheduling },
+    booktitle = { Proceedings of the 14th Python in Science Conference },
+    pages = { 130 - 136 },
+    year = { 2015 },
+    editor = { Kathryn Huff and James Bergstra },
+    doi = {10.25080/Majora-7b98e3ed-013},
+}
+
+
 @techreport{hagberg2008exploring,
     title = {Exploring network structure, dynamics, and function using NetworkX},
     author = {Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
