+++ inst/paper.bib
@@ -6,15 +6,14 @@
     doi = {10.1093/nar/gkaa554},
 }
 
-@software{reback2020pandas,
-    author = {The pandas development team},
-    title = {pandas-dev/pandas: Pandas},
-    month = feb,
-    year = 2020,
-    publisher = {Zenodo},
-    version = {latest},
-    doi = {10.5281/zenodo.3509134},
-    url = {https://doi.org/10.5281/zenodo.3509134}
+@InProceedings{mckinney-proc-scipy-2010,
+    author = { {W}es {M}c{K}inney },
+    title = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+    booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+    pages = { 56 - 61 },
+    year = { 2010 },
+    editor = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+    doi = { 10.25080/Majora-92bf1922-00a }
 }
 
 @InProceedings{matthew_rocklin-proc-scipy-2015,
