### 2021-05-09 #31 @arfon
This is now published in JOSS 🥳
