### 2022-09-28 #31 @JonnyTran
> Hi there wave @JonnyTran @ksielemann fill our our onboarding / feedback survey for this review i'd greatly appreciate it!! Many thanks in advance for your time. it really helps our organization! Jonny, if there are other maintainers involved in this package please ping them here as well. we would like to keep in touch with our maintainers to ensure the package is still being kept up to date AND to understand how we can better support you !

Thank you for the reminder! I've just submitted the survey. Looking forward to future contributions to PyOpenSci
